### 2025-05-05 #8162 @editorialbot
[PRE REVIEW]: ManipulaPy: A GPU-Accelerated Python Framework for Robotic Manipulation, Perception, and Control
**Submitting author:** @boelnasr (<a href="http://orcid.org/0000-0002-1768-2031">Mohamed Aboelnasr</a>)
**Repository:** https://github.com/boelnasr/ManipulaPy
**Branch with paper.md** (empty if default branch):
**Version:** V 1.1.0
