### 2025-05-05 #8162 @crvernon
:wave: @boelnasr - Thanks for you submission to JOSS.  While I am getting you setup with a topic editor, please reduce the length of your paper to <= 1000 words.  Also, please check out the pdf generation error above and see if we can get that compiling.  Thanks!
