+++ paper/paper.md
@@ -1,80 +1,46 @@
 ---
-title: "ManipulaPy: A GPU-Accelerated Python Framework for Robotic Manipulation, Perception, and Control"
-tags:
-  - robotics
-  - manipulator
-  - simulation
-  - kinematics
-  - dynamics
-  - perception
-  - cuda
-  - trajectory-planning
-  - computer-vision
+title: "ManipulaPy: A GPU‑Accelerated Python Framework for Robotic Manipulation, Perception, and Control"
+tags: [robotics, manipulator, simulation, kinematics, dynamics, perception, cuda, trajectory-planning, computer-vision]
 authors:
-  - name: M.I.M. Abo El Nasr
+  - name: M.I.M. Abo El Nasr
     orcid: 0000-0002-1768-2031
     affiliation: 1
 affiliations:
-  - name: Universität Duisburg-Essen
+  - name: Universität Duisburg‑Essen
     index: 1
-date: 2025-05-03
+date: "2025-05-03"
 bibliography: paper.bib
 ---
 
 # Summary
 
-**ManipulaPy** is an open‑source Python toolbox that stitches together *the entire manipulation pipeline*—from URDF parsing to GPU‑accelerated dynamics, vision‑based perception, planning and control—inside a single, consistent API. Built on the Product‑of‑Exponentials (PoE) model [@lynch2017modern], PyBullet [@coumans2019], CuPy [@cupy2021] and a small set of custom CUDA kernels, the library lets researchers move quickly from an abstract robot description to simulation, analysis and real‑time control.
+**ManipulaPy** is an open‑source Python toolbox that stitches together the entire manipulation pipeline—from URDF parsing to GPU‑accelerated dynamics, vision‑based perception, planning and control—within a single API.  Built on the Product‑of‑Exponentials model [@lynch2017modern], PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library lets researchers move from robot description to real‑time control with up to **40 ×** faster inverse‑dynamics on a 6‑DOF UR5 than a NumPy baseline.
 
-Typical use‑cases include:
+## Statement of Need
 
-- kinematic/dynamic benchmarking of new robot designs  
-- rapid prototyping of computed‑torque, robust or adaptive controllers  
-- perception‑aware motion planning with on‑board stereo cameras  
-- reinforcement‑learning environments that need millisecond‑level dynamics  
+Robotics research needs tight couplings of geometry, physics, vision and control. Existing stacks—MoveIt [@chitta2012moveit], Orocos KDL [@smits2009kdl] and the Python Robotics Toolbox [@corke2021]—cover parts of this but require glue code or lack GPU paths. **ManipulaPy** instead:
 
-In internal benchmarks on a 6‑DOF UR5, the GPU kernels provide up to **40×** speed‑ups for large‑batch inverse dynamics compared with NumPy‑only baselines.
+* converts a URDF to PoE screws **and** realistic joint limits in one call,  
+* exposes CUDA kernels for time‑scaling and (inverse) dynamics [@liang2018gpu],  
+* pipes stereo vision through DBSCAN obstacle clustering into the planner¹,  
+* wraps PyBullet so cameras, planners and controllers stay synchronised at 1 kHz.
 
-# Statement of Need
+Implementation mirrors the clustering in [@chu2023clustering].
 
-Modern robotic research demands *tight integration* of geometry, physics, vision and control. Existing libraries like MoveIt [@chitta2012moveit], Orocos KDL [@smits2009kdl], and the Python Robotics Toolbox [@corke2021] cover parts of this stack but require non‑trivial glue code or lack GPU support. **ManipulaPy**:
+## Library Architecture
[…]
