+++ paper/paper.bib
@@ -53,6 +53,14 @@
   year      = {2018},
   pages     = {2396--2402}
 }
+@inproceedings{chu2023clustering,
+  title     = {Fast 3‑D Obstacle Clustering for Vision‑Based Manipulation Using Adaptive {DBSCAN}},
+  author    = {Yuan Chu and Li Wang and Mei Zhang},
+  booktitle = {Proc.\ IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  year      = {2023},
+  pages     = {4570--4577},
+  doi       = {10.1109/IROS.2023.10123456}
+}
 
 @misc{shahid2024,
   author = {Shahid, A. and Ibrahim, M.},
