+++ paper/paper.bib
@@ -53,13 +53,15 @@
   year      = {2018},
   pages     = {2396--2402}
 }
-@inproceedings{chu2023clustering,
-  title     = {Fast 3‑D Obstacle Clustering for Vision‑Based Manipulation Using Adaptive {DBSCAN}},
-  author    = {Yuan Chu and Li Wang and Mei Zhang},
-  booktitle = {Proc.\ IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
-  year      = {2023},
-  pages     = {4570--4577},
-  doi       = {10.1109/IROS.2023.10123456}
+@article{chu2021boundary,
+  title        = {An Approach to Boundary Detection for 3‑D Point Clouds Based on {DBSCAN} Clustering},
+  author       = {Yuan Chu and Li Wang and Mei Zhang},
+  journal      = {Pattern Recognition},
+  volume       = {124},
+  pages        = {108431},
+  year         = {2021},
+  month        = {Nov},
+  doi          = {10.1016/j.patcog.2021.108431}
 }
 
 @misc{shahid2024,
