+++ paper/paper.md
@@ -16,7 +16,7 @@ bibliography: paper.bib
 
 **ManipulaPy** is an open‑source Python toolbox that stitches together the entire manipulation pipeline—from URDF parsing to GPU‑accelerated dynamics, vision‑based perception, planning and control—within a single API.  Built on the Product‑of‑Exponentials model [@lynch2017modern], PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library lets researchers move from robot description to real‑time control with up to **40 ×** faster inverse‑dynamics on a 6‑DOF UR5 than a NumPy baseline.
 
-## Statement of need
+## Statement of need
 
 Robotics research needs tight couplings of geometry, physics, vision and control. Existing stacks—MoveIt [@chitta2012moveit], Orocos KDL [@smits2009kdl] and the Python Robotics Toolbox [@corke2021]—cover parts of this but require glue code or lack GPU paths. **ManipulaPy** instead:
 
@@ -25,7 +25,7 @@ Robotics research needs tight couplings of geometry, physics, vision and control
 * pipes stereo vision through DBSCAN obstacle clustering into the planner¹,  
 * wraps PyBullet so cameras, planners and controllers stay synchronised at 1 kHz.
 
-Implementation mirrors the clustering in [@chu2023clustering].
+Implementation mirrors the clustering in [@chu2021boundary].
 
 ## Library Architecture
 
