### 2025-05-08 #8162 @boelnasr
> 👋 [@boelnasr](https://github.com/boelnasr) - Thanks for you submission to JOSS. While I am getting you setup with a topic editor, please reduce the length of your paper to <= 1000 words. Also, please check out the pdf generation error above and see if we can get that compiling. Thanks!

Thanks for the  feedback, Chris!

• Paper length – I’ve condensed paper.md to ≈ 780 words, well under the 1 000-word limit.
• PDF build – The latest push compiles cleanly.
• Reviewer suggestions – Below are several suggestions for reviewers:

- Alexander Fabisch
- Abdelhak Bougouffa
- CameronDevine
- Martin Huber
