### 2025-06-20 #8162 @lockwo
Hi 👋  @sea-bass, @ishaanamahajan, @cadojo would two (or more) of you be willing to review this submission for JOSS?
