### 2025-06-23 #8162 @ishaanamahajan
> Hi 👋 [@sea-bass](https://github.com/sea-bass), [@ishaanamahajan](https://github.com/ishaanamahajan), [@cadojo](https://github.com/cadojo) would two (or more) of you be willing to review this submission for JOSS?

Hi @lockwo thanks for inviting me to this JOSS review but I unfortunately won't be able to review it due to other commitments made beforehand.
