### 2025-06-28 boelnasr/ManipulaPy#2 @sea-bass
[JOSS Review] Ultralytics invalidates MIT license in this repo
Unfortunately, Ultralytics is AGPL 3.0 licensed ([see here](https://github.com/ultralytics/ultralytics/blob/main/LICENSE)), which means that this repo cannot be licensed under its existing MIT license.

The easiest thing would be to change your own project to also have an AGPL 3.0 license.

If you would like to keep your project as MIT, you can either find alternative implementations of YOLO detectors, or provide the YOLO detector interface as an optional interface that users can download separately.

For more information, see https://github.com/ultralytics/ultralytics/pull/2031

---

https://github.com/openjournals/joss-reviews/issues/8490
