### 2025-06-28 boelnasr/ManipulaPy#3 @sea-bass
[JOSS Review] Installation failed without `nvcc` and `cudnn` available
I went through the installation process as-is, and found some errors when installing `pycuda` as one of the package dependencies.

Narrowing it down, the errors pointed to `nvcc` not being available. After running `sudo apt install nvidia-cuda-toolkit`, installation worked as expected.

Please consider adding this -- or something similar -- to the installation steps, in case others also run into this issue.

EDIT: Based on my experience with https://github.com/boelnasr/ManipulaPy/issues/4, I think additionally `cudnn` needs to be installed by the user for this all to work.

It may go a long way to have some kind of "checkpoint" in the instructions for users to check that their installation works. For example, "enter `nvidia-smi` and check that the output contains XYZ"...

---

https://github.com/openjournals/joss-reviews/issues/8490
