### 2025-06-28 boelnasr/ManipulaPy#4 @sea-bass
[JOSS Review] Add `torch`, `numba`, `ultralytics`, and `cupy` as dependencies
After going through the installation process, and unblocking myself from https://github.com/boelnasr/ManipulaPy/issues/3, the next thing I tried was

```
python3
>>> import ManipulaPy
```

At which point I got this error:

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/sebastian/workspace/ManipulaPy/ManipulaPy/__init__.py", line 11, in <module>
    from ManipulaPy.kinematics import *
  File "/home/sebastian/workspace/ManipulaPy/ManipulaPy/kinematics.py", line 35, in <module>
    import torch
ModuleNotFoundError: No module named 'torch'
```

After manually installing `torch`, I tried again and got:

```
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/home/sebastian/workspace/ManipulaPy/ManipulaPy/__init__.py", line 13, in <module>
    from ManipulaPy.singularity import *
  File "/home/sebastian/workspace/ManipulaPy/ManipulaPy/singularity.py", line 6, in <module>
    from numba import cuda
ModuleNotFoundError: No module named 'numba'
```

Then, I kept going and also got errors about `ultralytics` and `cupy`

So, I think `torch`, `numba`, `ultralytics`, and `cupy` also need be included as a dependency for this project. Please add them.

Also, I couldn't install `cupy` right away because I was missing the cuDNN headers, at which point I stopped. Some instructions will be needed here too.

---

https://github.com/openjournals/joss-reviews/issues/8490
