### 2025-06-28 boelnasr/ManipulaPy#5 @sea-bass
[JOSS Review] Do not include YOLOv8 weights in repo
Right now, the files `yolov8m.pt` and `yolov8n.pt` files are simply available in the top level of the repo.

Not only are these files quite large, but also I imagine if users install this project from PyPi they do not get these files, right?

I would highly recommend having this either have download instructions, or automatically download weights if not available.

---

https://github.com/openjournals/joss-reviews/issues/8490
