### 2025-06-28 boelnasr/ManipulaPy#6 @sea-bass
[JOSS Review] Add official documentation
Right now, all the documentation seems to be available in the repo itself, under `examples/doc.IPYNB`.

To pass the quality bar for JOSS, I believe the documentation needs to be available as an actual page -- either in a well-organized doc folder within the repo, with different sections, or through an externally hosted page, for example, using [ReadTheDocs](https://app.readthedocs.org/) or similar.

---

https://github.com/openjournals/joss-reviews/issues/8490
