### 2025-06-28 boelnasr/ManipulaPy#7 @sea-bass
[JOSS Review] Do not include `.vscode` folders in repo
The repo right now has a bunch of `.vscode` folders in them, which I think you should remove.

I recommend adding a `.gitignore` file at the top-level of your repo where you can enter files/folders to not include under source control. I believe you can add

```
**/.vscode
```

to that file to prevent them from being added -- but you'll need to delete the existing ones first!

---

https://github.com/openjournals/joss-reviews/issues/8490
