### 2025-06-28 boelnasr/ManipulaPy#8 @sea-bass
[JOSS Review] Clarify the difference between `perception.py` and `vision.py` files
Right now, the `ManipulaPy` module has both a `perception.py` and a `vision.py` -- which are very similar to each other in name! Looking at the code, I'm also not immediately sure about the difference should be.

I would consider either consolidating them, or being more precise about their naming to help people reading the code.

---

https://github.com/openjournals/joss-reviews/issues/8490
