### 2025-06-28 boelnasr/ManipulaPy#9 @sea-bass
[JOSS Review] Tests do not sufficiently cover the tool's functionality
Looking at the way testing is setup, most of the GPU intensive modules are mocked (likely due to free CI runners not having GPU support). For example:

```
# Mock third-party modules that are unavailable in CI or test environments
MOCK_MODULES = [
    "torch",
    "cupy",
    "pycuda",
    "pycuda.driver",
    "pycuda.autoinit",
    "numba",
    "numba.cuda",
    "pybullet",
    "urchin",
    "urchin.urdf",
    "cv2",
    "ultralytics",
    "sklearn.cluster",
]
```

I think this results in the functionality not being properly tested. Specifically for tools like `cv2`, `torch`, `numba`, `sklearn`, and `ultralytics`, there should be a path to getting those working on CPU and should be probably taken off the list of mock modules, along with accompanying tests.

---

https://github.com/openjournals/joss-reviews/issues/8490
