### 2025-06-28 boelnasr/ManipulaPy#10 @sea-bass
[JOSS Review] Expand contributing guidelines
Per JOSS standard, the contributing guidelines state:

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Right now, there is only one sentence in the README describing that the repo should be forked and PEP8 style should be followed; however, there should be more information that covers the above criteria.

Thanks!

---

https://github.com/openjournals/joss-reviews/issues/8490
