### 2025-06-28 boelnasr/ManipulaPy#11 @sea-bass
[JOSS Review] Clean up example file naming and add top-level comments
Right now, the examples seem to follow inconsistent naming styles. For instance:

* A few examples are named `EX_XXXXX.py`, others are named `xxxxx_ex.py`, and others named `test_xxxx.py`.
* There is an `example.py` and an `Example.py` (different casing in the first letters)

Consider adopting a consistent style to help readers of the code navigate the examples.

I would also suggest for each example file to at least add 1 sentence as a top-level comment explaining what each example does. That goes a long way!

---

https://github.com/openjournals/joss-reviews/issues/8490
