### 2025-06-28 boelnasr/ManipulaPy#12 @sea-bass
[JOSS Review] Paper feedback
Thanks for submitting this repository to JOSS! Here is my feedback on the paper part of the submission:

# General
- [x] The paper mentions up to 40x faster than a NumPy baseline -- can one reproduce these experimental results using your repo? If so, please include instructions in the repo itself.
- [x] If mentioning product-of-exponentials approaches, you should cite [Pinocchio](https://github.com/stack-of-tasks/pinocchio) in this work as a popular tool that also adopts this notation. However, you can mention that your tool has GPU support whereas Pinocchio does not.
- [x] If possible, your paper should include at least one image of an example in action -- e.g., with a PyBullet sim window performing a task.
- [x] I think it would be helpful to better discuss elements of your vision pipeline. I see certain things like YOLO detection and DBSCAN obstacle clustering scattered throughout, but it may be beneficial to explain exactly what is possible here. For example, some manipulation pipelines deal with only geometric primitives; others support signed distance fields; others can take a point cloud of the environment and treat it as an "unstructured" obstacle via representations like octrees/octomaps. What exactly does ManipulaPy support?

# Per Section comments

- [x] The claim that other tools "require glue code" seems quite unsubstantiated. Could you be more specific with some examples of glue code?
- [x] I can imagine that there are many other GPU accelerated manipulation libraries that you could also cite here. Notably there is [cuRobo](https://curobo.org/) which you should absolutely compare to! You may also want to speak to the newer set of Jax accelerated Python tools for manipulation, such as [PyRoKi](https://pyroki-toolkit.github.io/) and [CBFPy](https://danielpmorton.github.io/cbfpy/).
- [x] It's a bit confusing that the citation to DBSCAN is available as a footnote, when it is short enough to just be inline in the same bullet and may be more clear that way.

## Library Architecture
- [x] I think in addition to just bullet points, an initial paragraph would go a long way.
- [x] As in https://github.com/boelnasr/ManipulaPy/issues/8, any clarification you can make about `perception.py` vs. `vision.py` would be good to address here too
- [x] Generally, I think this section is too specifically tied to the existing implementation. As you add more functionality to this tool or refactor the code, simply listing out the snapshot of files at the time of submission will not age well. I recommend switching this section from having such a list to speaking more generally about concepts, and possibly including an architecture diagram of your intended manipulation pipeline.

## Acknowledgements
- [x] The citation to Ultralytics YOLO is broken, please fix it.

---

https://github.com/openjournals/joss-reviews/issues/8490
