### 2025-06-28 boelnasr/ManipulaPy#4 @sea-bass
Update: I notice that the repo has a `requirements.txt` with many of these requirements already available, but doing `pip install` on the package won't automatically pick these up.

You should move these requirements to the `pyproject.toml` itself so they get picked up appropriately!
