+++ pyproject.toml
@@ -1,7 +1,152 @@
+# ──────────────────────────────────────────────────────────────
+# Project build configuration for ManipulaPy
+# ──────────────────────────────────────────────────────────────
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+# ------------------------------------------------------------------
+[project]
+name = "ManipulaPy"                    # exact PyPI casing
+version = "1.1.0"                      
+authors = [
+    { name = "Mohamed Aboelnar", email = "aboelnasr1997@gmail.com" },
+]
+description = "A comprehensive Python package for robotic manipulator analysis and control"
+readme = "README.md"
+license = { text = "AGPL-3.0-or-later" }
+requires-python = ">=3.8"
+
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+    "Topic :: Scientific/Engineering :: Physics",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+]
+keywords = [
+    "robotics",
+    "kinematics",
+    "dynamics",
+    "trajectory-planning",
+    "simulation",
+    "cuda",
+]
+
+# ------------------------------------------------------------------
+# Core runtime dependencies - FIXED: Removed problematic GPU dependencies from core
+dependencies = [
+    "numpy>=1.19.2",
+    "scipy>=1.5.2",
+    "pybullet>=3.0.6",
+    "trimesh>=3.9.14",
+    "urchin>=0.0.27",
+    "numba>=0.55",
+    "matplotlib>=3.3",
+    "scikit-learn>=1.0",
+    "opencv-python>=4.5",
+    "ultralytics>=8.0",
+    "torch>=1.8.0",
+]
+
+[project.optional-dependencies]
+# GPU support - Multiple options to handle different CUDA versions
+gpu-cuda11 = ["cupy-cuda11x>=10.0.0"]
+gpu-cuda12 = ["cupy-cuda12x>=12.0.0"]
+gpu-rocm = ["cupy-rocm-4-3>=10.0.0"]
+# Legacy PyCUDA (alternative to CuPy)
+gpu-pycuda = ["pycuda>=2021.1"]
+# All GPU options combined
+gpu = ["cupy-cuda11x>=10.0.0"]  # Default to CUDA 11
+
+# Development dependencies
+dev = [
+    "pytest>=6.0.0",
+    "pytest-cov>=2.0.0",
+    "black>=21.0.0",
+    "flake8>=3.8.0",
+    "isort>=5.0.0",
+    "mypy>=0.910",
+    "pre-commit>=2.15.0",
+]
+
+# Documentation dependencies
+docs = [
+    "sphinx>=4.0.0",
+    "sphinx-rtd-theme>=1.0.0",
+    "myst-parser>=0.15.0",
+]
+
+# Testing with GPU support
+test-gpu = [
+    "ManipulaPy[dev,gpu-cuda11]"
+]
+
+# All optional dependencies (excluding GPU to avoid installation issues)
+all = ["ManipulaPy[dev,docs]"]
+
+[project.urls]
[…]
