### 2025-06-29 35949e4
Issue #1
Solving the Licencing issue
