+++ ManipulaPy/cuda_kernels.py
@@ -1,225 +1,441 @@
-from numba import cuda, float32
+#!/usr/bin/env python3
+# SPDX-License-Identifier: AGPL-3.0-or-later
+"""
+CUDA Kernels Module - ManipulaPy
+
+This module provides CUDA-accelerated functions for trajectory planning and dynamics
+computation. All CUDA functionality is optional and gracefully degrades to CPU implementations.
+
+Copyright (c) 2025 Mohamed Aboelnar
+Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
+
+This file is part of ManipulaPy.
+
+ManipulaPy is free software: you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+ManipulaPy is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU Affero General Public License for more details.
+
+You should have received a copy of the GNU Affero General Public License
+along with ManipulaPy. If not, see <https://www.gnu.org/licenses/>.
+"""
+
 import numpy as np
+import warnings
+
+# Optional CUDA imports with graceful fallback
+try:
+    from numba import cuda, float32
+    CUDA_AVAILABLE = True
+except ImportError:
+    CUDA_AVAILABLE = False
+    # Create mock objects to prevent import errors
+    class MockCuda:
+        @staticmethod
+        def jit(func):
+            """Mock decorator for when CUDA is not available"""
+            def wrapper(*args, **kwargs):
+                raise RuntimeError(
+                    "CUDA functionality not available. Install CUDA support:\n"
+                    "pip install ManipulaPy[gpu-cuda11]  # For CUDA 11.x\n"
+                    "pip install ManipulaPy[gpu-cuda12]  # For CUDA 12.x\n"
+                    "Ensure you have NVIDIA GPU drivers and CUDA toolkit installed."
+                )
+            return wrapper
+        
+        @staticmethod
+        def grid(dim):
+            """Mock grid function"""
+            return 0
+        
+        @staticmethod
+        def device_array(*args, **kwargs):
+            """Mock device array function"""
+            raise RuntimeError("CUDA not available - cannot create device arrays")
+        
+        @staticmethod
+        def to_device(*args, **kwargs):
+            """Mock to_device function"""
+            raise RuntimeError("CUDA not available - cannot transfer to device")
+        
+        @staticmethod
+        def local():
+            """Mock local memory"""
+            class LocalMock:
+                @staticmethod
+                def array(*args, **kwargs):
+                    raise RuntimeError("CUDA not available - cannot create local arrays")
+            return LocalMock()
+    
+    cuda = MockCuda()
+    float32 = np.float32
+
+# Check for CuPy availability (separate from numba.cuda)
+try:
+    import cupy as cp
+    CUPY_AVAILABLE = True
+except ImportError:
+    CUPY_AVAILABLE = False
+    cp = None
 
 
-@cuda.jit
-def trajectory_kernel(
-    thetastart, thetaend, traj_pos, traj_vel, traj_acc, Tf, N, method
-):
+def check_cuda_availability():
     """
-    CUDA kernel to compute positions, velocities, and accelerations using cubic or quintic time scaling.
[…]
