+++ ManipulaPy/urdf_processor.py
@@ -1,4 +1,30 @@
 #!/usr/bin/env python3
+# SPDX-License-Identifier: AGPL-3.0-or-later
+"""
+URDF Processor Module - ManipulaPy
+
+This module provides comprehensive URDF (Unified Robot Description Format) processing
+capabilities including conversion to SerialManipulator objects, extraction of kinematic
+and dynamic parameters, and integration with PyBullet for simulation and visualization.
+
+Copyright (c) 2025 Mohamed Aboelnar
+
+This file is part of ManipulaPy.
+
+ManipulaPy is free software: you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+ManipulaPy is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU Affero General Public License for more details.
+
+You should have received a copy of the GNU Affero General Public License
+along with ManipulaPy. If not, see <https://www.gnu.org/licenses/>.
+"""
+
 
 from urchin.urdf import URDF
 import numpy as np
