### 2025-06-29 a8af738
Issue #1
Solving the Licencing issue
