### 2025-06-29 bb82e40
Issue #1
Solving the Licencing issue
