+++ README.md
@@ -1,168 +1,571 @@
 # ManipulaPy
 
-
 [![PyPI](https://img.shields.io/pypi/v/ManipulaPy)](https://pypi.org/project/ManipulaPy/)
-![Tested](https://img.shields.io/badge/tested-yes-brightgreen.svg)
+[![Python](https://img.shields.io/badge/python-3.8%2B-blue.svg)](https://www.python.org/downloads/)
+[![License](https://img.shields.io/badge/License-AGPL%20v3-blue.svg)](https://www.gnu.org/licenses/agpl-3.0)
 ![CI](https://github.com/boelnasr/ManipulaPy/actions/workflows/test.yml/badge.svg?branch=main)
+![Test Status](https://img.shields.io/badge/tests-passing-brightgreen)
+[![status](https://joss.theoj.org/papers/e0e68c2dcd8ac9dfc1354c7ee37eb7aa/status.svg)](https://joss.theoj.org/papers/e0e68c2dcd8ac9dfc1354c7ee37eb7aa)
 
-ManipulaPy is a modular, GPU-accelerated Python package for robotic manipulator simulation, planning, control, and perception. It supports kinematics, dynamics, path planning, control strategies, PyBullet simulation, and even 3D vision with stereo processing and YOLO-based obstacle detection.
+**ManipulaPy** is a comprehensive, GPU-accelerated Python package for robotic manipulator analysis, simulation, planning, control, and perception. It provides a unified framework combining advanced kinematics, dynamics modeling, trajectory planning, control strategies, PyBullet simulation, and computer vision capabilities with optional CUDA acceleration.
 
 ---
 
-## 🚀 Features
+## 🎯 Statement of Need
 
-- **Kinematic Analysis**: Forward and inverse kinematics for serial robots
-- **Dynamic Modeling**: Mass matrix, Coriolis/centrifugal, gravity force computation
-- **Path Planning**: Joint and Cartesian trajectory generation with time scaling
-- **Singularity Analysis**: Detect singularities, plot manipulability ellipsoids, and estimate workspace
-- **URDF Processing**: Convert URDF files into manipulatable Python models
-- **Control Strategies**: PD, PID, computed torque, robust, adaptive, Kalman filter, and feedforward control
-- **Simulation**: Real-time PyBullet simulation of joint-space trajectories
-- **Vision & Perception**: Stereo camera modeling, depth/disparity map computation, DBSCAN clustering, YOLO-based object detection
-- **Visualization Tools**: Trajectory plotting, steady-state response analysis, 3D cluster rendering
+Modern robotics research and development requires sophisticated tools for manipulator analysis and control. Existing solutions often lack integration between kinematic analysis, dynamic modeling, trajectory planning, and perception systems. ManipulaPy addresses these critical gaps by providing:
+
+- **🔧 Unified Framework**: Seamless integration of kinematics, dynamics, control, and perception
+- **⚡ GPU Acceleration**: High-performance CUDA kernels for real-time applications
+- **🔬 Research-Ready**: Comprehensive algorithms suitable for academic and industrial research
+- **🧩 Modular Design**: Use individual components or the complete integrated system
+- **📖 Well-Documented**: Extensive documentation with theoretical background and practical examples
+- **🆓 Open Source**: AGPL-3.0 licensed for transparency, collaboration, and academic use
+
[…]
