### 2025-06-30 08d7aef
Issue #4
Remove the Yolov8m.pt and Yolovn8.pt
