### 2025-06-30 e815faf
Issue #4
Remove the Yolov8m.pt and Yolovn8.pt
