### 2025-06-30 0b0750a
Issue #6
remove .vscode
