### 2025-06-30 131f593
Issue #6
remove .vscode
