### 2025-06-30 6f4d35a
Issue #4
Delete yolo files
