### 2025-06-30 7f02d1c
Issue #4
Delete yolo files
