### 2025-07-01 e556c4e
Issue #4
Update the yolo integration
