+++ tests/test_control.py
@@ -9,8 +9,27 @@ from ManipulaPy.urdf_processor import URDFToSerialManipulator
 from ManipulaPy.ManipulaPy_data.xarm import urdf_file as xarm_urdf_file
 
 
+def is_module_available(module_name):
+    """Check if a module is available and not mocked."""
+    try:
+        module = __import__(module_name)
+        return not hasattr(module, '_name') or module._name != f"MockModule({module_name})"
+    except ImportError:
+        return False
+
+
 class TestManipulatorController(unittest.TestCase):
     def setUp(self):
+        # Determine backend
+        if is_module_available('cupy'):
+            self.backend = 'cupy'
+            self.cp = cp
+            print("Using cupy backend for testing")
+        else:
+            self.backend = 'numpy'
+            self.cp = np
+            print("Using numpy backend for testing")
+
         # Use the built-in xarm urdf file from the library
         self.urdf_path = xarm_urdf_file
 
@@ -26,9 +45,9 @@ class TestManipulatorController(unittest.TestCase):
 
             # Get the number of joints from the dynamics
             num_joints = len(self.dynamics.Glist)
-            self.thetalist = np.zeros(num_joints)
-            self.dthetalist = np.zeros(num_joints)
-            self.ddthetalist = np.zeros(num_joints)
+            self.thetalist = np.zeros(num_joints, dtype=np.float64)  # Fix dtype
+            self.dthetalist = np.zeros(num_joints, dtype=np.float64)  # Fix dtype
+            self.ddthetalist = np.zeros(num_joints, dtype=np.float64)  # Fix dtype
 
             # Default joint and torque limits if not available
             self.joint_limits = np.array([[-np.pi, np.pi]] * num_joints)
@@ -107,9 +126,9 @@ class TestManipulatorController(unittest.TestCase):
         self.g = np.array([0, 0, -9.81])
         self.Ftip = np.array([0, 0, 0, 0, 0, 0])
         self.dt = 0.01
-        self.thetalist = np.array([0.1, 0.2])
-        self.dthetalist = np.array([0, 0])
-        self.ddthetalist = np.array([0, 0])
+        self.thetalist = np.array([0.1, 0.2], dtype=np.float64)  # Fix dtype
+        self.dthetalist = np.array([0, 0], dtype=np.float64)  # Fix dtype
+        self.ddthetalist = np.array([0, 0], dtype=np.float64)  # Fix dtype
         self.joint_limits = np.array([[-np.pi, np.pi], [-np.pi, np.pi]])
         self.torque_limits = np.array([[-10, 10], [-10, 10]])
 
@@ -117,7 +136,7 @@ class TestManipulatorController(unittest.TestCase):
         """Test PID control convergence to a setpoint."""
         # Set up test parameters
         num_joints = len(self.thetalist)
-        thetalistd = np.array([0.5, 0.7] if num_joints == 2 else [0.5] * num_joints)
+        thetalistd = np.array([0.5, 0.7] if num_joints == 2 else [0.5] * num_joints, dtype=np.float64)
         dthetalistd = np.zeros_like(thetalistd)
 
         # Define gains
@@ -126,27 +145,31 @@ class TestManipulatorController(unittest.TestCase):
         Kd = np.array([1.0] * num_joints)
 
         # Simulate a simple control loop
-        thetalist = np.copy(self.thetalist)
-        dthetalist = np.copy(self.dthetalist)
+        thetalist = np.copy(self.thetalist).astype(np.float64)  # Ensure float64
+        dthetalist = np.copy(self.dthetalist).astype(np.float64)  # Ensure float64
         history = []
         steps = 500
 
         for _ in range(steps):
             tau = self.controller.pid_control(
[…]
