+++ tests/test_trajectory_planning.py
@@ -1,418 +1,826 @@
 #!/usr/bin/env python3
 
-import unittest
+from numba import cuda, float32
 import numpy as np
 import matplotlib.pyplot as plt
-import os
-from ManipulaPy.path_planning import TrajectoryPlanning
-from ManipulaPy.urdf_processor import URDFToSerialManipulator
-from ManipulaPy.ManipulaPy_data.xarm import urdf_file as xarm_urdf_file
+from scipy.spatial import ConvexHull
+from ManipulaPy.utils import (
+    TransToRp,
+    MatrixLog3,
+    MatrixExp3,
+    CubicTimeScaling,
+    QuinticTimeScaling,
+)
+from urchin.urdf import URDF
+from ManipulaPy.cuda_kernels import trajectory_kernel, inverse_dynamics_kernel, forward_dynamics_kernel, cartesian_trajectory_kernel
+from ManipulaPy.potential_field import CollisionChecker, PotentialField
+import logging
 
+# Set up logging
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
+    datefmt='%Y-%m-%d %H:%M:%S'
+)
+logger = logging.getLogger(__name__)
 
-class TestTrajectoryPlanning(unittest.TestCase):
-    def setUp(self):
-        # Use the built-in xarm urdf file from the library
-        self.urdf_path = xarm_urdf_file
+class TrajectoryPlanning:
+    def __init__(self, serial_manipulator, urdf_path, dynamics, joint_limits, torque_limits=None):
+        """
+        Initializes a TrajectoryPlanning object.
 
+        Args:
+            serial_manipulator (SerialManipulator): An instance of SerialManipulator.
+            urdf_path (str): The path to the URDF file.
+            dynamics (ManipulatorDynamics): An instance of ManipulatorDynamics.
+            joint_limits (list): A list of tuples representing the joint limits.
+            torque_limits (list, optional): A list of tuples representing the torque limits. Defaults to None.
+        """
+        self.serial_manipulator = serial_manipulator
+        self.dynamics = dynamics
+        self.joint_limits = np.array(joint_limits)
+        self.torque_limits = (
+            np.array(torque_limits)
+            if torque_limits
+            else np.array([[-np.inf, np.inf]] * len(joint_limits))
+        )
+        
+        # Initialize collision checker and potential field with error handling
         try:
-            self.urdf_processor = URDFToSerialManipulator(self.urdf_path)
-            self.robot = self.urdf_processor.serial_manipulator
-            self.dynamics = self.urdf_processor.dynamics
-
-            # Get the number of joints
-            self.num_joints = len(self.dynamics.Glist)
-
-            # Default joint and torque limits
-            self.joint_limits = np.array([[-np.pi, np.pi]] * self.num_joints)
-
-            # Use None for torque_limits to avoid the boolean context issue
-            self.torque_limits = None
-
-            # Create the trajectory planner
-            self.trajectory_planner = TrajectoryPlanning(
-                self.robot,
-                self.urdf_path,
-                self.dynamics,
-                self.joint_limits,
-                self.torque_limits,
-            )
-
-            # Common parameters for testing
-            self.g = np.array([0, 0, -9.81])
-            self.Ftip = np.array([0, 0, 0, 0, 0, 0])
-
+            self.collision_checker = CollisionChecker(urdf_path)
         except Exception as e:
-            print(f"Error initializing test: {e}")
-            self.create_mock_objects()
+            logger.warning(f"Failed to initialize collision checker: {e}")
+            self.collision_checker = None
+        
+        try:
+            self.potential_field = PotentialField()
+        except Exception as e:
+            logger.warning(f"Failed to initialize potential field: {e}")
[…]
