### 2025-07-03 935d327
Issue #9
introducing more comprehensive testing
