+++ tests/test_perception.py
@@ -1,0 +1,543 @@
+#!/usr/bin/env python3
+"""
+test_perception.py - Comprehensive tests for the Perception module.
+Tests the higher-level perception functionality that uses Vision.
+"""
+
+import unittest
+import numpy as np
+import sys
+import os
+from unittest.mock import Mock, patch
+
+# Add the package to path
+sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+
+def is_module_available(module_name):
+    """Check if a module is really available (not mocked)."""
+    try:
+        module = __import__(module_name)
+        return not hasattr(module, '_name') or not str(module._name).startswith('Mock')
+    except ImportError:
+        return False
+
+class TestPerceptionInitialization(unittest.TestCase):
+    """Test Perception module initialization and basic functionality."""
+    
+    def setUp(self):
+        """Set up test fixtures."""
+        # Create mock vision instance
+        self.mock_vision = Mock()
+        self.mock_vision.capture_image.return_value = (
+            np.random.randint(0, 255, (480, 640, 3), dtype=np.uint8),
+            np.random.uniform(0.1, 5.0, (480, 640)).astype(np.float32)
+        )
+        self.mock_vision.detect_obstacles.return_value = (
+            np.random.randn(10, 3),  # 10 obstacle points
+            np.arange(10)  # Labels 0-9
+        )
+    
+    def test_perception_basic_initialization(self):
+        """Test basic Perception initialization."""
+        try:
+            from ManipulaPy.perception import Perception
+            
+            perception = Perception(vision_instance=self.mock_vision)
+            
+            self.assertIsNotNone(perception, "Perception should initialize")
+            self.assertIsNotNone(perception.logger, "Perception should have logger")
+            self.assertEqual(perception.vision, self.mock_vision, "Should store vision instance")
+            
+            print("✅ Perception basic initialization working")
+            
+        except ImportError as e:
+            self.skipTest(f"Perception module not available: {e}")
+    
+    def test_perception_requires_vision_instance(self):
+        """Test that Perception requires a valid vision instance."""
+        try:
+            from ManipulaPy.perception import Perception
+            
+            # Should raise ValueError when vision_instance is None
+            with self.assertRaises(ValueError):
+                Perception(vision_instance=None)
+            
+            print("✅ Perception properly validates vision instance")
+            
+        except ImportError as e:
+            self.skipTest(f"Perception module not available: {e}")
+    
+    def test_perception_logger_setup(self):
+        """Test that Perception sets up logging correctly."""
+        try:
+            from ManipulaPy.perception import Perception
+            
+            perception = Perception(vision_instance=self.mock_vision, logger_name="TestLogger")
+            
+            self.assertIsNotNone(perception.logger)
+            self.assertEqual(perception.logger.name, "TestLogger")
+            
+            print("✅ Perception logger setup working")
+            
+        except ImportError as e:
+            self.skipTest(f"Perception module not available: {e}")
+
+class TestPerceptionObstacleDetection(unittest.TestCase):
+    """Test Perception obstacle detection and clustering functionality."""
+    
+    def setUp(self):
+        """Set up test fixtures."""
+        self.mock_vision = Mock()
+        self.mock_vision.capture_image.return_value = (
+            np.random.randint(0, 255, (480, 640, 3), dtype=np.uint8),
+            np.random.uniform(0.1, 5.0, (480, 640)).astype(np.float32)
+        )
+        self.mock_vision.detect_obstacles.return_value = (
+            np.random.randn(10, 3),
+            np.arange(10)
+        )
+    
+    def test_detect_and_cluster_obstacles_pipeline(self):
+        """Test the full obstacle detection and clustering pipeline."""
+        try:
+            from ManipulaPy.perception import Perception
+            
[…]
