### 2025-07-03 9950a54
Issue #9
introducing more comprehensive testing
