+++ tests/conftest.py
@@ -1,10 +1,400 @@
-# tests/conftest.py
-import sys
-import types
-import os
+#!/usr/bin/env python3
+"""
+Fixed conftest.py with bug fixes for all the test failures.
+"""
 
-# Mock third-party modules that are unavailable in CI or test environments
-MOCK_MODULES = [
+import sys
+import os
+from unittest.mock import Mock, MagicMock
+import warnings
+import numpy as np
+
+# Suppress warnings during testing
+warnings.filterwarnings("ignore", category=DeprecationWarning)
+warnings.filterwarnings("ignore", category=UserWarning)
+
+# Add the package to the path
+sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+
+class MockModule:
+    """Improved mock module that handles iteration properly."""
+    
+    def __init__(self, name=None):
+        self._name = name or "MockModule"
+        
+    def __getattr__(self, name):
+        return MockModule(f"{self._name}.{name}")
+    
+    def __call__(self, *args, **kwargs):
+        return MockModule(f"{self._name}()")
+    
+    def __iter__(self):
+        return iter([])
+    
+    def __bool__(self):
+        return True
+    
+    def __len__(self):
+        return 0
+    
+    def __getitem__(self, key):
+        return MockModule(f"{self._name}[{key}]")
+
+class CuPyArrayMock:
+    """Fixed CuPy array mock with proper numpy compatibility."""
+    
+    def __init__(self, data):
+        # Use float64 to avoid casting issues
+        self._data = np.asarray(data, dtype=np.float64)
+    
+    def get(self):
+        """CuPy's method to convert GPU array to CPU (numpy) array."""
+        return self._data
+    
+    def __getattr__(self, name):
+        return getattr(self._data, name)
+    
+    def __array__(self):
+        return self._data
+    
+    def __repr__(self):
+        return f"CuPyArrayMock({self._data})"
+    
+    def __str__(self):
+        return str(self._data)
+    
+    # Fixed math operations with proper type handling
+    def __add__(self, other):
+        if hasattr(other, '_data'):
+            result = self._data + other._data
+        else:
+            result = self._data + np.asarray(other, dtype=self._data.dtype)
+        return CuPyArrayMock(result)
+    
+    def __sub__(self, other):
+        if hasattr(other, '_data'):
+            result = self._data - other._data
+        else:
+            result = self._data - np.asarray(other, dtype=self._data.dtype)
+        return CuPyArrayMock(result)
+    
+    def __mul__(self, other):
+        if hasattr(other, '_data'):
+            result = self._data * other._data
+        else:
+            result = self._data * np.asarray(other, dtype=self._data.dtype)
+        return CuPyArrayMock(result)
+    
+    def __matmul__(self, other):
+        if hasattr(other, '_data'):
+            result = self._data @ other._data
+        else:
+            result = self._data @ np.asarray(other, dtype=self._data.dtype)
+        return CuPyArrayMock(result)
+    
+    def __iadd__(self, other):
+        if hasattr(other, '_data'):
+            self._data = (self._data + other._data).astype(self._data.dtype)
+        else:
+            self._data = (self._data + np.asarray(other, dtype=self._data.dtype)).astype(self._data.dtype)
+        return self
+    
+    def __isub__(self, other):
+        if hasattr(other, '_data'):
+            self._data = (self._data - other._data).astype(self._data.dtype)
+        else:
+            self._data = (self._data - np.asarray(other, dtype=self._data.dtype)).astype(self._data.dtype)
+        return self
+    
+    def __imul__(self, other):
+        if hasattr(other, '_data'):
[…]
