+++ tests/test_vision.py
@@ -1,0 +1,605 @@
+#!/usr/bin/env python3
+"""
+test_vision_perception.py - Comprehensive tests for Vision and Perception modules
+Tests both mocked and real functionality when modules are available
+"""
+
+import unittest
+import numpy as np
+import sys
+import os
+from unittest.mock import Mock, patch, MagicMock
+
+# Add the package to path
+sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+
+class TestVisionModule(unittest.TestCase):
+    """Test the Vision module with both real and mocked dependencies."""
+    
+    def setUp(self):
+        """Set up test fixtures."""
+        self.test_image_rgb = np.random.randint(0, 255, (480, 640, 3), dtype=np.uint8)
+        self.test_image_depth = np.random.uniform(0.1, 5.0, (480, 640)).astype(np.float32)
+        
+        # Add some structure to depth image for obstacle detection
+        self.test_image_depth[200:280, 300:380] = 1.0  # Close obstacle
+        self.test_image_depth[350:400, 150:250] = 0.8  # Closer obstacle
+    
+    def test_vision_initialization_basic(self):
+        """Test basic Vision initialization."""
+        try:
+            from ManipulaPy.vision import Vision
+            
+            # Test default initialization
+            vision = Vision(use_pybullet_debug=False, show_plot=False)
+            
+            self.assertIsNotNone(vision, "Vision should initialize")
+            self.assertIsNotNone(vision.logger, "Vision should have logger")
+            self.assertIsInstance(vision.cameras, dict, "Vision should have cameras dict")
+            
+            print("✅ Vision basic initialization working")
+            
+        except ImportError as e:
+            self.skipTest(f"Vision module not available: {e}")
+    
+    def test_vision_camera_configuration(self):
+        """Test camera configuration setup."""
+        try:
+            from ManipulaPy.vision import Vision
+            
+            camera_config = {
+                "name": "test_camera",
+                "translation": [1.0, 2.0, 3.0],
+                "rotation": [10, 20, 30],
+                "fov": 45,
+                "near": 0.1,
+                "far": 10.0,
+                "intrinsic_matrix": np.eye(3, dtype=np.float32),
+                "distortion_coeffs": np.zeros(5, dtype=np.float32),
+                "use_opencv": False,
+                "device_index": 0
+            }
+            
+            vision = Vision(
+                camera_configs=[camera_config],
+                use_pybullet_debug=False,
+                show_plot=False
+            )
+            
+            self.assertIn(0, vision.cameras, "Camera should be configured")
+            self.assertEqual(vision.cameras[0]["name"], "test_camera")
+            self.assertEqual(vision.cameras[0]["fov"], 45)
+            np.testing.assert_array_equal(vision.cameras[0]["translation"], [1.0, 2.0, 3.0])
+            
+            print("✅ Vision camera configuration working")
+            
+        except ImportError as e:
+            self.skipTest(f"Vision module not available: {e}")
+    
+    def test_vision_extrinsic_matrix_computation(self):
+        """Test extrinsic matrix computation."""
+        try:
+            from ManipulaPy.vision import Vision
+            
+            vision = Vision(use_pybullet_debug=False, show_plot=False)
+            
+            # Test identity transformation
+            T = vision._make_extrinsic_matrix([0, 0, 0], [0, 0, 0])
+            np.testing.assert_array_almost_equal(T, np.eye(4), decimal=5)
+            
+            # Test pure translation
+            T = vision._make_extrinsic_matrix([1, 2, 3], [0, 0, 0])
+            expected = np.eye(4)
+            expected[:3, 3] = [1, 2, 3]
[…]
