### 2025-07-07 6f9e59d
adding Documentaion
Adding  extensive documentation addressing issue #6
