+++ docs/source/user_guide/Kinematics.rst
@@ -1,0 +1,721 @@
+Kinematics User Guide
+=====================
+
+This guide covers the **SerialManipulator** class in ManipulaPy, which provides forward kinematics, inverse kinematics, and Jacobian computations for serial robot manipulators using modern screw theory.
+
+.. contents:: **Quick Navigation**
+   :local:
+   :depth: 2
+
+What is Robot Kinematics?
+-------------------------
+
+**Robot kinematics** studies the geometry of manipulator motion without considering forces. The key problems are:
+
+- **Forward Kinematics (FK):** Given joint angles → find end-effector pose
+- **Inverse Kinematics (IK):** Given desired pose → find joint angles  
+- **Jacobian Analysis:** Relationship between joint velocities and end-effector motion
+
+ManipulaPy uses the **Product of Exponentials (PoE)** formulation with screw theory for numerically stable computations.
+
+Mathematical Foundation
+-----------------------
+
+Product of Exponentials Formula
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The forward kinematics is computed using:
+
+**Space Frame:**
+
+.. math::
+   T(\boldsymbol\theta)
+     = \exp\!\bigl([\mathcal S_{1}]\;\theta_{1}\bigr)\,
+       \exp\!\bigl([\mathcal S_{2}]\;\theta_{2}\bigr)\,
+       \cdots\,
+       \exp\!\bigl([\mathcal S_{n}]\;\theta_{n}\bigr)\,M
+
+**Body Frame:**
+
+.. math::
+   T(\boldsymbol\theta)
+     = M\,
+       \exp\!\bigl([\mathcal B_{1}]\;\theta_{1}\bigr)\,
+       \exp\!\bigl([\mathcal B_{2}]\;\theta_{2}\bigr)\,
+       \cdots\,
+       \exp\!\bigl([\mathcal B_{n}]\;\theta_{n}\bigr)
+
+where
+
+.. math::
+   [\mathcal S]
+     = \begin{bmatrix}
+         [\boldsymbol\omega]_\times & \mathbf v \\
+         \mathbf0^T                & 0
+       \end{bmatrix}
+   \quad\text{and}\quad
+   [\mathcal B]
+     = \begin{bmatrix}
+         [\boldsymbol\omega]_\times & \mathbf v \\
+         \mathbf0^T                & 0
+       \end{bmatrix}
+
+with :math:`[\boldsymbol\omega]_\times` the usual 3×3 skew‐symmetric matrix of :math:`\boldsymbol\omega`.
+
+Screw Axes
+~~~~~~~~~~
+
+Each joint's motion is encoded as a 6-vector
+
+.. math::
+   \mathcal S
+     = \begin{bmatrix}\boldsymbol\omega\\\mathbf v\end{bmatrix}\,\in\mathbb R^6
+
+with three canonical cases:
+
+- **Revolute joint** about axis :math:`\boldsymbol\omega` through point :math:`\mathbf r`  
+
+  .. math::
+     \mathbf v = \mathbf r \times \boldsymbol\omega,\qquad
+     \mathcal S
+       = \begin{bmatrix}
+           \boldsymbol\omega \\
+           \mathbf r \times \boldsymbol\omega
+         \end{bmatrix}
+
+- **Prismatic joint** translating along unit vector :math:`\mathbf v`  
+
+  .. math::
+     \mathcal S
+       = \begin{bmatrix}
+           \mathbf0 \\
+           \mathbf v
+         \end{bmatrix},\quad \|\mathbf v\|=1
+
+- **General (helical) screw** with pitch :math:`h`  
+
+  .. math::
+     \mathcal S
[…]

[…]
