### 2025-07-07 f421aa2
adding Documentaion
Adding  extensive documentation addressing issue #6
