+++ docs/source/user_guide/URDF_Processor.rst
@@ -1,0 +1,493 @@
+URDF Processor User Guide
+==========================
+
+This guide covers the URDF Processor module in ManipulaPy, which converts URDF (Unified Robot Description Format) files into SerialManipulator and ManipulatorDynamics objects for analytical robotics computations.
+
+Introduction
+-----------
+
+The URDF Processor bridges the gap between URDF robot descriptions and ManipulaPy's analytical framework. It automatically extracts kinematic and dynamic parameters from URDF files and creates the necessary objects for robotics analysis.
+
+**Key Features:**
+- Automatic parameter extraction from URDF files
+- Kinematic chain analysis and screw axis computation  
+- Inertial property extraction for dynamics
+- Joint limit handling with PyBullet integration
+- Conversion to SerialManipulator and ManipulatorDynamics objects
+
+Basic Usage
+----------
+
+Simple URDF Loading
+~~~~~~~~~~~~~~~~~~
+
+.. code-block:: python
+
+   from ManipulaPy.urdf_processor import URDFToSerialManipulator
+   
+   # Load URDF and create objects
+   processor = URDFToSerialManipulator("path/to/robot.urdf")
+   
+   # Access the created objects
+   robot = processor.serial_manipulator      # SerialManipulator instance
+   dynamics = processor.dynamics             # ManipulatorDynamics instance
+   
+   # Use the robot for computations
+   import numpy as np
+   theta = np.array([0.1, 0.3, -0.2])
+   T = robot.forward_kinematics(theta)
+   
+   print(f"End-effector position: {T[:3, 3]}")
+
+Using Built-in Models
+~~~~~~~~~~~~~~~~~~~~
+
+.. code-block:: python
+
+   from ManipulaPy.ManipulaPy_data.xarm import urdf_file as xarm_urdf
+   
+   # Load built-in xArm robot
+   processor = URDFToSerialManipulator(xarm_urdf)
+   robot = processor.serial_manipulator
+   
+   print(f"Robot has {len(robot.joint_limits)} joints")
+
+URDFToSerialManipulator Class
+----------------------------
+
+Class Constructor
+~~~~~~~~~~~~~~~~
+
+.. code-block:: python
+
+   URDFToSerialManipulator(urdf_name, use_pybullet_limits=True)
+
+**Parameters:**
+- ``urdf_name`` (str): Path to the URDF file
+- ``use_pybullet_limits`` (bool): Extract joint limits from PyBullet simulation
+
+**Attributes:**
+- ``serial_manipulator``: SerialManipulator object for kinematics
+- ``dynamics``: ManipulatorDynamics object for dynamics  
+- ``robot_data``: Dictionary containing extracted parameters
+- ``urdf_name``: Path to the loaded URDF file
+- ``robot``: Loaded URDF object from urchin library
+
+Extracted Parameters
+~~~~~~~~~~~~~~~~~~~
+
+The ``robot_data`` dictionary contains:
+
+.. code-block:: python
+
+   processor = URDFToSerialManipulator("robot.urdf")
+   data = processor.robot_data
+   
+   print(f"Degrees of freedom: {data['actuated_joints_num']}")
+   print(f"Home configuration shape: {data['M'].shape}")          # (4, 4)
+   print(f"Space screw axes shape: {data['Slist'].shape}")        # (6, n)
+   print(f"Body screw axes shape: {data['Blist'].shape}")         # (6, n)  
+   print(f"Number of inertia matrices: {len(data['Glist'])}")     # n links
+
+Core Methods
+-----------
+
+load_urdf()
+~~~~~~~~~~
+
+Extracts kinematic and dynamic parameters from the URDF file:
+
+.. code-block:: python
+
[…]
