+++ docs/source/user_guide/Trajectory_Planning.rst
@@ -1,0 +1,1577 @@
+Trajectory Planning User Guide
+===============================
+
+This guide covers the trajectory planning capabilities in ManipulaPy, including joint-space and Cartesian-space trajectory generation, dynamics integration, and collision avoidance using CUDA acceleration.
+
+Introduction
+-----------
+
+The TrajectoryPlanning class provides comprehensive trajectory generation and execution capabilities for robotic manipulators. It combines kinematic trajectory planning with dynamic analysis and collision avoidance to generate feasible, smooth robot motions.
+
+**Key Features:**
+- Joint-space trajectory generation with cubic/quintic time scaling
+- Cartesian-space trajectory planning for end-effector paths
+- CUDA-accelerated computations for real-time performance
+- Integrated dynamics analysis (forward/inverse dynamics)
+- Collision detection and avoidance using potential fields
+- Support for various trajectory optimization objectives
+
+Mathematical Background
+~~~~~~~~~~~~~~~~~~~~~~~
+
+This section summarizes the key mathematical constructs behind joint-space and Cartesian trajectory generation, as well as obstacle avoidance via potential fields.
+
+Joint-Space Time Scaling
+------------------------
+
+Given start :math:`\boldsymbol\theta_{0}` and end :math:`\boldsymbol\theta_{f}`, total duration :math:`T`, define a scalar time-scaling function
+
+.. math::
+
+   s = \frac{t}{T}, 
+   \quad s\in[0,1].
+
+Common choices:
+
+- **Cubic** (zero end-velocity):
+  
+  .. math::
+
+     \sigma_{3}(s)
+       = 3s^{2} - 2s^{3},
+
+- **Quintic** (zero end-velocity & zero end-acceleration):
+
+  .. math::
+
+     \sigma_{5}(s)
+       = 10s^{3} - 15s^{4} + 6s^{5}.
+
+Joint-space trajectory:
+
+.. math::
+
+   \boldsymbol\theta(t)
+     = \boldsymbol\theta_{0}
+       + \bigl(\boldsymbol\theta_{f} - \boldsymbol\theta_{0}\bigr)\,\sigma_{m}(s),
+
+with
+
+- velocity:  
+  :math:`\dot{\boldsymbol\theta}(t)
+     = \frac{1}{T}\bigl(\boldsymbol\theta_{f}-\boldsymbol\theta_{0}\bigr)\,\sigma_{m}'(s)`
+
+- acceleration:  
+  :math:`\ddot{\boldsymbol\theta}(t)
+     = \frac{1}{T^{2}}\bigl(\boldsymbol\theta_{f}-\boldsymbol\theta_{0}\bigr)\,\sigma_{m}''(s)`
+
+where :math:`m=3` or :math:`5` for cubic/quintic.
+
+Cartesian-Space Interpolation
+-----------------------------
+
+Given start/end end-effector poses :math:`X_{0},X_{f}\in SE(3)`, define the relative transform
+
+.. math::
+
+   \Delta X = X_{0}^{-1}X_{f}
+   = \exp\bigl(\,[\Xi]\,\bigr),
+
+with twist :math:`\Xi\in\mathfrak{se}(3)` given by the matrix logarithm.  Then
+
+.. math::
+
+   X(t)
+     = X_{0}\,\exp\!\bigl([\Xi]\;\sigma_{m}(s)\bigr).
+
+Extract :
+
+- position:  the translational part of :math:`X(t)`  
+- orientation:  the rotational part via Rodrigues’ formula on :math:`[\Xi]`  
+
+Dynamics-Aware Torque Computation
+---------------------------------
+
+Once joint positions, velocities and accelerations are known, the required torques along the trajectory follow by inverse dynamics:
+
+.. math::
+
+   \boldsymbol\tau(t)
[…]
