+++ ManipulaPy/cuda_kernels.py
@@ -1,43 +1,46 @@
 #!/usr/bin/env python3
 # SPDX-License-Identifier: AGPL-3.0-or-later
 """
-CUDA Kernels Module - ManipulaPy
+Optimized CUDA Kernels Module - ManipulaPy
 
-This module provides CUDA-accelerated functions for trajectory planning and dynamics
+This module provides highly optimized CUDA-accelerated functions for trajectory planning and dynamics
 computation. All CUDA functionality is optional and gracefully degrades to CPU implementations.
+
+Optimizations include:
+- 2D grid parallelization across time and joint indices
+- Shared memory utilization for 6x6 matrix operations
+- Register-optimized computations avoiding spilling
+- Fused kernels to reduce memory bandwidth requirements
+- Adaptive launch configurations for optimal occupancy
+- Pinned memory transfers for improved PCIe bandwidth
+- Optimized scalar time-scaling with shared memory
+- Fast math operations with rsqrt for improved performance
 
 Copyright (c) 2025 Mohamed Aboelnar
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
-
-This file is part of ManipulaPy.
-
-ManipulaPy is free software: you can redistribute it and/or modify
-it under the terms of the GNU Affero General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-ManipulaPy is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU Affero General Public License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with ManipulaPy. If not, see <https://www.gnu.org/licenses/>.
 """
 
 import numpy as np
 import warnings
+import math
+import os
+from typing import Tuple, List, Optional
+from numba import config as _nb_cfg
+_nb_cfg.CUDA_CACHE_SIZE = "1024"  # cache compiled kernels
+
+# Environment toggle for strict FP
+FAST_MATH = bool(int(os.getenv("MANIPULAPY_FASTMATH", "1")))
 
 # Optional CUDA imports with graceful fallback
 try:
-    from numba import cuda, float32
+    from numba import cuda, float32, int32, float16
     CUDA_AVAILABLE = True
 except ImportError:
     CUDA_AVAILABLE = False
     # Create mock objects to prevent import errors
     class MockCuda:
         @staticmethod
-        def jit(func):
+        def jit(func=None, device=False, inline=False, fastmath=False):
             """Mock decorator for when CUDA is not available"""
             def wrapper(*args, **kwargs):
                 raise RuntimeError(
@@ -46,34 +49,56 @@ except ImportError:
                     "pip install ManipulaPy[gpu-cuda12]  # For CUDA 12.x\n"
                     "Ensure you have NVIDIA GPU drivers and CUDA toolkit installed."
                 )
-            return wrapper
+            return wrapper if func is None else wrapper(func)
         
         @staticmethod
         def grid(dim):
-            """Mock grid function"""
             return 0
         
         @staticmethod
         def device_array(*args, **kwargs):
-            """Mock device array function"""
[…]
