### 2025-07-09 ef0286c
Cuda Kernels Optmization
optimization of the cuda kernels and the path planning
