+++ ManipulaPy/path_planning.py
@@ -1,32 +1,29 @@
 #!/usr/bin/env python3
 # SPDX-License-Identifier: AGPL-3.0-or-later
 """
-Path Planning Module - ManipulaPy
+Optimized Path Planning Module - ManipulaPy
 
-This module provides trajectory planning capabilities including joint space and Cartesian
-space trajectory generation with CUDA acceleration and collision avoidance.
+This module provides highly optimized trajectory planning capabilities including joint space 
+and Cartesian space trajectory generation with CUDA acceleration and collision avoidance.
+
+Key optimizations:
+- Adaptive grid sizing for optimal GPU occupancy
+- Memory pooling to reduce allocation overhead
+- Batch processing for multiple trajectories
+- Fused kernels to minimize memory bandwidth
+- Intelligent fallback to CPU when beneficial
+- 2D parallelization for better GPU utilization
 
 Copyright (c) 2025 Mohamed Aboelnar
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
-
-This file is part of ManipulaPy.
-
-ManipulaPy is free software: you can redistribute it and/or modify
-it under the terms of the GNU Affero General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-ManipulaPy is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU Affero General Public License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with ManipulaPy. If not, see <https://www.gnu.org/licenses/>.
 """
-from numba import cuda
+
+from numba import njit, prange
 import numpy as np
 import matplotlib.pyplot as plt
+import warnings
+import time
+from typing import Optional 
 from .utils import (
     TransToRp,
     MatrixLog3,
@@ -35,111 +32,437 @@ from .utils import (
     QuinticTimeScaling,
 )
 from .cuda_kernels import (
-    trajectory_kernel,
-    inverse_dynamics_kernel,
-    forward_dynamics_kernel,
-    cartesian_trajectory_kernel,
+    CUDA_AVAILABLE,
+    check_cuda_availability,
+    make_1d_grid,
+    make_2d_grid,
+    get_gpu_properties,
+    optimized_trajectory_generation,
+    optimized_potential_field,
+    optimized_batch_trajectory_generation,
+    get_cuda_array,
+    return_cuda_array,
+    profile_start,
+    profile_stop,
+    _best_2d_config,
+    _h2d_pinned,
 )
+
+# Import CUDA functions only if available
+if CUDA_AVAILABLE:
+    from numba import cuda
+    from .cuda_kernels import (
+        trajectory_kernel,
+        inverse_dynamics_kernel,
+        forward_dynamics_kernel,
+        cartesian_trajectory_kernel,
+        fused_potential_gradient_kernel,
+        batch_trajectory_kernel,
+    )
+else:
+    # Create dummy functions for when CUDA is not available
+    def trajectory_kernel(*args, **kwargs):
+        raise RuntimeError("CUDA not available")
+    def inverse_dynamics_kernel(*args, **kwargs):
+        raise RuntimeError("CUDA not available")
+    def forward_dynamics_kernel(*args, **kwargs):
+        raise RuntimeError("CUDA not available")
+    def cartesian_trajectory_kernel(*args, **kwargs):
+        raise RuntimeError("CUDA not available")
+    def fused_potential_gradient_kernel(*args, **kwargs):
+        raise RuntimeError("CUDA not available")
[…]
