+++ ManipulaPy/cuda_kernels.py
@@ -112,7 +112,17 @@ except ImportError:
 float_t = float16 if os.getenv("MANIPULAPY_USE_FP16", "0") == "1" else float32
 
 def check_cuda_availability():
-    """Check if CUDA is available and provide helpful error messages."""
+    """
+        Check if CUDA is available and provide helpful diagnostic information.
+        
+        Checks the availability of CUDA and emits a user-friendly warning with
+        installation instructions if GPU acceleration is not possible. Useful for
+        gracefully handling scenarios where GPU support cannot be initialized.
+        
+        Returns:
+            bool: True if CUDA is available, False otherwise. Triggers a UserWarning
+            with detailed setup guidance when CUDA is not detected.
+        """
     if not CUDA_AVAILABLE:
         warnings.warn(
             "CUDA not available. GPU-accelerated trajectory planning will not work.\n"
@@ -127,7 +137,13 @@ def check_cuda_availability():
     return CUDA_AVAILABLE
 
 def check_cupy_availability():
-    """Check if CuPy is available for GPU array operations."""
+    """
+        Check if CuPy is available for GPU array operations.
+        
+        Returns:
+            bool: True if CuPy is available, False otherwise. Raises a UserWarning
+            if CuPy is not available, providing installation instructions.
+        """
     if not CUPY_AVAILABLE:
         warnings.warn(
             "CuPy not available. Some GPU array operations will not work.\n"
@@ -156,7 +172,20 @@ def _h2d_pinned(arr):
 
 # Optimized launch configuration helpers
 def make_1d_grid(size, threads=256):
-    """Create 1D grid configuration with optimal occupancy."""
+    """
+        Create a 1D grid configuration for CUDA kernel launch with optimal thread and block sizing.
+        
+        Calculates the appropriate number of blocks and threads for a CUDA kernel based on the problem size.
+        Ensures minimum occupancy by adjusting thread count for small datasets and rounding to power of 2.
+        
+        Args:
+            size (int): Total number of elements or work items to process
+            threads (int, optional): Desired number of threads per block. Defaults to 256.
+        
+        Returns:
+            Tuple[Tuple[int], Tuple[int]]: A tuple containing (blocks,), (threads,) for kernel launch configuration
+        """
+
     if size <= 0:
         return (1,), (1,)
     
@@ -169,8 +198,22 @@ def make_1d_grid(size, threads=256):
 
 def make_2d_grid(N: int,
                  num_joints: int,
-                 block_size: Tuple[int, int] = (64, 8)):
+                 block_size: Tuple[int, int] = (128, 8)):
     """
+        Compute optimal 2D grid configuration for CUDA kernel launch.
+        
+        Calculates grid and block dimensions that ensure:
+            • Thread-block dimensions are powers of two (≥ 4)
+            • Block thread count respects device limits
+            • Provides good GPU occupancy by targeting at least 2x streaming multiprocessors
+        
+        Args:
+            N (int): First dimension of problem space
+            num_joints (int): Second dimension of problem space
+            block_size (Tuple[int, int], optional): Initial suggested block dimensions. Defaults to (128, 8).
+        
+        Returns:
+            Tuple[Tuple[int, int], Tuple[int, int]]: ((blocks_x, blocks_y), (threads_x, threads_y))
     Choose (grid, block) for a 2-D kernel so that
       • thread-block dimensions are powers of two (≥ 4)
       • block.x * block.y  ≤  device.MAX_THREADS_PER_BLOCK
@@ -225,7 +268,18 @@ def make_2d_grid(N: int,
     return (blocks_x, blocks_y), (threads_x, threads_y)
 
 def get_gpu_properties():
-    """Get GPU properties for optimization decisions."""
+    """
[…]
