+++ docs/source/api/cuda_kernels.rst
@@ -4,56 +4,20 @@
 CUDA Kernels API Reference
 ==========================
 
-This page documents **ManipulaPy.cuda_kernels**, the module providing CUDA-accelerated functions for trajectory planning and dynamics computation.
-
-.. tip::
-   All CUDA functionality is optional and gracefully degrades to CPU implementations when GPU support is unavailable.
-
------------------
-Quick Navigation
------------------
-
-.. contents::
-   :local:
-   :depth: 2
-
-----------------
-Module Overview
-----------------
-
 .. currentmodule:: ManipulaPy.cuda_kernels
 
-The CUDA kernels module provides high-performance GPU-accelerated functions for:
+This module provides CUDA-accelerated functions for trajectory planning and dynamics computation with automatic CPU fallback.
 
-- **Trajectory Generation**: Cubic and quintic time scaling with parallel computation
-- **Dynamics Computation**: Inverse and forward dynamics for trajectory batches
-- **Potential Fields**: Attractive and repulsive potential field calculations
-- **Automatic Fallback**: CPU implementations when CUDA is unavailable
-
-.. note::
-   **GPU Requirements**: CUDA kernels require:
-   
-   - NVIDIA GPU with compute capability 3.5+
-   - CUDA toolkit 11.0+ or 12.0+
-   - Installation: ``pip install ManipulaPy[gpu-cuda11]`` or ``pip install ManipulaPy[gpu-cuda12]``
-
------------------
+--------------------
 Availability Check
------------------
+--------------------
 
 .. autofunction:: check_cuda_availability
 
-   Check if CUDA is available and provide helpful error messages.
+   Check if CUDA is available and provide helpful diagnostic information.
 
    **Returns:**
      - **bool** -- True if CUDA is available, False otherwise
-
-   **Example:**
-     >>> from ManipulaPy.cuda_kernels import check_cuda_availability
-     >>> if check_cuda_availability():
-     ...     print("CUDA acceleration available")
-     ... else:
-     ...     print("Falling back to CPU implementation")
 
 .. autofunction:: check_cupy_availability
 
@@ -62,127 +26,135 @@ Availability Check
    **Returns:**
      - **bool** -- True if CuPy is available, False otherwise
 
------------------
-CUDA Kernels
------------------
+.. autofunction:: get_gpu_properties
+
+   Retrieve current CUDA device properties for kernel optimization and resource allocation.
+
+   **Returns:**
+     - **dict or None** -- GPU device properties including multiprocessor count, memory limits, etc.
+
+-------------------
+Core CUDA Kernels
+-------------------
 
 Trajectory Kernels
-------------------
+--------------------
 
 .. autofunction:: trajectory_kernel
 
-   CUDA kernel to compute joint trajectories with cubic or quintic time scaling.
+   CUDA kernel for generating joint trajectory points with time-scaling.
 
    **Parameters:**
-     - **thetastart** (*cuda.device_array*) -- Starting joint angles (n,)
-     - **thetaend** (*cuda.device_array*) -- Ending joint angles (n,)
-     - **traj_pos** (*cuda.device_array*) -- Output positions (N, n)
-     - **traj_vel** (*cuda.device_array*) -- Output velocities (N, n)
-     - **traj_acc** (*cuda.device_array*) -- Output accelerations (N, n)
+     - **thetastart** (*cuda.device_array*) -- Starting joint angles
[…]
