### 2025-07-13 2e018f7
Cuda Kernel Optmization
optimizing the workflow o the CUDA optimization be improving the memory management
