+++ ManipulaPy/cuda_kernels.py
@@ -31,23 +31,84 @@ _nb_cfg.CUDA_CACHE_SIZE = "1024"  # cache compiled kernels
 # Environment toggle for strict FP
 FAST_MATH = bool(int(os.getenv("MANIPULAPY_FASTMATH", "1")))
 
-# Optional CUDA imports with graceful fallback
-try:
-    from numba import cuda, float32, int32, float16
-    CUDA_AVAILABLE = True
-except ImportError:
-    CUDA_AVAILABLE = False
-    # Create mock objects to prevent import errors
+# IMPROVED CUDA DETECTION LOGIC WITH BETTER ERROR HANDLING
+def _detect_cuda_capability():
+    """
+    Comprehensive CUDA detection that handles various failure modes gracefully.
+    
+    Returns:
+        tuple: (cuda_available, cuda_module, float32, int32, error_message)
+    """
+    try:
+        # Step 1: Try importing numba.cuda - avoid float16 completely
+        from numba import cuda, float32, int32
+        
+        # Step 2: Check if CUDA is available in numba (this might fail with CUDA_ERROR_NO_DEVICE)
+        try:
+            cuda_available = cuda.is_available()
+            if not cuda_available:
+                return False, None, None, None, "numba.cuda.is_available() returned False - likely driver/GPU issue"
+        except Exception as e:
+            # Common issues: CUDA_ERROR_NO_DEVICE, driver problems, etc.
+            return False, None, None, None, f"numba.cuda.is_available() failed: {e}"
+        
+        # Step 3: Try to detect actual GPU devices
+        try:
+            devices = cuda.list_devices()
+            if not devices:
+                return False, None, None, None, "No CUDA devices detected by numba"
+        except Exception as e:
+            return False, None, None, None, f"Failed to list CUDA devices: {e}"
+        
+        # Step 4: Try a simple GPU operation to verify functionality
+        try:
+            # Test basic GPU memory allocation
+            test_array = cuda.device_array(10, dtype=np.float32)
+            
+            # Test basic kernel compilation (but don't execute)
+            @cuda.jit
+            def test_kernel(arr):
+                idx = cuda.grid(1)
+                if idx < arr.shape[0]:
+                    arr[idx] = idx
+            
+            # Try to get device properties
+            current_device = cuda.get_current_device()
+            _ = current_device.MULTIPROCESSOR_COUNT
+            
+            # Clean up test array
+            del test_array
+            
+            return True, cuda, float32, int32, None
+            
+        except Exception as e:
+            return False, None, None, None, f"CUDA functionality test failed: {e}"
+            
+    except ImportError as e:
+        return False, None, None, None, f"numba.cuda import failed: {e}"
+    except Exception as e:
+        return False, None, None, None, f"Unexpected error in CUDA detection: {e}"
+
+# Perform the detection - no float16!
+CUDA_AVAILABLE, cuda, float32, int32, _cuda_error = _detect_cuda_capability()
+
+# Set up the rest based on detection results
+if not CUDA_AVAILABLE:
+    # CUDA not available - create mock objects
+    _cuda_error_msg = _cuda_error or "CUDA not available"
+    
     class MockCuda:
         @staticmethod
         def jit(func=None, device=False, inline=False, fastmath=False):
             """Mock decorator for when CUDA is not available"""
             def wrapper(*args, **kwargs):
                 raise RuntimeError(
-                    "CUDA functionality not available. Install CUDA support:\n"
-                    "pip install ManipulaPy[gpu-cuda11]  # For CUDA 11.x\n"
-                    "pip install ManipulaPy[gpu-cuda12]  # For CUDA 12.x\n"
-                    "Ensure you have NVIDIA GPU drivers and CUDA toolkit installed."
[…]
