### 2025-07-13 00a3d93
Cuda Kernel Optmization
optimizing the workflow o the CUDA optimization be improving the memory management
