### 2025-07-13 89314be
Examples
Adding more extensive examples
