+++ Examples/EX_traj.py
@@ -1,191 +1,0 @@
-#!/usr/bin/env python3
-"""
-Example: joint-space trajectory generation + combined JS + Cartesian PF avoidance
---------------------------------------------------------------------
-Run as:
-    python example_trajectory_planning.py
-Produces:
-    - joint_trajectory.png        # original joint plots
-    - ee_path.png                 # TCP path with obstacles & avoidance
-    - torque_profile.png          # torques for avoidance trajectory
-"""
-import os
-import numpy as np
-import matplotlib
-matplotlib.use("Agg")
-import matplotlib.pyplot as plt
-
-from ManipulaPy.ManipulaPy_data.xarm import urdf_file as xarm_urdf_file
-from ManipulaPy.urdf_processor import URDFToSerialManipulator
-from ManipulaPy.path_planning import TrajectoryPlanning
-from ManipulaPy.utils import TransInv
-
-# ------------------------------------------------------------------
-# 1.  Robot model
-# ------------------------------------------------------------------
-urdf_proc          = URDFToSerialManipulator(xarm_urdf_file)
-serial_manipulator = urdf_proc.serial_manipulator
-dynamics           = urdf_proc.dynamics
-joint_limits       = serial_manipulator.joint_limits      # list[(min,max)]
-torque_limits      = [(-np.inf, np.inf)] * len(joint_limits)
-
-planner = TrajectoryPlanning(
-    serial_manipulator,
-    xarm_urdf_file,
-    dynamics,
-    joint_limits,
-    torque_limits
-)
-pf = planner.potential_field
-
-# ------------------------------------------------------------------
-# 2.  Define motion
-# ------------------------------------------------------------------
-thetastart = np.deg2rad([  0, -90,  90,   0,  90,   0])
-thetaend   = np.deg2rad([ 45, -45,  60,   0, 110,  45])
-Tf, N, method = 10.0, 200, 5
-
-traj = planner.joint_trajectory(thetastart, thetaend, Tf, N, method)
-
-# ------------------------------------------------------------------
-# 3.  Plot joint position/velocity/acceleration (original)
-# ------------------------------------------------------------------
-planner.plot_trajectory(
-    trajectory_data=traj,
-    Tf=Tf,
-    title="xArm joint trajectory (quintic)",
-    labels=[f"Joint {i+1}" for i in range(len(joint_limits))]
-)
-plt.gcf().savefig("joint_trajectory.png", dpi=400)
-
-# ------------------------------------------------------------------
-# 4.  Define joint-space obstacles
-# ------------------------------------------------------------------
-obs_degrees = [
-    [ 30, -30,  45,   0,  60,   0],
-    [ 15,  15,  90,   0,  45,   0],
-]
-obs_js = [np.deg2rad(q) for q in obs_degrees]
-
-# ------------------------------------------------------------------
-# 4b. Cartesian obstacles *relative to robot base*
-# ------------------------------------------------------------------
-# If your robot base in the world is not at the origin, set world_T_base accordingly:
-world_T_base = np.eye(4, dtype=np.float32)
-# e.g. world_T_base[:3,3] = [0.1, 0.0, 0.2]
-
-# define in world coordinates
-cart_obs_world = np.array([
-    [1.20,  0.40, 0.80],
-    [0.40,  0.20, 0.30],
-], dtype=np.float32)
-
-# transform into base frame
-h       = np.hstack((cart_obs_world, np.ones((len(cart_obs_world),1),dtype=np.float32)))
[…]
