+++ Examples/intermediate_examples/control_comparison_intermediate_demo.py
@@ -1,0 +1,30 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: AGPL-3.0-or-later
+"""
+Intermediate Control Comparison Demo: Multiple Control Strategies Analysis
+
+This example demonstrates comparison of different control strategies including
+PID, computed torque, adaptive, and robust control methods.
+
+Usage:
+    python control_comparison_intermediate_demo.py
+
+Author: ManipulaPy Development Team
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from ManipulaPy.control import ManipulatorController
+
+def main():
+    """Demonstrate intermediate control comparison operations."""
+    print("=== ManipulaPy: Intermediate Control Comparison Demo ===")
+    
+    # TODO: Implement control comparison demonstration
+    # - Multiple control strategies
+    # - Performance comparison
+    # - Stability analysis
+    print("🎮 Intermediate control comparison functionality demonstration")
+
+if __name__ == "__main__":
+    main()
