+++ Examples/advanced_examples/optimal_control_advanced_demo.py
@@ -1,0 +1,30 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: AGPL-3.0-or-later
+"""
+Advanced Optimal Control Demo: State-of-the-Art Control Strategies
+
+This example demonstrates advanced control methods including optimal control,
+model predictive control, and adaptive control with parameter estimation.
+
+Usage:
+    python optimal_control_advanced_demo.py
+
+Author: ManipulaPy Development Team
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from ManipulaPy.control import ManipulatorController
+
+def main():
+    """Demonstrate advanced optimal control operations."""
+    print("=== ManipulaPy: Advanced Optimal Control Demo ===")
+    
+    # TODO: Implement optimal control demonstration
+    # - Optimal control strategies
+    # - Model predictive control
+    # - Adaptive parameter estimation
+    print("🎮 Advanced optimal control functionality demonstration")
+
+if __name__ == "__main__":
+    main()
