+++ Examples/advanced_examples/real_robot_integration_advanced_demo.py
@@ -1,0 +1,29 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: AGPL-3.0-or-later
+"""
+Advanced Real Robot Integration Demo: Hardware Interface and Control
+
+This example demonstrates real robot integration including hardware communication,
+real-time control, and safety monitoring for physical robot systems.
+
+Usage:
+    python real_robot_integration_advanced_demo.py [--robot-ip IP]
+
+Author: ManipulaPy Development Team
+"""
+
+import numpy as np
+from ManipulaPy.control import ManipulatorController
+
+def main():
+    """Demonstrate advanced real robot integration operations."""
+    print("=== ManipulaPy: Advanced Real Robot Integration Demo ===")
+    
+    # TODO: Implement real robot integration demonstration
+    # - Hardware communication
+    # - Real-time control
+    # - Safety monitoring
+    print("🤖 Advanced real robot integration functionality demonstration")
+
+if __name__ == "__main__":
+    main()

[…]
