+++ Examples/intermediate_examples/singularity_analysis_intermediate_demo.py
@@ -1,0 +1,1806 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: AGPL-3.0-or-later
+"""
+Intermediate Singularity Analysis Demo - ManipulaPy
+
+This demo showcases comprehensive singularity analysis capabilities including:
+- Singularity detection and classification
+- Manipulability ellipsoid visualization
+- Condition number analysis
+- Workspace boundary analysis
+- Singularity avoidance strategies
+- Real-time singularity monitoring
+- Dexterity measures and optimization
+- GPU-accelerated workspace sampling
+
+Copyright (c) 2025 Mohamed Aboelnar
+Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from mpl_toolkits.mplot3d import Axes3D
+import time
+import logging
+from pathlib import Path
+import matplotlib
+matplotlib.use('TkAgg')
+
+# ManipulaPy imports
+try:
+    from ManipulaPy.kinematics import SerialManipulator
+    from ManipulaPy.dynamics import ManipulatorDynamics
+    from ManipulaPy.singularity import Singularity
+    from ManipulaPy.urdf_processor import URDFToSerialManipulator
+    from ManipulaPy.cuda_kernels import CUDA_AVAILABLE, check_cuda_availability
+    from ManipulaPy import utils
+    from ManipulaPy.ManipulaPy_data.xarm import urdf_file
+except ImportError as e:
+    print(f"Error importing ManipulaPy modules: {e}")
+    print("Please ensure ManipulaPy is properly installed.")
+    exit(1)
+
+# Optional GPU acceleration
+try:
+    import cupy as cp
+    from numba import cuda
+    from numba.cuda.random import create_xoroshiro128p_states, xoroshiro128p_uniform_float32
+    GPU_AVAILABLE = True
+    print("✅ GPU acceleration available")
+except ImportError:
+    GPU_AVAILABLE = False
+    print("⚠️ GPU acceleration not available, using CPU only")
+
+# Scientific computing imports
+try:
+    from scipy.linalg import svd, qr
+    from scipy.spatial import ConvexHull
+    from scipy.optimize import minimize
+    from sklearn.cluster import DBSCAN
+    SCIPY_AVAILABLE = True
+except ImportError:
+    SCIPY_AVAILABLE = False
+    print("⚠️ SciPy/sklearn not available, some features will be limited")
+
+# Set up logging
+logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+logger = logging.getLogger(__name__)
+
+class IntermediateSingularityDemo:
+    """
+    Demonstrates advanced singularity analysis techniques for robotic manipulators.
+    """
+    
+    def __init__(self, use_simple_robot=False):
+        """
+        Initialize the singularity analysis demo.
+        
+        Args:
+            use_simple_robot: If True, creates a simple 3-DOF robot. 
+                             If False (default), uses the built-in XArm robot.
+        """
+        self.use_simple_robot = use_simple_robot
+        self.setup_robot()
+        self.setup_singularity_analyzer()
+        
+    def setup_robot(self):
+        """Set up the robot model (either XArm or simple)."""
+        if self.use_simple_robot:
+            self.setup_simple_robot()
+        else:
+            self.setup_xarm_robot()
+            
+    def setup_xarm_robot(self):
+        """Load the built-in XArm robot from ManipulaPy data."""
+        logger.info("Setting up XArm robot from built-in data...")
+        
+        try:
+            # Load XArm robot using built-in URDF
[…]
