+++ docs/docs/requirements.txt
@@ -1,16 +1,0 @@
-# Minimal requirements for Read the Docs
-# Core documentation tools
-sphinx>=7.1.0
-sphinx-rtd-theme>=1.3.0
-sphinx-autodoc-typehints>=1.24.0
-sphinx-copybutton>=0.5.2
-myst-parser>=2.0.0
-sphinx-design>=0.5.0
-
-# Core dependencies for autodoc (minimal versions)
-numpy>=1.21.0
-scipy>=1.7.0
-matplotlib>=3.5.0
-
-# Mock heavy dependencies in conf.py instead of installing them
-# This prevents build failures on Read the Docs
