+++ docs/requirements.txt
@@ -1,14 +1,16 @@
-numpy>=1.19.2
-scipy>=1.5.2
-pybullet>=3.0.6
-pycuda>=2021.1
-trimesh>=3.9.14
-urchin>=0.0.27
-numba>=0.55
-matplotlib>=3.3
-scikit-learn>=1.0
-opencv-python>=4.5
-ultralytics>=8.0  # For YOLO-based perception
-cupy>=9.0         # CUDA-accelerated arrays/kernels
-torch>=1.8.0      # Required for YOLO (PyTorch-based)
+# Documentation build requirements for Read the Docs
+# Core documentation tools
+sphinx>=7.1.0
+sphinx-rtd-theme>=1.3.0
+sphinx-autodoc-typehints>=1.24.0
+sphinx-copybutton>=0.5.2
+myst-parser>=2.0.0
+sphinx-design>=0.5.0
 
+# Core dependencies for autodoc (minimal versions)
+numpy>=1.21.0
+scipy>=1.7.0
+matplotlib>=3.5.0
+
+# Mock heavy dependencies in conf.py instead of installing them
+# This prevents build failures on Read the Docs
