+++ docs/source/conf.py
@@ -5,6 +5,35 @@
 import os
 import sys
 import datetime
+from unittest.mock import MagicMock
+
+# ── CRITICAL: Mock heavy dependencies FIRST ─────────────
+class Mock(MagicMock):
+    @classmethod
+    def __getattr__(cls, name):
+        return MagicMock()
+
+# Mock all heavy dependencies before any other imports
+MOCK_MODULES = [
+    # Deep Learning / GPU
+    "torch", "cupy", "numba", "numba.cuda", "numba.cuda.random", "numba.config",
+    "pycuda", "pycuda.driver", "pycuda.autoinit", "tensorflow",
+    
+    # Robotics / Simulation
+    "pybullet", "pybullet_data", "urchin", "urchin.urdf",
+    
+    # Computer Vision
+    "cv2", "ultralytics", "opencv-python",
+    
+    # Scientific Computing (keep matplotlib for docs)
+    "sklearn", "sklearn.cluster", "scipy", "scipy.spatial", "scipy.linalg",
+    
+    # Optional dependencies
+    "plotly", "seaborn", "PIL", "Pillow",
+]
+
+for mod_name in MOCK_MODULES:
+    sys.modules[mod_name] = Mock()
 
 # ── Path handling ────────────────────────────────────────
 DOCS_DIR = os.path.abspath(os.path.dirname(__file__))
@@ -26,20 +55,7 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.intersphinx",
     "sphinx.ext.mathjax",
-    
-    # Enhanced documentation features
-    "sphinx.ext.autosectionlabel",   # every heading gets an explicit label
     "sphinx.ext.githubpages",        # .nojekyll file for GitHub Pages
-    "sphinx.ext.extlinks",           # external link shortcuts
-    "sphinx.ext.todo",               # TODO directives
-    "sphinx.ext.ifconfig",           # conditional content
-    
-    # Graphviz for block diagrams
-    "sphinx.ext.graphviz",           # dot/graphviz diagram support
-    
-    # Third-party extensions (check availability)
-    "myst_parser",                   # markdown support
-    "sphinx_design",                 # nice HTML components (<sd-*)
 ]
 
 # ── Safely add optional extensions ──────────────────────
@@ -54,24 +70,45 @@ def try_add_extension(ext_name):
         print(f"⚠️  Optional extension not available: {ext_name}")
         return False
 
-# Try to add optional extensions
+# Try to add optional extensions - be conservative for RTD
 _optional_extensions = {
+    "myst_parser": try_add_extension("myst_parser"),
     "sphinx_copybutton": try_add_extension("sphinx_copybutton"),
-    "sphinx_tabs": try_add_extension("sphinx_tabs"), 
-    "sphinx_togglebutton": try_add_extension("sphinx_togglebutton"),
-    "sphinxext.opengraph": try_add_extension("sphinxext.opengraph"),
-    "sphinx_sitemap": try_add_extension("sphinx_sitemap"),
+    "sphinx_design": try_add_extension("sphinx_design"),
 }
 
-# ── Theme selection (furo -> pydata -> RTD -> default) ──
+# Only add these if we're not on RTD (they may not be available)
+if not os.environ.get('READTHEDOCS'):
+    _optional_extensions.update({
+        "sphinx.ext.autosectionlabel": try_add_extension("sphinx.ext.autosectionlabel"),
+        "sphinx.ext.extlinks": try_add_extension("sphinx.ext.extlinks"), 
+        "sphinx.ext.todo": try_add_extension("sphinx.ext.todo"),
[…]
