+++ .readthedocs.yaml
@@ -1,20 +1,29 @@
+# Read the Docs configuration file for ManipulaPy
+# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
+
 version: 2
 
+# Set the OS, Python version and other tools you might need
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.8"
+
+# Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/source/conf.py
+  fail_on_warning: false
 
+# Optionally build your docs in additional formats such as PDF and ePub
+formats:
+  - pdf
+  - epub
+
+# Python requirements and installation
 python:
-  version: 3.8
   install:
     - requirements: docs/requirements.txt
     - method: pip
       path: .
-
-artifacts:
-  - path: docs/build/html
-    destination: html
-
-build:
-  image: latest
-  use-virtualenv: true
-
+      extra_requirements:
+        - docs
