+++ docs/requirements.txt
@@ -1,16 +1,22 @@
-# Documentation build requirements for Read the Docs
-# Core documentation tools
-sphinx>=7.1.0
-sphinx-rtd-theme>=1.3.0
-sphinx-autodoc-typehints>=1.24.0
-sphinx-copybutton>=0.5.2
-myst-parser>=2.0.0
-sphinx-design>=0.5.0
+# Core documentation dependencies
+sphinx>=4.0.0,<7.0.0
+sphinx-rtd-theme>=1.0.0,<2.0.0
+myst-parser>=0.17.0,<3.0.0
 
-# Core dependencies for autodoc (minimal versions)
-numpy>=1.21.0
-scipy>=1.7.0
-matplotlib>=3.5.0
+# For better API documentation
+sphinx-autodoc-typehints>=1.12.0,<2.0.0
+sphinx-copybutton>=0.5.0,<1.0.0
 
-# Mock heavy dependencies in conf.py instead of installing them
-# This prevents build failures on Read the Docs
+# For mathematical expressions
+sphinxcontrib-katex>=0.8.6,<1.0.0
+
+# Basic scientific computing (lighter versions for docs)
+numpy>=1.20.0,<2.0.0
+matplotlib>=3.3.0,<4.0.0
+
+# Required for proper documentation build
+packaging>=21.0
+docutils>=0.17,<0.20
+
+# For proper reStructuredText parsing
+Pygments>=2.7.0,<3.0.0
