### 2025-07-16 a3a76d0
paper Updates
Updates to the paper
