### 2025-07-16 754fb0e
paper Updates
Updates to the paper
