+++ docs/source/user_guide/Perception.rst
@@ -1,5 +1,5 @@
 Perception User Guide
-====================
+=======================
 
 .. _user_guide_perception:
 
@@ -10,7 +10,7 @@ This comprehensive guide covers the Perception module in ManipulaPy, which provi
    :depth: 2
 
 Overview
---------
+----------
 
 The Perception module serves as a higher-level interface that builds upon the Vision module to provide sophisticated environmental understanding capabilities for robotic systems. It transforms raw visual data into actionable information for robot control and navigation.
 
@@ -47,7 +47,7 @@ The Perception module serves as a higher-level interface that builds upon the Vi
    </div>
 
 Key Features
-------------
+--------------
 
 **Environmental Understanding**
   - Real-time obstacle detection and classification
@@ -65,10 +65,10 @@ Key Features
   - Coordinate frame transformations for robot control
 
 Getting Started
----------------
+-----------------
 
 Basic Perception Setup
-~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~
 
 The Perception module requires a Vision instance to function:
 
@@ -87,7 +87,7 @@ The Perception module requires a Vision instance to function:
    print("🧠 Perception system initialized successfully!")
 
 Simple Obstacle Detection
-~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Detect and cluster obstacles in the robot's environment:
 
@@ -123,10 +123,10 @@ Detect and cluster obstacles in the robot's environment:
        print(f"  📏 Size: [{cluster_size[0]:.2f}, {cluster_size[1]:.2f}, {cluster_size[2]:.2f}] m")
 
 Core Functionality
-------------------
+----------------------
 
 Obstacle Detection and Clustering
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 The primary function of the Perception module is to detect and cluster obstacles:
 
@@ -174,7 +174,7 @@ The detection pipeline follows these steps:
 6. **Filtering**: Remove noise and invalid clusters
 
 Stereo Vision Processing
-~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 For systems with stereo cameras, generate detailed 3D point clouds:
 
@@ -208,7 +208,7 @@ For systems with stereo cameras, generate detailed 3D point clouds:
        print("⚠️ Stereo vision not enabled - using monocular detection")
 
 Advanced Clustering Control
-~~~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 Fine-tune clustering parameters for different environments:
 
@@ -243,10 +243,10 @@ Fine-tune clustering parameters for different environments:
        return labels, num_clusters
 
 Data Flow Architecture
---------------------
+------------------------
 
 Understanding the Data Pipeline
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
[…]
