### 2025-07-20 boelnasr/ManipulaPy#16 @cadojo
[JOSS Review] Code Coverage Reporting
This project has a lot of great unit tests! Have you considered tracking and reporting code coverage with a free tool, e.v. Coveralls or CodeCov? This would be a nice addition, but I do not consider this addition required for paper approval.

Could you please point me to how the [coverage statistics](https://github.com/boelnasr/ManipulaPy?tab=readme-ov-file#coverage-statistics) discussed in the README are measured? If no such documentation exists for repeated measurement, please add documentation for how these metrics are recorded. Ideally, these metrics would be recorded with formal coverage utilities (as discussed above).

---

- [ ] [optional] CodeCov or Coveralls Integration
- [x] Documentation for repeatable coverage metrics discussed in README.

---

This issue was opened as part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8490).
