### 2025-07-20 #8490 @cadojo
Thank you for the ping @lockwo! It looks like @boelnasr has made a lot of great progress toward @sea-bass's review comments. The documentation in particular looks great.

I echo @sea-bass's comment about removing some CPU-compatible platforms from the mock list in unit tests. I also recommend adding [code coverage reporting](https://github.com/boelnasr/ManipulaPy/issues/16), though I don't consider that item required for purposes of review.

I am in the process of attempting installation with GPU support on my end. I will consider my first review pass complete after I attempt that installation, and after I attempt to replicate performance claims and functionality.
