### 2025-07-20 boelnasr/ManipulaPy#17 @cadojo
[JOSS Review] Discuss Limitations of ManipulaPy in Paper
It would be helpful to expand on alternative projects --- the paper briefly mentions MoveIt!, Orocos KDL, and Robotics Toolbox --- and any limitations ManipulaPy has that some alternatives do not. For example: many roboticists use ROS as a middle layer across applications. Is ManipulaPy compatible with use inside ROS applications? This specific example (ROS) is not required for review, but I believe adding more paper content which compares and contrasts ManipulaPy with alternatives would be helpful for readers.

---

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8490).
