+++ ManipulaPy/sim.py
@@ -6,6 +6,11 @@ Simulation Module - ManipulaPy
 This module provides PyBullet-based simulation capabilities for robotic manipulators
 including real-time visualization, physics simulation, and interactive control.
 
+UPDATED VERSION with VISIBLE TRAJECTORY SPLINE:
+- Replaced addUserDebugLine() with real capsule geometry
+- Trajectory splines now appear in getCameraImage() screenshots
+- Added proper cleanup for trajectory visualization
+
 Copyright (c) 2025 Mohamed Aboelnar
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 
@@ -55,6 +60,10 @@ class Simulation:
         self.joint_params = []
         self.reset_button = None
         self.home_position = None
+        
+        # NEW: Track trajectory visualization bodies for cleanup
+        self.trajectory_body_ids = []
+        
         self.setup_simulation()
 
     def setup_logger(self):
@@ -149,6 +158,18 @@ class Simulation:
             ]
             self.home_position = np.zeros(len(self.non_fixed_joints))
 
+    def set_robot_models(self, robot, dynamics):
+        """
+        Set pre-existing robot models to avoid reprocessing.
+        
+        Args:
+            robot: SerialManipulator instance
+            dynamics: ManipulatorDynamics instance
+        """
+        self.robot = robot
+        self.dynamics = dynamics
+        self.logger.info("Pre-existing robot models set successfully.")
+
     def initialize_planner_and_controller(self):
         """
         Initializes the trajectory planner and the manipulator controller.
@@ -206,6 +227,275 @@ class Simulation:
         ]
         return np.array(joint_positions)
 
+    def _capsule_line(self, a, b, radius=0.006, rgba=(1, 0.5, 0, 1)):
+        """
+        Create a thin capsule between point a and b; returns body-id.
+        This creates REAL GEOMETRY that appears in getCameraImage() screenshots.
+        
+        Args:
+            a: Start point [x, y, z]
+            b: End point [x, y, z]
+            radius: Capsule radius in world units
+            rgba: Color as [r, g, b, a] where values are 0-1
+        
+        Returns:
+            int: PyBullet body ID, or -1 if failed
+        """
+        a, b = np.array(a), np.array(b)
+        v = b - a
+        L = np.linalg.norm(v)
+        
+        if L < 1e-6:
+            return -1
+        
+        # Calculate orientation to align capsule with the line direction
+        z = v / L  # Direction vector
+        
+        # Find perpendicular vectors
+        x = np.cross([0, 0, 1], z)
+        if np.linalg.norm(x) < 1e-6:
+            x = np.cross([0, 1, 0], z)
+        x = x / np.linalg.norm(x)
+        y = np.cross(z, x)
+        
+        # Calculate proper orientation for capsule
+        # PyBullet capsules are aligned with Z-axis by default
+        if abs(z[2]) > 0.99:  # Nearly vertical
+            orn = p.getQuaternionFromEuler([0, 0, 0])
+        else:
+            # Calculate rotation to align Z-axis with direction vector
+            angle = np.arccos(np.clip(z[2], -1, 1))
+            if angle > 1e-6:
+                axis = np.cross([0, 0, 1], z)
+                axis_norm = np.linalg.norm(axis)
+                if axis_norm > 1e-6:
+                    axis = axis / axis_norm
+                    orn = p.getQuaternionFromAxisAngle(axis, angle)
+                else:
+                    orn = p.getQuaternionFromEuler([0, 0, 0])
+            else:
+                orn = p.getQuaternionFromEuler([0, 0, 0])
+        
+        # Midpoint of the line segment
[…]
