+++ Benchmark/quick_benchmark.py
@@ -1,0 +1,1076 @@
+#!/usr/bin/env python3
+"""
+ManipulaPy Comprehensive Benchmark Suite
+
+This benchmark suite tests performance across all ManipulaPy modules:
+- Kinematics (forward/inverse, Jacobian computation)
+- Dynamics (mass matrix, inverse/forward dynamics) 
+- Trajectory Planning (joint/Cartesian trajectories, CUDA kernels)
+- Control (PID, computed torque, adaptive control)
+- Vision & Perception (obstacle detection, clustering)
+- URDF Processing (loading and conversion)
+- Singularity Analysis (workspace generation, manipulability)
+
+Usage:
+    python benchmark_manipulapy.py [--module MODULE] [--iterations N] [--save-results]
+"""
+
+import time
+import numpy as np
+import matplotlib.pyplot as plt
+import pandas as pd
+import argparse
+import os
+import sys
+import warnings
+from pathlib import Path
+from typing import Dict, List, Tuple, Optional, Any
+import json
+import psutil
+import platform
+import matplotlib
+matplotlib.use('Agg')  # Non-interactive backend to avoid Qt issues
+# Suppress warnings for cleaner output
+warnings.filterwarnings("ignore")
+
+# Add ManipulaPy to path if needed
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+
+try:
+    import cupy as cp
+    CUDA_AVAILABLE = True
+    print("✅ CUDA/CuPy available for GPU acceleration")
+except ImportError:
+    CUDA_AVAILABLE = False
+    print("⚠️ CUDA/CuPy not available - using CPU only")
+
+class BenchmarkTimer:
+    """Context manager for timing operations"""
+    def __init__(self, name: str, warmup: bool = False):
+        self.name = name
+        self.warmup = warmup
+        self.start_time = None
+        self.end_time = None
+        
+    def __enter__(self):
+        if not self.warmup:
+            print(f"🚀 Running {self.name}...")
+        self.start_time = time.perf_counter()
+        return self
+        
+    def __exit__(self, *args):
+        self.end_time = time.perf_counter()
+        self.elapsed = self.end_time - self.start_time
+        if not self.warmup:
+            print(f"✅ {self.name}: {self.elapsed:.4f}s")
+
+class MockTrajectoryPlanner:
+    """Mock trajectory planner that avoids the boolean array issue"""
+    
+    def __init__(self, robot, dynamics, joint_limits, torque_limits):
+        self.serial_manipulator = robot
+        self.dynamics = dynamics
+        self.joint_limits = np.array(joint_limits)
+        # Fix the boolean array issue by properly handling torque_limits
+        if torque_limits is None:
+            self.torque_limits = np.array([[-np.inf, np.inf]] * len(joint_limits))
+        else:
+            self.torque_limits = np.array(torque_limits)
+    
+    def joint_trajectory(self, thetastart, thetaend, Tf, N, method):
+        """Generate joint trajectory using simple polynomial interpolation"""
+        thetastart = np.array(thetastart, dtype=np.float32)
+        thetaend = np.array(thetaend, dtype=np.float32)
+        
+        # Time vector
+        t = np.linspace(0, Tf, N)
+        
+        # Initialize output arrays
+        traj_pos = np.zeros((N, len(thetastart)))
+        traj_vel = np.zeros((N, len(thetastart)))
+        traj_acc = np.zeros((N, len(thetastart)))
+        
+        for i in range(N):
+            # Normalized time
+            s = t[i] / Tf
+            
+            if method == 3:  # Cubic
[…]
