+++ Benchmark/README.MD
@@ -1,0 +1,1177 @@
+# ManipulaPy Benchmark Suite
+
+This directory contains comprehensive benchmarking tools for evaluating ManipulaPy's performance, accuracy, and efficiency across different hardware configurations and problem sizes.
+
+## 📊 Available Benchmarks
+
+### 🚀 Performance Benchmark (`performance_benchmark.py`)
+Comprehensive performance evaluation across different problem sizes and hardware configurations.
+
+**Features:**
+- GPU vs CPU performance comparison
+- Trajectory planning optimization analysis```markdown
+# ManipulaPy Benchmarking Suite
+
+<div align="center">
+
+**Performance evaluation and accuracy validation tools for ManipulaPy's GPU-accelerated robotics framework**
+
+</div>
+
+---
+
+## 📊 Overview
+
+The ManipulaPy Benchmarking Suite provides essential tools to evaluate performance characteristics and validate computational accuracy within the ManipulaPy framework. This suite helps researchers and developers understand GPU acceleration benefits, assess numerical precision, and perform quick performance validation for continuous integration workflows.
+
+## 🚀 Key Features
+
+- **Performance Profiling**: Comprehensive timing analysis for kinematics, dynamics, and trajectory planning
+- **GPU vs CPU Comparison**: Direct performance comparison between CUDA-accelerated and CPU implementations  
+- **Accuracy Validation**: Numerical precision testing between different computational backends
+- **Quick Validation**: Fast performance checks suitable for CI/CD pipelines
+- **Automated Reporting**: Generate detailed reports with visualizations and statistical analysis
+
+---
+
+## 📁 Benchmark Files
+
+| File | Description | Use Case |
+|------|-------------|----------|
+| **`performance_benchmark.py`** | Comprehensive performance analysis suite | Full system performance evaluation |
+| **`accuracy_benchmark.py`** | Numerical accuracy validation between backends | Computational precision verification |
+| **`quick_benchmark.py`** | Fast performance check for CI/CD pipelines | Continuous integration testing |
+
+---
+
+## 🏃‍♂️ Quick Start
+
+### Prerequisites
+
+```bash
+# Install ManipulaPy with GPU support
+pip install ManipulaPy[gpu-cuda12]  # or gpu-cuda11
+
+# Install additional benchmarking dependencies
+pip install matplotlib seaborn tabulate psutil
+```
+
+### Running Benchmarks
+
+```bash
+# Navigate to benchmark directory
+cd ManipulaPy/Benchmark/
+
+# Run comprehensive performance benchmark
+python performance_benchmark.py
+
+# Validate computational accuracy
+python accuracy_benchmark.py
+
+# Quick performance check
+python quick_benchmark.py
+```
+
+---
+
+## 📈 Benchmark Types
+
+### 1. Performance Benchmark (`performance_benchmark.py`)
+
+Comprehensive evaluation of all ManipulaPy components with detailed profiling:
+
+```python
+# Example output sections:
+✅ System Information
+✅ Kinematics Benchmark (Forward/Inverse)
+✅ Dynamics Benchmark (Mass Matrix, Coriolis, Gravity)
+✅ Trajectory Planning Benchmark (Joint/Cartesian Space)
+✅ Control Systems Benchmark (PID, Computed Torque)
+✅ GPU vs CPU Performance Comparison
+✅ Memory Usage Analysis
[…]
