+++ Benchmark/performance_benchmark.py
@@ -1,0 +1,1539 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: AGPL-3.0-or-later
+"""
+Performance Benchmark Suite for ManipulaPy Path Planning
+
+This comprehensive benchmark compares CPU vs GPU performance for various path planning
+tasks including joint trajectory generation, inverse dynamics, forward dynamics,
+Cartesian trajectory planning, and batch processing.
+
+Features:
+- Systematic benchmarking across different problem sizes
+- Statistical analysis with confidence intervals
+- Performance profiling and memory usage tracking
+- Visual performance comparison charts
+- Scalability analysis
+- Hardware capability assessment
+
+Copyright (c) 2025 Mohamed Aboelnar
+Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
+"""
+
+import time
+import numpy as np
+import matplotlib.pyplot as plt
+import seaborn as sns
+import pandas as pd
+import psutil
+import gc
+import warnings
+from typing import Dict, List, Tuple, Optional
+from dataclasses import dataclass
+from pathlib import Path
+import json
+import logging
+import matplotlib
+matplotlib.use("Agg")
+# ManipulaPy imports
+try:
+    from ManipulaPy.path_planning import OptimizedTrajectoryPlanning, compare_implementations
+    from ManipulaPy.cuda_kernels import (
+        CUDA_AVAILABLE, check_cuda_availability, get_gpu_properties,
+        benchmark_kernel_performance, profile_start, profile_stop
+    )
+    from ManipulaPy.kinematics import SerialManipulator
+    from ManipulaPy.dynamics import ManipulatorDynamics
+    from ManipulaPy.urdf_processor import URDFToSerialManipulator
+    from ManipulaPy import utils
+except ImportError as e:
+    print(f"Error importing ManipulaPy: {e}")
+    print("Please ensure ManipulaPy is properly installed")
+    exit(1)
+
+# Set up logging
+logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+logger = logging.getLogger(__name__)
+
+# Suppress warnings for cleaner output
+warnings.filterwarnings('ignore')
+
+@dataclass
+class BenchmarkConfig:
+    """Configuration for benchmark parameters"""
+    test_name: str
+    trajectory_points: List[int]
+    joint_counts: List[int]
+    batch_sizes: List[int]
+    num_runs: int
+    timeout_seconds: float
+    enable_profiling: bool
+    save_results: bool
+    output_dir: str
+
+@dataclass
+class BenchmarkResult:
+    """Individual benchmark result"""
+    test_name: str
+    implementation: str  # 'cpu' or 'gpu'
+    n_points: int
+    n_joints: int
+    batch_size: int
+    execution_time: float
+    memory_usage_mb: float
+    success: bool
+    error_message: Optional[str] = None
+
+class PathPlanningBenchmark:
+    """
+    Comprehensive benchmark suite for ManipulaPy path planning performance.
+    """
+    
+    def __init__(self, config: BenchmarkConfig):
+        """
+        Initialize the benchmark suite.
+        
+        Args:
+            config: Benchmark configuration parameters
+        """
+        self.config = config
+        self.results: List[BenchmarkResult] = []
+        self.system_info = self._get_system_info()
+        
+        # Create output directory in the same directory as the script
+        script_dir = Path(__file__).parent if hasattr(Path(__file__), 'parent') else Path.cwd()
+        self.output_path = script_dir / config.output_dir
+        self.output_path.mkdir(parents=True, exist_ok=True)
+        
+        # Create subdirectories for organized data storage
[…]
