+++ Benchmark/accuracy_benchmark.py
@@ -1,0 +1,1703 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: AGPL-3.0-or-later
+"""
+Comprehensive Accuracy Benchmark for ManipulaPy
+
+This script provides comprehensive accuracy benchmarks for the ManipulaPy library,
+testing various components including kinematics, dynamics, trajectory planning,
+control algorithms, and GPU vs CPU implementations using real robot data.
+
+Copyright (c) 2025 Mohamed Aboelnar
+Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
+"""
+
+import numpy as np
+import time
+import logging
+import warnings
+import matplotlib.pyplot as plt
+import seaborn as sns
+from typing import Dict, List, Tuple, Optional, Any
+import json
+import os
+from datetime import datetime
+import pandas as pd
+import matplotlib
+matplotlib.use("Agg")
+
+# ManipulaPy imports
+try:
+    from ManipulaPy.kinematics import SerialManipulator
+    from ManipulaPy.dynamics import ManipulatorDynamics
+    from ManipulaPy.control import ManipulatorController
+    from ManipulaPy.path_planning import OptimizedTrajectoryPlanning
+    from ManipulaPy.urdf_processor import URDFToSerialManipulator
+    from ManipulaPy.cuda_kernels import (
+        CUDA_AVAILABLE, 
+        check_cuda_availability,
+        optimized_trajectory_generation,
+        trajectory_cpu_fallback
+    )
+    from ManipulaPy import utils
+    from ManipulaPy.ManipulaPy_data.xarm import urdf_file
+    MANIPULAPY_AVAILABLE = True
+except ImportError as e:
+    print(f"Warning: ManipulaPy not fully available: {e}")
+    MANIPULAPY_AVAILABLE = False
+
+# Set up logging
+logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
+logger = logging.getLogger(__name__)
+
+# Suppress warnings for cleaner output
+warnings.filterwarnings('ignore', category=UserWarning)
+warnings.filterwarnings('ignore', category=RuntimeWarning)
+
+
+class PatchedSingularity:
+    """
+    Patched version of Singularity class that handles non-square Jacobians properly.
+    """
+    def __init__(self, serial_manipulator):
+        self.serial_manipulator = serial_manipulator
+
+    def singularity_analysis(self, thetalist):
+        """
+        Analyze if the manipulator is at a singularity based on the manipulability measure.
+        """
+        try:
+            J = self.serial_manipulator.jacobian(thetalist, frame="space")
+            
+            # For non-square Jacobians (typical case for 6-DOF end-effector in n-DOF space)
+            if J.shape[0] != J.shape[1]:
+                # Use manipulability index: sqrt(det(J * J^T))
+                JJT = J @ J.T
+                det_JJT = np.linalg.det(JJT)
+                manipulability = np.sqrt(abs(det_JJT))
+                # Singularity when manipulability is near zero
+                return manipulability < 1e-4
+            else:
+                # Square Jacobian case (rare)
+                det_J = np.linalg.det(J)
+                return abs(det_J) < 1e-4
+        except (np.linalg.LinAlgError, ValueError):
+            # If calculation fails, assume singular
+            return True
+
+    def condition_number(self, thetalist):
+        """
+        Calculate the condition number of the Jacobian for a given set of joint angles.
+        For non-square Jacobians, uses the condition number of J*J^T.
+        """
+        try:
+            J = self.serial_manipulator.jacobian(thetalist, frame="space")
+            
+            if J.shape[0] == J.shape[1]:
+                # Square Jacobian
[…]
