+++ Benchmark/benchmark_results_20250720_145750/README.md
@@ -1,0 +1,40 @@
+# ManipulaPy Benchmark Results
+
+## Directory Structure
+
+- **`data/`** - Raw benchmark data in JSON and CSV formats
+  - `benchmark_results_latest.json` - Complete results in JSON format
+  - `benchmark_results_latest.csv` - Tabular data for analysis
+  - `summary_statistics_latest.json` - Statistical summary
+  - `system_info.json` - System configuration and hardware info
+  - `failed_tests_*.json` - Information about any failed tests
+
+- **`plots/`** - Visualization and charts
+  - `benchmark_results_overview.*` - Main performance comparison charts
+  - `detailed_analysis.*` - In-depth analysis plots
+  - `*_analysis.*` - Individual test type analysis
+
+- **`reports/`** - Written reports and documentation
+  - `benchmark_report_latest.md` - Comprehensive benchmark report
+  - Timestamped reports for historical comparison
+
+- **`logs/`** - Execution logs and debugging information
+  - `benchmark_*.log` - Detailed execution logs
+
+## Quick Access Files
+
+- 📊 **Latest Results**: `data/benchmark_results_latest.csv`
+- 📈 **Main Charts**: `plots/benchmark_results_overview.png`
+- 📋 **Full Report**: `reports/benchmark_report_latest.md`
+- ⚙️ **System Info**: `data/system_info.json`
+
+## Benchmark Configuration
+
+- **Test Date**: 2025-07-20 16:10:55
+- **Trajectory Points**: [10000, 50000, 100000]
+- **Joint Counts**: [3, 6]
+- **Batch Sizes**: [1, 2, 3]
+- **Runs per Test**: 3
+- **CUDA Available**: True
+
+Generated by ManipulaPy Performance Benchmark Suite

[…]
