+++ tests/test_trajectory_planning.py
@@ -1,826 +1,645 @@
 #!/usr/bin/env python3
 
-from numba import cuda, float32
+import unittest
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.spatial import ConvexHull
-from ManipulaPy.utils import (
-    TransToRp,
-    MatrixLog3,
-    MatrixExp3,
-    CubicTimeScaling,
-    QuinticTimeScaling,
-)
-from urchin.urdf import URDF
-from ManipulaPy.cuda_kernels import trajectory_kernel, inverse_dynamics_kernel, forward_dynamics_kernel, cartesian_trajectory_kernel
-from ManipulaPy.potential_field import CollisionChecker, PotentialField
-import logging
+import os
+from ManipulaPy.path_planning import OptimizedTrajectoryPlanning, TrajectoryPlanning
+from ManipulaPy.urdf_processor import URDFToSerialManipulator
+from ManipulaPy.ManipulaPy_data.xarm import urdf_file as xarm_urdf_file
 
-# Set up logging
-logging.basicConfig(
-    level=logging.INFO,
-    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-    datefmt='%Y-%m-%d %H:%M:%S'
-)
-logger = logging.getLogger(__name__)
 
-class TrajectoryPlanning:
-    def __init__(self, serial_manipulator, urdf_path, dynamics, joint_limits, torque_limits=None):
-        """
-        Initializes a TrajectoryPlanning object.
+def is_module_available(module_name):
+    """Check if a module is available and not mocked."""
+    try:
+        module = __import__(module_name)
+        return not hasattr(module, '_name') or module._name != f"MockModule({module_name})"
+    except ImportError:
+        return False
 
-        Args:
-            serial_manipulator (SerialManipulator): An instance of SerialManipulator.
-            urdf_path (str): The path to the URDF file.
-            dynamics (ManipulatorDynamics): An instance of ManipulatorDynamics.
-            joint_limits (list): A list of tuples representing the joint limits.
-            torque_limits (list, optional): A list of tuples representing the torque limits. Defaults to None.
-        """
-        self.serial_manipulator = serial_manipulator
-        self.dynamics = dynamics
-        self.joint_limits = np.array(joint_limits)
-        self.torque_limits = (
-            np.array(torque_limits)
-            if torque_limits
-            else np.array([[-np.inf, np.inf]] * len(joint_limits))
-        )
-        
-        # Initialize collision checker and potential field with error handling
-        try:
-            self.collision_checker = CollisionChecker(urdf_path)
-        except Exception as e:
-            logger.warning(f"Failed to initialize collision checker: {e}")
-            self.collision_checker = None
-        
-        try:
-            self.potential_field = PotentialField()
-        except Exception as e:
-            logger.warning(f"Failed to initialize potential field: {e}")
-            self.potential_field = None
 
-        # Check if CUDA is available
-        try:
-            cuda.current_context()
-            self.cuda_available = True
-            logger.info("CUDA is available for trajectory planning")
-        except Exception:
-            self.cuda_available = False
-            logger.info("CUDA not available, will use NumPy fallback for trajectory planning")
-
-    def _fallback_joint_trajectory(self, thetastart, thetaend, Tf, N, method):
-        """
-        Pure NumPy implementation of joint trajectory generation as fallback.
-        """
-        logger.info("Using NumPy fallback for joint trajectory generation")
-        
-        thetastart = np.array(thetastart, dtype=np.float32)
-        thetaend = np.array(thetaend, dtype=np.float32)
-        num_joints = len(thetastart)
-
-        traj_pos = np.zeros((N, num_joints), dtype=np.float32)
-        traj_vel = np.zeros((N, num_joints), dtype=np.float32)
[…]
