+++ paper/paper.md
@@ -6,7 +6,7 @@ authors:
     orcid: 0000-0002-1768-2031
     affiliation: 1
 affiliations:
-  - name: Universität Duisburg‑Essen
+  - name: Universität Duisburg‑Essen
     index: 1
 date: "2025-05-03"
 bibliography: paper.bib
@@ -14,150 +14,133 @@ bibliography: paper.bib
 
 # Summary
 
-**ManipulaPy** is an open‑source Python toolbox that stitches together the entire manipulation pipeline—from URDF parsing to GPU‑accelerated dynamics, vision‑based perception, planning and control—within a single API.  Built on the Product‑of‑Exponentials model [@lynch2017modern], PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library lets researchers move from robot description to real‑time control with up to **40 ×** faster inverse‑dynamics on a 6‑DOF UR5 than a NumPy baseline.
+**ManipulaPy** is an open‑source Python toolbox that unifies the entire manipulation pipeline—from URDF parsing to GPU‑accelerated dynamics, vision‑based perception, planning and control—within a single API. Built on the Product‑of‑Exponentials model [@lynch2017modern] (similar to Pinocchio [@Pinocchio2025] but with GPU acceleration), PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library enables researchers to move from robot description to real‑time control with up to **13× overall performance improvement** and **3600× faster inverse dynamics** on a 6‑DOF UR5 compared to NumPy baseline. Performance claims are reproducible via benchmarks in `benchmarks/README.md`.
 
 # Statement of Need
 
-Robotics research needs tight couplings of geometry, physics, vision and control. Existing stacks—MoveIt [@chitta2012moveit], Orocos KDL [@smits2009kdl] and the Python Robotics Toolbox [@corke2021]—cover parts of this but require glue code or lack GPU paths. **ManipulaPy** instead:
+Modern manipulation research requires tight integration of geometry, dynamics, perception, planning, and control—ideally within a single, real-time computational loop on GPU hardware. However, existing open-source tools address only portions of this pipeline, forcing researchers to write substantial integration code:
 
-* converts a URDF to PoE screws **and** realistic joint limits in one call,  
-* exposes CUDA kernels for time‑scaling and (inverse) dynamics [@liang2018gpu],  
-* pipes stereo vision through DBSCAN obstacle clustering into the planner¹,  
-* wraps PyBullet so cameras, planners and controllers stay synchronised at 1 kHz.
+| Library | Core Strengths | Integration Challenges |
+|---------|---------------|------------------------|
+| **MoveIt** [@chitta2012moveit] | Mature sampling-based planners | Requires custom ROS nodes to bridge sensor data with planning; external plugins needed for real-time dynamics; no native GPU acceleration |
+| **Pinocchio** [@Pinocchio2025] | High-performance PoE dynamics (C++) | CPU-only; separate perception and planning libraries must be manually synchronized; requires Python bindings for integration |
+| **CuRobo** [@sundaralingam2023curobo] | GPU-accelerated collision checking and trajectory optimization | Planning-focused; lacks perception pipeline and closed-loop control; requires external sensor processing |
+| **Python Robotics Toolbox** [@corke2021] | Educational algorithms with clear APIs | CPU-only implementation; users must implement their own simulators, controllers, and sensor processing |
+| **PyRoKi** [@pyroki2025] | JAX-accelerated kinematics | Early development stage; limited dynamics and no perception support |
[…]
