+++ setup.py
@@ -5,95 +5,209 @@ with open("README.md", "r", encoding="utf-8") as fh:
 
 setup(
     name="ManipulaPy",
-    version="1.1.0",  # keep at 1.1.0
+    version="1.1.0",  # Major feature release with GPU acceleration and vision pipeline
     author="Mohamed Aboelnasr",
     author_email="aboelnasr1997@gmail.com",
-    description="A modular, GPU-accelerated Python package for robotic manipulator simulation and control",
+    description="A comprehensive, GPU-accelerated Python framework for robotic manipulation, perception, and control",
     long_description=long_description,
     long_description_content_type="text/markdown",
     url="https://github.com/boelnasr/ManipulaPy",
     packages=find_packages(),
+    
+    # Core dependencies - required for basic functionality
     install_requires=[
-        "numpy>=1.24,<2.0",
-        "scipy>=1.10,<1.13",
-        "pybullet>=3.2.5",
-        "urchin>=0.0.28",
-        "trimesh>=4.0,<4.2",
-        "opencv-python>=4.5,<5.0",
-        "scikit-learn>=1.3,<1.6",
-        "matplotlib>=3.3",
-        "ultralytics>=8.0",    # YOLO-based perception
-        "torch>=1.8.0",        # required by ultralytics
-        "cupy-cuda11x>=10.0.0",# optional CUDA arrays & kernels
+        "numpy>=1.20.0,<2.0",         # Core numerical operations
+        "scipy>=1.7.0,<2.0",          # Scientific computing
+        "matplotlib>=3.3.0,<4.0",     # Plotting and visualization
+        "pybullet>=3.2.5,<4.0",       # Physics simulation
+        "urchin>=0.0.27,<1.0",        # URDF processing
+        "trimesh>=3.15.0,<5.0",       # 3D mesh processing
+        "opencv-python>=4.5.0,<5.0",  # Computer vision
+        "scikit-learn>=1.0.0,<2.0",   # Machine learning (clustering)
+        "numba>=0.56.0,<1.0",         # JIT compilation for performance
     ],
+    
+    # Optional dependencies for enhanced functionality
     extras_require={
+        # GPU acceleration with CUDA 11.x
+        "gpu-cuda11": [
+            "cupy-cuda11x>=10.0.0,<13.0",  # CUDA arrays and kernels
+        ],
+        
+        # GPU acceleration with CUDA 12.x  
+        "gpu-cuda12": [
+            "cupy-cuda12x>=12.0.0,<13.0",  # CUDA arrays and kernels
+        ],
+        
+        # Legacy GPU support
         "gpu": [
-            "pycuda>=2021.1",    # legacy PyCUDA support
+            "cupy-cuda11x>=10.0.0,<13.0",  # Default to CUDA 11.x
+            "pycuda>=2021.1,<2025.0",      # Legacy PyCUDA support
         ],
+        
+        # Vision and perception capabilities
+        "vision": [
+            "ultralytics>=8.0.0,<9.0",     # YOLO object detection
+            "torch>=1.8.0,<3.0",           # Required by ultralytics
+            "torchvision>=0.9.0,<1.0",     # Computer vision models
+        ],
+        
+        # Development and testing tools
         "dev": [
-            "pytest>=6.0.0",
-            "pytest-cov>=2.0.0",
-            "black>=21.0.0",
-            "flake8>=3.8.0",
-            "isort>=5.0.0",
-            "mypy>=0.910",
-            "pre-commit>=2.15.0",
+            "pytest>=7.0.0,<8.0",
+            "pytest-cov>=4.0.0,<5.0", 
+            "pytest-benchmark>=4.0.0,<5.0",
[…]
