+++ pyproject.toml
@@ -9,7 +9,7 @@ build-backend = "setuptools.build_meta"
 # ------------------------------------------------------------------
 [project]
 name = "ManipulaPy"                    # exact PyPI casing
-version = "1.1.0"                      
+version = "1.1.1"                      
 authors = [
     { name = "Mohamed Aboelnar", email = "aboelnasr1997@gmail.com" },
 ]
