### 2025-07-20 boelnasr/ManipulaPy#12 @boelnasr
Hi @sea-bass, thanks so much for the thorough feedback! I’ve just pushed updates addressing every point:

---

### 1. Reproducibility of the 40× claim

* Added a **Reproducibility** section to `benchmarks/README.md` with step‑by‑step instructions (prereqs, install, benchmark command) so anyone can reproduce the NumPy vs GPU timings.

### 2. PoE / Pinocchio citation

* In the Summary and Theory sections I now explicitly cite Pinocchio (\[@Pinocchio2025]) when introducing the product‑of‑exponentials model, and note that ManipulaPy adds GPU acceleration.

### 3. Example simulation screenshot

* Added a new figure (`figure_1_manipulapy_trajectory.png`) showing a UR5 in PyBullet avoiding a dynamic obstacle at 1 kHz.

### 4. Vision pipeline details

* Pre‑pended a high‑level overview paragraph in “Vision and Perception Pipeline,” and spelled out all supported obstacle representations (primitives, point clouds, SDFs, octrees/octomaps, hybrid).
* Moved the DBSCAN citation inline in Stage 4 (no more footnote).

### 5. Concrete “glue code” examples

* Replaced the vague “glue code” line with concrete examples (e.g. custom OpenCV → ROS → MoveIt pipelines, PyBullet URDF → Pinocchio wrappers, CuPy tensor serialization).

### 6. cuRobo / PyRoKi / CBFPy comparisons

* Expanded the Statement of Need table to include **CuRobo**, **PyRoKi**, and **CBFPy** with their core strengths and integration gaps.

### 7. Library Architecture overview

* Added an opening paragraph describing the five conceptual layers (Model, Dynamics, Perception, Planning, Control) and how they map to a stable GPU‑first API.
* Included a block‑diagram of the pipeline as Figure Y.

### 8. Clarify `vision.py` vs. `perception.py`

* In the Architecture section I now explain that `vision.py` handles low‑level camera ops (capture, rectification, stereo), while `perception.py` does detection (YOLO), clustering (DBSCAN), and obstacle representation.

### 9. Generalized module listing

* Replaced the full file listing with a short summary of key modules and how new functionality plugs in via defined interfaces.

### 10. Fixed Ultralytics YOLO citation

* Corrected the BibTeX entry and in‑text link for Ultralytics YOLO v8.

Please let me know if there’s anything I missed or if you’d like a final consistency pass on figures/references. Thanks again! 🙌
