### 2025-07-20 boelnasr/ManipulaPy#5 @boelnasr
The file was removed, the models will be automatically downloaded when the vision module is launched
