### 2025-07-20 boelnasr/ManipulaPy#7 @boelnasr
**/.vscode
is delted now
