### 2025-07-20 boelnasr/ManipulaPy#10 @boelnasr
Thank you for this important feedback! I have now added comprehensive contributing guidelines to the README that address all three required criteria:
✅ 1. Contributing to the software:

Detailed fork and pull request workflow with step-by-step commands
Development environment setup instructions (pip install -e .[dev])
Code quality requirements (PEP 8, Black formatting, testing)
Pre-commit hooks and CI validation process

✅ 2. Reporting issues or problems:

Clear guidance on opening GitHub Issues with required information
Template for bug reports (descriptive title, reproduction steps, expected vs. actual behavior)
Log and screenshot requirements for effective debugging

✅ 3. Seeking support:

Multiple support channels clearly listed:

GitHub Discussions for design questions
GitHub Issues for bug reports
Direct email contact ([aboelnasr1997@gmail.com](mailto:aboelnasr1997@gmail.com))
Documentation links and examples directory

Code of Conduct reference for community standards

Additional improvements made:

Contribution areas clearly defined (bugs, features, docs, performance, testing, visualization)
Development guidelines with specific tools and practices
Community engagement pathways for different types of users
Contact information with multiple channels for different needs

The guidelines now provide clear, actionable paths for all forms of community engagement while maintaining professional standards for code quality and documentation.
I believe this now meets the JOSS community guidelines requirements. Please let me know if any additional details would be helpful!
