### 2025-07-20 9bb1ea5
readme update
insallation guide lines
