### 2025-07-20 boelnasr/ManipulaPy#3 @boelnasr
Thanks for testing the installer and catching this gap! You’re absolutely right – we were assuming CUDA/cuDNN were already in place, which isn’t always true. I’ve just pushed an update to the README that includes:

1. **Prerequisite steps** for installing the NVIDIA CUDA Toolkit (so that `nvcc` is on your `$PATH`) and cuDNN
2. A quick “Verify your GPU” check with `nvidia-smi` and `nvcc --version`
3. A simple post‑install Python snippet to confirm that ManipulaPy actually sees your GPU

Here’s the new “Prerequisites” + “Post‑Install Check” sections we added:

````markdown
## 🚀 Quick Start

### Prerequisites

Before installing ManipulaPy, make sure your system has:

1. **NVIDIA Drivers & CUDA Toolkit**
   - Install `nvcc` (e.g. `sudo apt install nvidia-cuda-toolkit` or via NVIDIA’s official CUDA installer).
   - Verify:
     ```bash
     nvidia-smi       # should list your GPU(s) and driver version
     nvcc --version   # should print your CUDA version
     ```

2. **cuDNN**
   - Download & install the appropriate cuDNN library for your CUDA version from NVIDIA.
   - Verify that headers and libraries are present under `/usr/include` and `/usr/lib`.

### Installation

```bash
# CPU-only
pip install ManipulaPy

# GPU support (CUDA 11.x)
pip install ManipulaPy[gpu-cuda11]

# GPU support (CUDA 12.x)
pip install ManipulaPy[gpu-cuda12]

# Dev install
git clone https://github.com/boelnasr/ManipulaPy.git
cd ManipulaPy
pip install -e .[dev]
````

### Post‑Install Check

```bash
python3 - <<EOF
import manipulapy
if manipulapy.cuda.is_available():
    print("✅ CUDA detected:", manipulapy.cuda.get_device_name())
else:
    raise RuntimeError("❌ CUDA not detected. Check CUDA Toolkit/cuDNN installation.")
EOF
```

Could you give this a try and let me know if it catches all of the edge cases you saw? Thanks again for helping us tighten up the docs!
