### 2025-07-20 boelnasr/ManipulaPy#6 @boelnasr
Thanks for the nudge on this! I’ve now moved our docs out of the single notebook and into a full Sphinx-based site, hosted on ReadTheDocs:

🔗 [https://manipulapy.readthedocs.io/en/latest/index.html](https://manipulapy.readthedocs.io/en/latest/index.html)

* **Structure in‑repo:**

  * A new `/docs/` folder contains `conf.py`, `.rst` source files, and examples broken out into sections.
  * We’ve added a Makefile so you can build HTML locally with `make -C docs html`.

* **External hosting:**

  * The site is automatically built on ReadTheDocs whenever we push to `main`.
  * Sidebar navigation organizes Tutorials, API Reference, Installation, and Examples.

Let me know if this layout meets the JOSS quality bar or if you’d like additional sections or reorganizations!
