### 2025-07-20 boelnasr/ManipulaPy#11 @boelnasr
Thank you for pointing this out! Consistent naming and in-file documentation makes the Examples much more approachable.

**✅ I have implemented a consistent naming convention and added descriptive headers:**

**🔧 Implemented Naming Convention:**
```
basic_examples/
├── kinematics_basic_demo.py          # ✅ Follows consistent pattern
├── dynamics_basic_demo.py            # ✅ Follows consistent pattern
├── control_basic_demo.py             # ✅ Follows consistent pattern
└── urdf_processing_basic_demo.py     # ✅ Follows consistent pattern

intermediate_examples/
├── trajectory_planning_intermediate_demo.py    # ✅ Follows consistent pattern
├── perception_intermediate_demo.py             # ✅ Follows consistent pattern
├── simulation_intermediate_demo.py             # ✅ Follows consistent pattern
└── singularity_analysis_intermediate_demo.py  # ✅ Follows consistent pattern

advanced_examples/
├── gpu_acceleration_advanced_demo.py          # ✅ Follows consistent pattern
├── batch_processing_advanced_demo.py          # ✅ Follows consistent pattern
├── collision_avoidance_advanced_demo.py       # ✅ Follows consistent pattern
└── real_robot_integration_advanced_demo.py    # ✅ Follows consistent pattern
```

**📝 Added Standardized File Headers:**
Each example now includes a descriptive header with:
```python
#!/usr/bin/env python3
"""
Kinematics Basic Demo - ManipulaPy Examples

This example demonstrates forward and inverse kinematics computations
for a 6-DOF robotic manipulator, including Jacobian analysis and
workspace visualization.

Features demonstrated:
- Forward kinematics with Product-of-Exponentials
- Inverse kinematics with damped least-squares solver
- Jacobian computation and singularity detection
- End-effector workspace visualization

Requirements: Basic ManipulaPy installation
Complexity: ⭐ (Beginner)
Runtime: ~30 seconds
"""
```

**🗂️ Improved Organization:**
- **Consistent naming**: `{topic}_{level}_demo.py` pattern implemented
- **Clear hierarchy**: basic → intermediate → advanced complexity structure
- **Descriptive headers**: Purpose, features, requirements, and complexity ratings added
- **Runtime estimates**: Help users choose appropriate examples

**📋 Updated Examples README:**
I've also updated the Examples README with a comprehensive table showing example descriptions, complexity levels (⭐⭐⭐), required dependencies, expected runtime, and key concepts demonstrated.

The examples are now much more navigable and educational for users at different skill levels, providing a professional, consistent example suite that serves as effective documentation and learning material.
