### 2025-07-21 dc94ec4
Read Me
read me update to the latest project structure
