+++ README.md
@@ -423,35 +423,98 @@ for name, result in results.items():
 
 ## 📁 Examples & Tutorials
 
-The `Examples/` directory contains comprehensive demonstrations:
+The `Examples/` directory contains comprehensive demonstrations organized into three levels:
 
-| Example | Description | Complexity |
-|---------|-------------|------------|
-| `basic_kinematics.py` | Forward/inverse kinematics with visualization | ⭐ |
-| `trajectory_planning_demo.py` | GPU-accelerated path planning | ⭐⭐ |
-| `dynamics_and_control.py` | Complete dynamics and control analysis | ⭐⭐ |
-| `simulation_complete.py` | Full PyBullet simulation pipeline | ⭐⭐⭐ |
-| `perception_stereo.py` | Stereo vision and object detection | ⭐⭐⭐ |
-| `performance_comparison.py` | GPU vs CPU performance analysis | ⭐⭐ |
-| `advanced_control.py` | Multiple control strategies comparison | ⭐⭐⭐ |
+### 🎯 Basic Examples (⭐)
+Perfect for getting started with ManipulaPy fundamentals.
 
-### Running Examples
+| Example | Description | Output |
+|---------|-------------|--------|
+| `kinematics_basic_demo.py` | Forward/inverse kinematics with visualization | Manipulability analysis plots |
+| `dynamics_basic_demo.py` | Mass matrix, Coriolis forces, gravity compensation | Complete robot analysis |
+| `control_basic_demo.py` | PID, computed torque, feedforward control | Control strategy comparison |
+| `urdf_processing_basic_demo.py` | URDF to SerialManipulator conversion | Configuration space analysis |
+| `visualization_basic_demo.py` | End-effector paths and workspace visualization | 3D trajectory plots |
+
+### 🔧 Intermediate Examples (⭐⭐)
+Advanced features and integrated systems.
+
+| Example | Description | Key Features |
+|---------|-------------|--------------|
+| `trajectory_planning_intermediate_demo.py` | Multi-segment trajectories and optimization | GPU acceleration, smoothing |
+| `singularity_analysis_intermediate_demo.py` | Workspace analysis and singularity avoidance | Manipulability ellipsoids |
+| `control_comparison_intermediate_demo.py` | Multiple control strategies benchmarking | Real-time monitoring |
+| `perception_intermediate_demo.py` | Computer vision pipeline with clustering | YOLO detection, stereo vision |
+| `simulation_intermediate_demo.py` | Complete PyBullet integration | Real-time physics simulation |
+
+### 🚀 Advanced Examples (⭐⭐⭐)
+Research-grade implementations and high-performance computing.
+
+| Example | Description | Advanced Features |
+|---------|-------------|-------------------|
+| `gpu_acceleration_advanced_demo.py` | CUDA kernels and performance optimization | Memory efficiency analysis |
+| `batch_processing_advanced_demo.py` | Large-scale trajectory generation | Batch scaling analysis |
+| `collision_avoidance_advanced_demo.py` | Real-time obstacle avoidance | Potential field visualization |
+| `optimal_control_advanced_demo.py` | Advanced control algorithms | Performance statistics |
+| `stereo_vision_advanced_demo.py` | 3D perception and point cloud processing | Advanced perception analysis |
+| `real_robot_integration_advanced_demo.py` | Hardware integration examples | Real-time simulation |
+
+### 🏃‍♂️ Running Examples
 
 ```bash
 cd Examples/
 
-# Basic kinematics
-python basic_kinematics.py
+# Basic Examples - Start here!
[…]
