+++ tests/test_control.py
@@ -4,6 +4,7 @@ import unittest
 import numpy as np
 import cupy as cp
 import matplotlib.pyplot as plt
+from unittest.mock import patch, MagicMock
 from ManipulaPy.control import ManipulatorController
 from ManipulaPy.urdf_processor import URDFToSerialManipulator
 from ManipulaPy.ManipulaPy_data.xarm import urdf_file as xarm_urdf_file
@@ -45,13 +46,13 @@ class TestManipulatorController(unittest.TestCase):
 
             # Get the number of joints from the dynamics
             num_joints = len(self.dynamics.Glist)
-            self.thetalist = np.zeros(num_joints, dtype=np.float64)  # Fix dtype
-            self.dthetalist = np.zeros(num_joints, dtype=np.float64)  # Fix dtype
-            self.ddthetalist = np.zeros(num_joints, dtype=np.float64)  # Fix dtype
+            self.thetalist = np.zeros(num_joints, dtype=np.float32)
+            self.dthetalist = np.zeros(num_joints, dtype=np.float32)
+            self.ddthetalist = np.zeros(num_joints, dtype=np.float32)
 
             # Default joint and torque limits if not available
-            self.joint_limits = np.array([[-np.pi, np.pi]] * num_joints)
-            self.torque_limits = np.array([[-10, 10]] * num_joints)
+            self.joint_limits = np.array([[-np.pi, np.pi]] * num_joints, dtype=np.float32)
+            self.torque_limits = np.array([[-10, 10]] * num_joints, dtype=np.float32)
 
         except Exception as e:
             print(f"Error loading URDF: {e}")
@@ -63,22 +64,19 @@ class TestManipulatorController(unittest.TestCase):
         # Create a simplified dynamics object for testing
         class MockDynamics:
             def __init__(self):
-                self.Glist = np.array([np.eye(6), np.eye(6)])  # Mock inertia matrices
+                self.Glist = np.array([np.eye(6), np.eye(6)])
                 self.S_list = np.array(
                     [[0, 0, 1, 0, 0, 0], [0, 0, 1, 0, 0.1, 0]]
-                ).T  # Mock screw axes
-                self.M_list = np.eye(4)  # Mock home position
+                ).T
+                self.M_list = np.eye(4)
 
             def mass_matrix(self, thetalist):
-                # Return a simple diagonal mass matrix
                 return np.diag([1.0, 0.8])
 
             def velocity_quadratic_forces(self, thetalist, dthetalist):
-                # Simple Coriolis term
                 return np.array([0.01 * dthetalist[1] ** 2, 0.01 * dthetalist[0] ** 2])
 
             def gravity_forces(self, thetalist, g):
-                # Simple gravity term
                 return np.array(
                     [
                         0.5 * g[2] * np.sin(thetalist[0]),
@@ -87,35 +85,43 @@ class TestManipulatorController(unittest.TestCase):
                 )
 
             def inverse_dynamics(self, thetalist, dthetalist, ddthetalist, g, Ftip):
-                # Simple implementation for testing
                 M = self.mass_matrix(thetalist)
                 c = self.velocity_quadratic_forces(thetalist, dthetalist)
                 grav = self.gravity_forces(thetalist, g)
                 return M.dot(ddthetalist) + c + grav
 
             def forward_dynamics(self, thetalist, dthetalist, taulist, g, Ftip):
-                # Simple implementation for testing
                 M = self.mass_matrix(thetalist)
                 c = self.velocity_quadratic_forces(thetalist, dthetalist)
                 grav = self.gravity_forces(thetalist, g)
[…]
