+++ tests/conftest.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
 """
-Fixed conftest.py with bug fixes for all the test failures.
+Smart conftest.py that only mocks what's truly unavailable or GPU-only.
+Allows proper testing of CPU-capable libraries while gracefully handling missing dependencies.
 """
 
 import sys
@@ -8,7 +9,8 @@ import os
 from unittest.mock import Mock, MagicMock
 import warnings
 import numpy as np
-
+import matplotlib
+matplotlib.use('Agg')
 # Suppress warnings during testing
 warnings.filterwarnings("ignore", category=DeprecationWarning)
 warnings.filterwarnings("ignore", category=UserWarning)
@@ -17,7 +19,7 @@ warnings.filterwarnings("ignore", category=UserWarning)
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 
 class MockModule:
-    """Improved mock module that handles iteration properly."""
+    """Enhanced mock module that handles iteration and common operations properly."""
     
     def __init__(self, name=None):
         self._name = name or "MockModule"
@@ -41,15 +43,25 @@ class MockModule:
         return MockModule(f"{self._name}[{key}]")
 
 class CuPyArrayMock:
-    """Fixed CuPy array mock with proper numpy compatibility."""
+    """Enhanced CuPy array mock with proper numpy compatibility and type safety."""
     
-    def __init__(self, data):
-        # Use float64 to avoid casting issues
-        self._data = np.asarray(data, dtype=np.float64)
+    def __init__(self, data, dtype=None):
+        # Convert to appropriate dtype, defaulting to float32 for GPU simulation
+        if dtype is None:
+            dtype = np.float32 if np.issubdtype(np.asarray(data).dtype, np.floating) else np.asarray(data).dtype
+        self._data = np.asarray(data, dtype=dtype)
     
     def get(self):
         """CuPy's method to convert GPU array to CPU (numpy) array."""
         return self._data
+    
+    def copy_to_host(self):
+        """Alternative CuPy method for GPU->CPU transfer."""
+        return self._data.copy()
+    
+    def copy_to_device(self, src):
+        """Mock GPU memory copy."""
+        self._data[:] = np.asarray(src, dtype=self._data.dtype)
     
     def __getattr__(self, name):
         return getattr(self._data, name)
@@ -63,54 +75,54 @@ class CuPyArrayMock:
     def __str__(self):
         return str(self._data)
     
-    # Fixed math operations with proper type handling
+    # Math operations with proper type handling
     def __add__(self, other):
         if hasattr(other, '_data'):
             result = self._data + other._data
         else:
             result = self._data + np.asarray(other, dtype=self._data.dtype)
-        return CuPyArrayMock(result)
+        return CuPyArrayMock(result, dtype=self._data.dtype)
     
     def __sub__(self, other):
         if hasattr(other, '_data'):
             result = self._data - other._data
         else:
             result = self._data - np.asarray(other, dtype=self._data.dtype)
-        return CuPyArrayMock(result)
+        return CuPyArrayMock(result, dtype=self._data.dtype)
     
     def __mul__(self, other):
         if hasattr(other, '_data'):
             result = self._data * other._data
         else:
             result = self._data * np.asarray(other, dtype=self._data.dtype)
-        return CuPyArrayMock(result)
+        return CuPyArrayMock(result, dtype=self._data.dtype)
     
     def __matmul__(self, other):
         if hasattr(other, '_data'):
             result = self._data @ other._data
         else:
             result = self._data @ np.asarray(other, dtype=self._data.dtype)
-        return CuPyArrayMock(result)
[…]
