### 2025-07-22 boelnasr/ManipulaPy#6 @sea-bass
Docs look awesome!

Overall, I am worried that all the code snippets that were added into the documentation will be extremely difficult to maintain as the software changes. Normally you would want to cross-reference or link to other code files, instead of putting so much code in your docs that is likely to get out of date. But I will not block on that.

There are some rendering errors though. See below:

There are a few syntax warnings rendering in the docs about line underlines in titles being too short. See for example:

* https://manipulapy.readthedocs.io/en/latest/user_guide/Kinematics.html#product-of-exponentials-formula (and below)
* https://manipulapy.readthedocs.io/en/latest/user_guide/index.html#guide-categories (and below)
* https://manipulapy.readthedocs.io/en/latest/user_guide/Dynamics.html#computing-mass-matrix (and below)
* https://manipulapy.readthedocs.io/en/latest/user_guide/Trajectory_Planning.html#introduction (and below)
* https://manipulapy.readthedocs.io/en/latest/user_guide/Simulation.html#troubleshooting-guide (and below)
* https://manipulapy.readthedocs.io/en/latest/user_guide/URDF_Processor.html#introduction (and below)
* https://manipulapy.readthedocs.io/en/latest/user_guide/vision.html#stereo-vision-pipeline (and below)
* https://manipulapy.readthedocs.io/en/latest/user_guide/Potential_Field.html#introduction-to-potential-fields (and below)
* https://manipulapy.readthedocs.io/en/latest/user_guide/CUDA_Kernels.html#installation-options (and below)

Also some indentation errors in the equations in
* https://manipulapy.readthedocs.io/en/latest/user_guide/Dynamics.html#mathematical-background
* https://manipulapy.readthedocs.io/en/latest/user_guide/Trajectory_Planning.html#joint-space-time-scaling

Certain numbered lists are not rendering correctly, likely due to incorrect indentation. See e.g., here: https://manipulapy.readthedocs.io/en/latest/user_guide/URDF_Processor.html#summary

The sub-titles here don't render correctly, this is likely still using Markdown format and not RST: https://manipulapy.readthedocs.io/en/latest/user_guide/Potential_Field.html#advanced-topics
