### 2025-07-22 boelnasr/ManipulaPy#12 @sea-bass
Thanks for all these updates, it looks great!. Some remaining notes:

- [ ] Figures 1 and 3 are not showing up in the paper
- [ ] The bulleted/numbered lists are all rendering in one paragraph instead of breaking lines, probably an indentation error. See here:
