### 2025-07-22 boelnasr/ManipulaPy#18 @sea-bass
[JOSS Review] Correct author name in files
In looking at the submission again, I noticed that you have added the following copyright statement in all the files:

```
Copyright (c) 2025 Mohamed Aboelnar
```

Looking at your GitHub user name and JOSS Paper submission, I think there is a letter missing, but please forgive me if I am not correct. Should your second name be corrected to Aboelnasr (or Abo El-Nasr / AboElNasr depending on how you prefer the English character spelling)?

https://github.com/openjournals/joss-reviews/issues/8490
