### 2025-07-22 boelnasr/ManipulaPy#3 @sea-bass
Thank you! This all worked perfectly, except for the commands:

```python
import manipulapy
if manipulapy.cuda.is_available():
  ...
```

Since Python is case-sensitive, these all have to be `ManipulaPy` with the upper-case letters.

I think the only place where this was not correctly updated was in the top-level README, so it's not too many corrections that need to happen.
