### 2025-07-22 bbae7b4
paper corrections
add the missing figure and correcting the indentation
