### 2025-07-22 1078304
test updates
adding CUDA kernels updates
