+++ docs/source/user_guide/CUDA_Kernels.rst
@@ -1,5 +1,5 @@
 CUDA Kernels User Guide
-=======================
+==========================
 
 This guide covers the **CUDA acceleration** features in ManipulaPy, which provide GPU-accelerated functions for high-performance robotic computations including trajectory planning, dynamics computation, and potential field calculations.
 
@@ -8,7 +8,7 @@ This guide covers the **CUDA acceleration** features in ManipulaPy, which provid
    :depth: 2
 
 What is GPU Acceleration?
--------------------------
+----------------------------
 
 **GPU acceleration** uses NVIDIA graphics cards to perform parallel computations much faster than traditional CPU processing. ManipulaPy leverages CUDA through:
 
@@ -24,10 +24,10 @@ Key benefits include:
 - **Seamless Integration**: Works directly with existing ManipulaPy classes
 
 Mathematical Foundation
------------------------
+--------------------------
 
 CUDA Kernel Implementation
-~~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 **Trajectory Generation Kernel:**
 
@@ -60,10 +60,10 @@ Parallel inverse dynamics using the Newton-Euler formulation:
 Each GPU thread processes one trajectory point simultaneously.
 
 Installation and Setup
-----------------------
+-------------------------
 
 System Requirements
-~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~
 
 **Hardware:**
 - NVIDIA GPU with Compute Capability 3.0 or higher
@@ -76,7 +76,7 @@ System Requirements
 - Python 3.8+
 
 Installation Options
-~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~
 
 **Option 1: Full GPU Support (Recommended)**
 
@@ -96,7 +96,7 @@ Installation Options
    pip install ManipulaPy
 
 Verification
-~~~~~~~~~~~
+~~~~~~~~~~~~~~
 
 .. code-block:: python
 
@@ -123,10 +123,10 @@ Verification
        CUPY_AVAILABLE = False
 
 Available CUDA Kernels
-----------------------
+-------------------------
 
 Trajectory Generation Kernels
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 **trajectory_kernel**
 
@@ -157,7 +157,7 @@ Generates Cartesian space trajectories for end-effector motion.
    def cartesian_trajectory_kernel(pstart, pend, traj_pos, traj_vel, traj_acc, Tf, N, method):
 
 Dynamics Computation Kernels
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 **inverse_dynamics_kernel**
 
@@ -187,7 +187,7 @@ Simulates robot motion given applied torques.
        Glist, Slist, M, thetamat, dthetamat, ddthetamat, joint_limits):
 
 Potential Field Kernels
-~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 **attractive_potential_kernel**
 
@@ -217,10 +217,10 @@ Computes gradients of potential fields for navigation.
[…]
