+++ docs/source/user_guide/Trajectory_Planning.rst
@@ -4,7 +4,7 @@ Trajectory Planning User Guide
 This guide covers the trajectory planning capabilities in ManipulaPy, including joint-space and Cartesian-space trajectory generation, dynamics integration, and collision avoidance using CUDA acceleration.
 
 Introduction
------------
+----------------
 
 The TrajectoryPlanning class provides comprehensive trajectory generation and execution capabilities for robotic manipulators. It combines kinematic trajectory planning with dynamic analysis and collision avoidance to generate feasible, smooth robot motions.
 
@@ -17,39 +17,32 @@ The TrajectoryPlanning class provides comprehensive trajectory generation and ex
 - Support for various trajectory optimization objectives
 
 Mathematical Background
-~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 This section summarizes the key mathematical constructs behind joint-space and Cartesian trajectory generation, as well as obstacle avoidance via potential fields.
 
 Joint-Space Time Scaling
-------------------------
+----------------------------
 
 Given start :math:`\boldsymbol\theta_{0}` and end :math:`\boldsymbol\theta_{f}`, total duration :math:`T`, define a scalar time-scaling function
 
 .. math::
-
-   s = \frac{t}{T}, 
-   \quad s\in[0,1].
+   s = \frac{t}{T}, \quad s \in [0,1]
 
 Common choices:
 
 - **Cubic** (zero end-velocity):
-  
-  .. math::
 
-     \sigma_{3}(s)
-       = 3s^{2} - 2s^{3},
+  .. math::
+     \sigma_{3}(s) = 3s^{2} - 2s^{3}
 
 - **Quintic** (zero end-velocity & zero end-acceleration):
 
   .. math::
-
-     \sigma_{5}(s)
-       = 10s^{3} - 15s^{4} + 6s^{5}.
-
-Joint-space trajectory:
+     \sigma_{5}(s) = 10s^{3} - 15s^{4} + 6s^{5}
 
 .. math::
+   \mathbf{q}(t) = \mathbf{q}_{\text{start}} + \sigma(s) \cdot (\mathbf{q}_{\text{end}} - \mathbf{q}_{\text{start}})
 
    \boldsymbol\theta(t)
      = \boldsymbol\theta_{0}
@@ -67,8 +60,9 @@ with
 
 where :math:`m=3` or :math:`5` for cubic/quintic.
 
+
 Cartesian-Space Interpolation
------------------------------
+--------------------------------
 
 Given start/end end-effector poses :math:`X_{0},X_{f}\in SE(3)`, define the relative transform
 
@@ -149,10 +143,10 @@ This mathematical framework underlies all high-level methods in the `TrajectoryP
 
 
 Basic Usage
-----------
+---------------
 
 Setting Up Trajectory Planning
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 .. code-block:: python
 
@@ -180,7 +174,7 @@ Setting Up Trajectory Planning
    print("Trajectory planner initialized successfully")
 
 Simple Joint Trajectory
-~~~~~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 .. code-block:: python
 
@@ -208,10 +202,10 @@ Simple Joint Trajectory
    np.testing.assert_allclose(trajectory['positions'][-1], theta_end, rtol=1e-3)
 
 TrajectoryPlanning Class
------------------------
[…]
