+++ docs/source/user_guide/Potential_Field.rst
@@ -8,26 +8,26 @@ This comprehensive guide covers the potential field path-planning tools in Manip
    and performance tips.
 
 Introduction to Potential Fields
--------------------------------
+--------------------------------------
 
 Potential field methods treat the robot as a point moving under the influence of an artificial
-potential surface.  The robot is “pulled” toward the goal by an _attractive_ potential and
-“pushed” away from obstacles by _repulsive_ potentials.  By following the negative gradient
+potential surface. The robot is "pulled" toward the goal by an **attractive** potential and
+"pushed" away from obstacles by **repulsive** potentials. By following the negative gradient
 of the combined field, the robot can find a collision-free path.
 
 Key components:
 
-- **Attractive potential**  draws the robot toward the goal.  
-- **Repulsive potential**   pushes the robot away from obstacles within an influence radius.  
-- **Gradient**             direction of steepest descent in the combined potential.  
-- **Collision checking**   rapid test whether a given configuration is in collision.  
+- **Attractive potential** draws the robot toward the goal.  
+- **Repulsive potential** pushes the robot away from obstacles within an influence radius.  
+- **Gradient** direction of steepest descent in the combined potential.  
+- **Collision checking** rapid test whether a given configuration is in collision.  
 
 Mathematical Background
 ~~~~~~~~~~~~~~~~~~~~~~~
 
-Let :math:`\mathbf q \in \mathbb R^n` be the robot’s configuration (e.g., joint angles).  
-Let :math:`\mathbf q_{\mathrm{goal}}` be the goal configuration, and let  
-:math:`\{\,\mathbf o_i\in\mathbb R^n\}` be a set of obstacle points.
+Let :math:`\mathbf{q} \in \mathbb{R}^n` be the robot's configuration (e.g., joint angles).  
+Let :math:`\mathbf{q}_{\mathrm{goal}}` be the goal configuration, and let  
+:math:`\{\mathbf{o}_i \in \mathbb{R}^n\}` be a set of obstacle points.
 
 Attractive Potential
 ^^^^^^^^^^^^^^^^^^^^
@@ -35,10 +35,10 @@ Attractive Potential
 A common quadratic attractive potential:
 
 .. math::
-   U_{\mathrm{att}}(\mathbf q)
-     = \tfrac12\,k_{\mathrm{att}}\,\|\mathbf q - \mathbf q_{\mathrm{goal}}\|^2
+   U_{\mathrm{att}}(\mathbf{q})
+     = \frac{1}{2} k_{\mathrm{att}} \|\mathbf{q} - \mathbf{q}_{\mathrm{goal}}\|^2
 
-where :math:`k_{\mathrm{att}}>0` is the attractive gain.
+where :math:`k_{\mathrm{att}} > 0` is the attractive gain.
 
 Repulsive Potential
 ^^^^^^^^^^^^^^^^^^^
@@ -46,16 +46,15 @@ Repulsive Potential
 An inverse-distance repulsive potential, active only within an influence distance :math:`d_0`:
 
 .. math::
-   U_{\mathrm{rep}}(\mathbf q)
+   U_{\mathrm{rep}}(\mathbf{q})
      = \sum_{i}
        \begin{cases}
-         \tfrac12\,k_{\mathrm{rep}}\!\bigl(\tfrac{1}{d_i} - \tfrac{1}{d_0}\bigr)^2
-           & d_i \le d_0,\\[6pt]
[…]
