### 2025-07-22 boelnasr/ManipulaPy#3 @boelnasr
After installation, confirm that ManipulaPy can see your GPU:

```bash
# Check that CUDA is available to ManipulaPy
python3 - <<EOF
from ManipulaPy import cuda_kernels

if cuda_kernels.check_cuda_availability():
    props = cuda_kernels.get_gpu_properties()
    print(f"✅ CUDA is available on device: {props['name']} "
          f"({props['multiprocessor_count']} SMs, "
          f"{props['max_threads_per_block']} max threads/block)")
else:
    raise RuntimeError("❌ CUDA not detected or not properly configured in ManipulaPy.")
EOF

```

If you see the ✅ message with your GPU name, you’re all set! Otherwise, double‑check the CUDA Toolkit and cuDNN installation steps above. \`\`\`
sorry for the oversight i have updated the syntax, and add the link for the cuDNN documentation

1. **NVIDIA Drivers & CUDA Toolkit**
   - `nvcc` on your `PATH` (e.g. via `sudo apt install nvidia-cuda-toolkit` or the [official NVIDIA CUDA installer](https://developer.nvidia.com/cuda-downloads)).
   - Verify with:
     ```bash
     nvidia-smi       # should list your GPU(s) and driver version
     nvcc --version   # should print CUDA version
     ```

2. **cuDNN**
   - Download and install cuDNN for your CUDA version from [NVIDIA's cuDNN installation guide](https://docs.nvidia.com/deeplearning/cudnn/installation/latest/).
   - Verify headers/libs under `/usr/include` and `/usr/lib/x86_64-linux-gnu` (or your distro’s equivalent).
