### 2025-07-22 a634850
docs fix
fixing some docs syntax errors
