### 2025-07-22 boelnasr/ManipulaPy#6 @boelnasr
Thank you for the detailed feedback! You're absolutely right about code maintenance being a problem.
Why Keep Code Snippets (For Now)
Robotics code is complex and users need to see working examples with real values - things like joint limits, control settings, and how different parts work together aren't obvious from just API docs. The code snippets help people actually get started instead of getting stuck on setup and configuration, which is really common in robotics libraries.
Future Plan
You're right about the maintenance issue. Here's our plan:
Now: Keep examples in docs to help users learn
Later: Auto-generate examples from tested code
Future: Interactive tutorials
The auto-generated approach you mentioned is exactly right - we can have real code files that get tested and automatically included in the docs. This way examples stay working and up-to-date without manual work.

i have also resolved the rendering issues in the docs
