### 2025-07-22 boelnasr/ManipulaPy#12 @sea-bass
Thank you for the fixes -- looks great now.
