### 2025-07-22 boelnasr/ManipulaPy#9 @sea-bass
Where is the coverage reported?

I looked at the CI logs and this is just running 2 skipped tests due to CUDA unavailability:

https://github.com/boelnasr/ManipulaPy/actions/runs/16456335540/job/46514177103
