+++ pyproject.toml
@@ -8,7 +8,7 @@ build-backend = "setuptools.build_meta"
 
 # ------------------------------------------------------------------
 [project]
-name = "ManipulaPy"                    # exact PyPI casing
+name = "ManipulaPy"
 version = "1.1.1"                      
 authors = [
     { name = "Mohamed Aboelnar", email = "aboelnasr1997@gmail.com" },
@@ -41,39 +41,59 @@ keywords = [
 ]
 
 # ------------------------------------------------------------------
-# Core runtime dependencies - FIXED: Removed problematic GPU dependencies from core
+# Core runtime dependencies
 dependencies = [
     "numpy>=1.19.2",
     "scipy>=1.5.2",
-    "pybullet>=3.0.6",
-    "trimesh>=3.9.14",
-    "urchin>=0.0.27",
-    "numba>=0.55",
     "matplotlib>=3.3",
-    "scikit-learn>=1.0",
-    "opencv-python>=4.5",
-    "ultralytics>=8.0",
-    "torch>=1.8.0",
+    "numba>=0.55",
 ]
 
 [project.optional-dependencies]
-# GPU support - Multiple options to handle different CUDA versions
+# Core optional dependencies
+core-optional = [
+    "scikit-learn>=1.0",
+]
+
+# Simulation dependencies
+simulation = [
+    "pybullet>=3.0.6",
+    "urchin>=0.0.27",
+    "trimesh>=3.9.14",
+]
+
+# Vision dependencies  
+vision = [
+    "opencv-python>=4.5",
+    "ultralytics>=8.0",
+    "pillow>=8.0.0",
+]
+
+# AI/ML dependencies
+ml = [
+    "torch>=1.8.0",
+    "scikit-learn>=1.0",
+]
+
+# GPU support
 gpu-cuda11 = ["cupy-cuda11x>=10.0.0"]
 gpu-cuda12 = ["cupy-cuda12x>=12.0.0"]
 gpu-rocm = ["cupy-rocm-4-3>=10.0.0"]
-# Legacy PyCUDA (alternative to CuPy)
 gpu-pycuda = ["pycuda>=2021.1"]
-# All GPU options combined
-gpu = ["cupy-cuda11x>=10.0.0"]  # Default to CUDA 11
+gpu = ["cupy-cuda11x>=10.0.0"]
 
-# Development dependencies
+# Development and testing dependencies
 dev = [
     "pytest>=6.0.0",
-    "pytest-cov>=2.0.0",
-    "black>=21.0.0",
-    "flake8>=3.8.0",
-    "isort>=5.0.0",
-    "mypy>=0.910",
+    "pytest-cov>=3.0.0",
+    "pytest-mock>=3.6.0",
+    "pytest-xvfb>=2.0.0",
+    "pytest-timeout>=2.1.0",
+    "coverage[toml]>=6.0.0",
+    "black>=22.0.0",
+    "flake8>=4.0.0",
+    "isort>=5.10.0",
+    "mypy>=0.950",
     "pre-commit>=2.15.0",
 ]
 
@@ -84,13 +104,15 @@ docs = [
     "myst-parser>=0.15.0",
 ]
 
-# Testing with GPU support
-test-gpu = [
-    "ManipulaPy[dev,gpu-cuda11]"
-]
+# Testing combinations
+test-core = ["ManipulaPy[dev]"]
+test-simulation = ["ManipulaPy[dev,simulation]"]
+test-vision = ["ManipulaPy[dev,vision]"]
+test-gpu = ["ManipulaPy[dev,gpu-cuda11]"]
+test-full = ["ManipulaPy[dev,simulation,vision,ml]"]
 
[…]
