+++ tests/conftest.py
@@ -1,7 +1,13 @@
 #!/usr/bin/env python3
 """
-Smart conftest.py that only mocks what's truly unavailable or GPU-only.
-Allows proper testing of CPU-capable libraries while gracefully handling missing dependencies.
+Smart conftest.py for ManipulaPy test suite.
+
+This configuration:
+- Only mocks what's truly unavailable or GPU-only
+- Allows proper testing of CPU-capable libraries when available
+- Provides comprehensive test fixtures and utilities
+- Handles CI environments gracefully
+- Enables proper coverage reporting
 """
 
 import sys
@@ -9,17 +15,31 @@ import os
 from unittest.mock import Mock, MagicMock
 import warnings
 import numpy as np
+import pytest
+import types
+
+# Set up matplotlib for headless testing
 import matplotlib
 matplotlib.use('Agg')
-# Suppress warnings during testing
+
+# Suppress common warnings during testing
 warnings.filterwarnings("ignore", category=DeprecationWarning)
-warnings.filterwarnings("ignore", category=UserWarning)
+warnings.filterwarnings("ignore", category=UserWarning) 
+warnings.filterwarnings("ignore", category=FutureWarning)
+warnings.filterwarnings("ignore", category=PendingDeprecationWarning)
 
 # Add the package to the path
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+
+# Force CPU-only execution
 os.environ['CUDA_VISIBLE_DEVICES'] = ''
 os.environ['NUMBA_DISABLE_CUDA'] = '1'
 os.environ['MANIPULAPY_FORCE_CPU'] = '1'
+os.environ['TORCH_USE_CUDA_DSA'] = '0'
+
+# ============================================================================
+# Enhanced Mock Classes
+# ============================================================================
 
 class MockModule:
     """Enhanced mock module that handles iteration and common operations properly."""
@@ -49,7 +69,6 @@ class CuPyArrayMock:
     """Enhanced CuPy array mock with proper numpy compatibility and type safety."""
     
     def __init__(self, data, dtype=None):
-        # Convert to appropriate dtype, defaulting to float32 for GPU simulation
         if dtype is None:
             dtype = np.float32 if np.issubdtype(np.asarray(data).dtype, np.floating) else np.asarray(data).dtype
         self._data = np.asarray(data, dtype=dtype)
@@ -74,9 +93,6 @@ class CuPyArrayMock:
     
     def __repr__(self):
         return f"CuPyArrayMock({self._data})"
-    
-    def __str__(self):
-        return str(self._data)
     
     # Math operations with proper type handling
     def __add__(self, other):
@@ -112,20 +128,6 @@ class CuPyArrayMock:
             self._data += other._data.astype(self._data.dtype)
         else:
             self._data += np.asarray(other, dtype=self._data.dtype)
-        return self
-    
-    def __isub__(self, other):
-        if hasattr(other, '_data'):
-            self._data -= other._data.astype(self._data.dtype)
-        else:
-            self._data -= np.asarray(other, dtype=self._data.dtype)
-        return self
-    
-    def __imul__(self, other):
-        if hasattr(other, '_data'):
-            self._data *= other._data.astype(self._data.dtype)
-        else:
-            self._data *= np.asarray(other, dtype=self._data.dtype)
         return self
     
     # Array properties
@@ -210,31 +212,6 @@ class CuPyMock:
             
         return CuPyArrayMock(np.clip(data, a_min, a_max))
     
-    # Math functions
-    def sqrt(self, arr):
-        data = arr._data if hasattr(arr, '_data') else np.asarray(arr)
-        return CuPyArrayMock(np.sqrt(data))
-    
[…]
