+++ .github/workflows/test.yml
@@ -1,173 +1,411 @@
-name: CI
+name: CI Tests
+
 on:
   push:
-    branches: [ main ]
+    branches: [ main, develop ]
   pull_request:
-    branches: [ main ]
+    branches: [ main, develop ]
+
 jobs:
   test:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.8, 3.9, '3.10']
+        python-version: [3.8, 3.9, "3.10", 3.11]
       fail-fast: false
+        
     steps:
     - name: Checkout code
-      uses: actions/checkout@v3
+      uses: actions/checkout@v4
     
     - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
+      uses: actions/setup-python@v5
       with:
         python-version: ${{ matrix.python-version }}
     
+    - name: Install system dependencies
+      run: |
+        sudo apt-get update
+        sudo apt-get install -y \
+          libgl1-mesa-dev \
+          libglu1-mesa-dev \
+          libegl1-mesa-dev \
+          libxrandr2 \
+          libxinerama1 \
+          libxcursor1 \
+          libxi6 \
+          libxtst6 \
+          xvfb
+
+    - name: Cache pip dependencies
+      uses: actions/cache@v4
+      with:
+        path: ~/.cache/pip
+        key: ${{ runner.os }}-pip-${{ matrix.python-version }}-${{ hashFiles('**/requirements.txt', '**/pyproject.toml') }}
+        restore-keys: |
+          ${{ runner.os }}-pip-${{ matrix.python-version }}-
+          ${{ runner.os }}-pip-
+
+    - name: Install base dependencies
+      run: |
+        python -m pip install --upgrade pip setuptools wheel
+        pip install pytest pytest-cov pytest-mock coverage[toml]
+
+    - name: Install PyTorch (CPU version for CI)
+      run: |
+        pip install torch --index-url https://download.pytorch.org/whl/cpu
+        python -c "import torch; print(f'✅ PyTorch {torch.__version__} installed successfully')"
+
+    - name: Install ManipulaPy in development mode
+      run: |
+        pip install -e .
+
     - name: Install test dependencies
       run: |
-        python -m pip install --upgrade pip
-        pip install pytest pytest-cov pytest-mock numpy scipy matplotlib
-    
-    - name: Create conftest.py for mocking and custom CLI flags
+        pip install \
+          numpy>=1.19.0 \
+          scipy>=1.6.0 \
+          matplotlib>=3.3.0 \
+          scikit-learn>=1.0.0 \
+          pillow>=8.0.0
+
+    - name: Install optional dependencies (continue on error)
+      continue-on-error: true
       run: |
-        mkdir -p tests
-        cat > tests/conftest.py << 'EOF'
+        # These may not be available in CI, but install if possible
+        pip install pybullet || echo "PyBullet installation failed (expected in CI)"
+        pip install opencv-python || echo "OpenCV installation failed"
+        pip install urchin || echo "URCHIN installation failed"
+        pip install ultralytics || echo "Ultralytics installation failed (expected in CI)"
+
+    - name: Verify ManipulaPy installation
+      run: |
+        # Test core package import
+        python -c "import ManipulaPy; print(f'ManipulaPy version: {ManipulaPy.__version__}')"
+        
+        # Test core modules (should always work)
+        python -c "from ManipulaPy import utils; print('✅ Utils module imported successfully')"
+        python -c "import torch; print(f'✅ PyTorch {torch.__version__} available')"
+        
+        # Test core functionality (requires torch)
+        python -c "
+        try:
+            from ManipulaPy import kinematics, dynamics
[…]
