+++ ManipulaPy/__init__.py
@@ -4,8 +4,8 @@
 """
 ManipulaPy Package
 
-This package provides tools for the analysis and manipulation of robotic systems, including kinematics,
-dynamics, singularity analysis, path planning, and URDF processing utilities.
+This package provides tools for the analysis and manipulation of robotic systems,
+including kinematics, dynamics, singularity analysis, path planning, and URDF processing utilities.
 
 License: GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 Copyright (c) 2025 Mohamed Aboelnar
@@ -24,44 +24,148 @@ You should have received a copy of the GNU Affero General Public License
 along with this program. If not, see <https://www.gnu.org/licenses/>.
 """
 
+import warnings
+import sys
 
-
-
-# ---------------------------------------------------------------------
-# Re-export key submodules for convenient top-level access
-# (import order chosen to avoid circular dependencies)
-# ---------------------------------------------------------------------
-from ManipulaPy.kinematics import *
-from ManipulaPy.dynamics import *
-from ManipulaPy.singularity import *
-from ManipulaPy.path_planning import *
-from ManipulaPy.utils import *
-from ManipulaPy.urdf_processor import *
-from ManipulaPy.vision import *
-from ManipulaPy.perception import *
-from ManipulaPy.control import *
-from ManipulaPy.sim import *
-from ManipulaPy.potential_field import *
-from ManipulaPy.cuda_kernels import *
-
-# ---------------------------------------------------------------------
 # Package metadata
-# ---------------------------------------------------------------------
-__version__ = "1.1.1"            
-__author__  = "Mohamed Aboelnar"
+__version__ = "1.1.1"
+__author__ = "Mohamed Aboelnar"
 __license__ = "AGPL-3.0-or-later"
 
-__all__ = [
-    "kinematics",
-    "dynamics",
-    "singularity",
-    "path_planning",
-    "utils",
-    "urdf_processor",
-    "vision",
-    "perception",
-    "control",
-    "sim",
-    "potential_field",
-    "cuda_kernels",
-]
+# Core modules that should always be available
+_CORE_MODULES = []
+_OPTIONAL_MODULES = []
+
+# Import core utilities (should always work)
+try:
+    from ManipulaPy.utils import *
+    from ManipulaPy import utils
+    _CORE_MODULES.append("utils")
+except ImportError as e:
+    warnings.warn(f"Failed to import utils module: {e}", ImportWarning)
+
+# Import core functionality that depends on torch
+try:
+    from ManipulaPy.kinematics import *
+    from ManipulaPy import kinematics
+    _CORE_MODULES.append("kinematics")
+except ImportError as e:
+    warnings.warn(f"Failed to import kinematics module: {e}. Ensure PyTorch is installed.", ImportWarning)
+
+try:
+    from ManipulaPy.dynamics import *
+    from ManipulaPy import dynamics
+    _CORE_MODULES.append("dynamics")
+except ImportError as e:
+    warnings.warn(f"Failed to import dynamics module: {e}. Ensure PyTorch is installed.", ImportWarning)
+
+try:
+    from ManipulaPy.singularity import *
+    from ManipulaPy import singularity
+    _CORE_MODULES.append("singularity")
+except ImportError as e:
+    warnings.warn(f"Failed to import singularity module: {e}", ImportWarning)
+
+try:
+    from ManipulaPy.path_planning import *
+    from ManipulaPy import path_planning
+    _CORE_MODULES.append("path_planning")
+except ImportError as e:
+    warnings.warn(f"Failed to import path_planning module: {e}", ImportWarning)
+
+try:
[…]
