+++ tests/__init__.py
@@ -1,0 +1,298 @@
+#!/usr/bin/env python3
+"""
+Test suite for ManipulaPy package initialization.
+
+This module tests the graceful import handling, version information,
+and module availability in the ManipulaPy package.
+"""
+
+import sys
+import warnings
+import importlib
+from unittest.mock import patch, MagicMock
+import pytest
+
+
+class TestManipulaPyInit:
+    """Test ManipulaPy package initialization and imports."""
+
+    def test_package_version(self):
+        """Test that ManipulaPy has correct version information."""
+        import ManipulaPy
+        
+        assert hasattr(ManipulaPy, '__version__')
+        assert hasattr(ManipulaPy, '__author__')
+        assert hasattr(ManipulaPy, '__license__')
+        
+        assert ManipulaPy.__version__ == "1.1.1"
+        assert ManipulaPy.__author__ == "Mohamed Aboelnar"
+        assert ManipulaPy.__license__ == "AGPL-3.0-or-later"
+
+    def test_package_import_success(self):
+        """Test that ManipulaPy can be imported successfully."""
+        import ManipulaPy
+        assert ManipulaPy is not None
+
+    def test_core_modules_available(self):
+        """Test that core modules are available."""
+        import ManipulaPy
+        
+        # These should always be available as core modules
+        expected_core = ['utils']  # utils should always work
+        
+        for module_name in expected_core:
+            assert hasattr(ManipulaPy, module_name), f"Core module {module_name} not available"
+
+    def test_torch_dependent_modules(self):
+        """Test modules that depend on PyTorch."""
+        import ManipulaPy
+        
+        # These modules require torch (which should be installed as core dependency now)
+        torch_dependent = ['kinematics', 'dynamics']
+        
+        for module_name in torch_dependent:
+            if hasattr(ManipulaPy, module_name):
+                module = getattr(ManipulaPy, module_name)
+                assert module is not None, f"Module {module_name} is None"
+
+    def test_optional_modules_graceful_handling(self):
+        """Test that optional modules are handled gracefully when missing."""
+        import ManipulaPy
+        
+        # These are optional modules that may or may not be available
+        optional_modules = ['vision', 'perception', 'sim', 'cuda_kernels']
+        
+        # Should not crash if missing, just not be in __all__
+        for module_name in optional_modules:
+            # It's OK if these don't exist - they're optional
+            if hasattr(ManipulaPy, module_name):
+                module = getattr(ManipulaPy, module_name)
+                # If present, should not be None
+                assert module is not None, f"Optional module {module_name} is present but None"
+
+    def test_all_list_contains_only_imported_modules(self):
+        """Test that __all__ only contains successfully imported modules."""
+        import ManipulaPy
+        
+        assert hasattr(ManipulaPy, '__all__')
+        assert isinstance(ManipulaPy.__all__, list)
+        
+        # Every item in __all__ should be available as an attribute
+        for module_name in ManipulaPy.__all__:
+            assert hasattr(ManipulaPy, module_name), f"Module {module_name} in __all__ but not available"
+
+    def test_import_with_missing_torch(self):
+        """Test graceful handling when PyTorch is missing."""
+        # This test simulates what happens if torch is not available
+        with patch.dict('sys.modules', {'torch': None}):
+            with warnings.catch_warnings(record=True) as w:
+                warnings.simplefilter("always", ImportWarning)
+                
+                # Force reimport
+                if 'ManipulaPy' in sys.modules:
+                    del sys.modules['ManipulaPy']
+                if 'ManipulaPy.kinematics' in sys.modules:
+                    del sys.modules['ManipulaPy.kinematics']
+                if 'ManipulaPy.dynamics' in sys.modules:
+                    del sys.modules['ManipulaPy.dynamics']
+                
+                import ManipulaPy
+                
[…]
