+++ pyproject.toml
@@ -1,457 +1,422 @@
-# ──────────────────────────────────────────────────────────────
-# Project build configuration for ManipulaPy
-# ──────────────────────────────────────────────────────────────
+name: CI Tests
 
-[build-system]
-requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta"
+on:
+  push:
+    branches: [ main, develop ]
+  pull_request:
+    branches: [ main, develop ]
 
-# ------------------------------------------------------------------
-[project]
-name = "ManipulaPy"
-version = "1.1.1"                      
-authors = [
-    { name = "Mohamed Aboelnar", email = "aboelnasr1997@gmail.com" },
-]
-description = "A comprehensive Python package for robotic manipulator analysis and control"
-readme = "README.md"
-license = { text = "AGPL-3.0-or-later" }
-requires-python = ">=3.8"
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.8, 3.9, "3.10", 3.11]
+      fail-fast: false
+        
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v4
+    
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v5
+      with:
+        python-version: ${{ matrix.python-version }}
+    
+    - name: Install system dependencies
+      run: |
+        sudo apt-get update
+        sudo apt-get install -y \
+          libgl1-mesa-dev \
+          libglu1-mesa-dev \
+          libegl1-mesa-dev \
+          libxrandr2 \
+          libxinerama1 \
+          libxcursor1 \
+          libxi6 \
+          libxtst6 \
+          xvfb
 
-classifiers = [
-    "Development Status :: 4 - Beta",
-    "Intended Audience :: Science/Research",
-    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Topic :: Scientific/Engineering :: Artificial Intelligence",
-    "Topic :: Scientific/Engineering :: Physics",
-    "Topic :: Software Development :: Libraries :: Python Modules",
-]
-keywords = [
-    "robotics",
-    "kinematics",
-    "dynamics",
-    "trajectory-planning",
-    "simulation",
-    "cuda",
-]
+    - name: Cache pip dependencies
+      uses: actions/cache@v4
+      with:
+        path: ~/.cache/pip
+        key: ${{ runner.os }}-pip-${{ matrix.python-version }}-${{ hashFiles('**/requirements.txt', '**/pyproject.toml') }}
+        restore-keys: |
+          ${{ runner.os }}-pip-${{ matrix.python-version }}-
+          ${{ runner.os }}-pip-
 
-# ------------------------------------------------------------------
-# Core runtime dependencies
-dependencies = [
-    "numpy>=1.19.2",
-    "scipy>=1.5.2",
-    "matplotlib>=3.3",
-    "numba>=0.55",
-]
+    - name: Install base dependencies
+      run: |
+        python -m pip install --upgrade pip setuptools wheel
+        pip install pytest pytest-cov pytest-mock pytest-timeout coverage[toml]
 
-[project.optional-dependencies]
-# Core optional dependencies
-core-optional = [
-    "scikit-learn>=1.0",
-]
+    - name: Install PyTorch (CPU version for CI)
+      run: |
+        pip install torch --index-url https://download.pytorch.org/whl/cpu
[…]
