+++ .github/workflows/test.yml
@@ -49,7 +49,7 @@ jobs:
     - name: Install base dependencies
       run: |
         python -m pip install --upgrade pip setuptools wheel
-        pip install pytest pytest-cov pytest-mock coverage[toml]
+        pip install pytest pytest-cov pytest-mock pytest-timeout coverage[toml]
 
     - name: Install PyTorch (CPU version for CI)
       run: |
@@ -58,7 +58,27 @@ jobs:
 
     - name: Install ManipulaPy in development mode
       run: |
-        pip install -e .
+        # First install core dependencies manually to avoid TOML parsing issues
+        pip install numpy>=1.19.2 scipy>=1.5.2 matplotlib>=3.3 numba>=0.55
+        
+        # Then install the package
+        pip install -e . || {
+          echo "❌ pip install -e . failed, trying alternative approach"
+          echo "Installing dependencies manually..."
+          pip install -r <(python -c "
+import sys
+deps = [
+    'numpy>=1.19.2',
+    'scipy>=1.5.2', 
+    'matplotlib>=3.3',
+    'numba>=0.55'
+]
+for dep in deps:
+    print(dep)
+")
+          export PYTHONPATH="${GITHUB_WORKSPACE}:$PYTHONPATH"
+          echo "Using PYTHONPATH approach instead of editable install"
+        }
 
     - name: Install test dependencies
       run: |
@@ -80,9 +100,40 @@ jobs:
 
     - name: Verify ManipulaPy installation
       run: |
+        # Test core package import
         python -c "import ManipulaPy; print(f'ManipulaPy version: {ManipulaPy.__version__}')"
-        python -c "from ManipulaPy import kinematics, dynamics, utils; print('✅ Core modules imported successfully')"
-        python -c "import torch; print(f'✅ PyTorch integration working: {torch.__version__}')"
+        
+        # Test core modules (should always work)
+        python -c "from ManipulaPy import utils; print('✅ Utils module imported successfully')"
+        python -c "import torch; print(f'✅ PyTorch {torch.__version__} available')"
+        
+        # Test core functionality (requires torch)
+        python -c "
+        try:
+            from ManipulaPy import kinematics, dynamics
+            print('✅ Core kinematics and dynamics modules imported successfully')
+        except ImportError as e:
+            print(f'⚠️  Core modules import failed: {e}')
+            exit(1)
+        "
+        
+        # Test optional modules (may fail gracefully)
+        python -c "
+        import warnings
+        warnings.filterwarnings('ignore', category=ImportWarning)
+        
+        try:
+            from ManipulaPy import vision
+            print('✅ Vision module imported successfully')
+        except ImportError:
+            print('ℹ️  Vision module not available (missing ultralytics)')
+        
+        try:
+            from ManipulaPy import simulation  
+            print('✅ Simulation module imported successfully')
+        except ImportError:
+            print('ℹ️  Simulation module not available (optional dependencies missing)')
+        "
 
     - name: Set up display for tests
       run: |
@@ -96,11 +147,22 @@ jobs:
         DISPLAY: ":99"
         QT_QPA_PLATFORM: "offscreen"
         MPLBACKEND: "Agg"
-        SKIP_CUDA_TESTS: "true"
+        SKIP_CUDA_TESTS: "true" 
         SKIP_VISION_TESTS: "true"
         SKIP_SIMULATION_TESTS: "true"
       run: |
-        # Run tests that don't require special dependencies
+        # Check if tests directory exists and list test files
+        echo "📁 Checking test directory contents:"
+        if [ -d "tests/" ]; then
+          find tests/ -name "*.py" -type f | head -10
[…]
