### 2025-07-23 90fbe3a
Revert "CI test covarage"
This reverts commit b6b8fc9a5291e6507d9f470041b21a41f7e33f38.
