### 2025-07-23 81a5c10
CI test discovery
