+++ pyproject.toml
@@ -30,6 +30,9 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Artificial Intelligence",
     "Topic :: Scientific/Engineering :: Physics",
     "Topic :: Software Development :: Libraries :: Python Modules",
+    "Operating System :: OS Independent",
+    "Intended Audience :: Developers",
+    "Intended Audience :: Education",
 ]
 keywords = [
     "robotics",
@@ -38,6 +41,9 @@ keywords = [
     "trajectory-planning",
     "simulation",
     "cuda",
+    "computer-vision",
+    "control-systems",
+    "manipulator",
 ]
 
 # ------------------------------------------------------------------
@@ -69,6 +75,13 @@ vision = [
     "pillow>=8.0.0",
 ]
 
+# Vision dependencies for headless environments (CI-friendly)
+vision-headless = [
+    "opencv-python-headless>=4.5",
+    "ultralytics>=8.0",
+    "pillow>=8.0.0",
+]
+
 # AI/ML dependencies
 ml = [
     "torch>=1.8.0",
@@ -89,12 +102,14 @@ dev = [
     "pytest-mock>=3.6.0",
     "pytest-xvfb>=2.0.0",
     "pytest-timeout>=2.1.0",
+    "pytest-benchmark>=4.0.0",
     "coverage[toml]>=6.0.0",
     "black>=22.0.0",
     "flake8>=4.0.0",
     "isort>=5.10.0",
     "mypy>=0.950",
     "pre-commit>=2.15.0",
+    "tox>=4.0.0",
 ]
 
 # Documentation dependencies
@@ -102,14 +117,30 @@ docs = [
     "sphinx>=4.0.0",
     "sphinx-rtd-theme>=1.0.0",
     "myst-parser>=0.15.0",
+    "sphinx-autodoc-typehints>=1.12.0",
+    "nbsphinx>=0.8.0",
+    "jupyter>=1.0.0",
+]
+
+# CI-specific dependencies
+ci = [
+    "pytest>=6.0.0",
+    "pytest-cov>=3.0.0",
+    "pytest-mock>=3.6.0",
+    "pytest-xvfb>=2.0.0",
+    "pytest-timeout>=2.1.0",
+    "coverage[toml]>=6.0.0",
+    "opencv-python-headless>=4.5",
+    "tomli>=1.2.0;python_version<'3.11'",
 ]
 
 # Testing combinations
 test-core = ["ManipulaPy[dev]"]
 test-simulation = ["ManipulaPy[dev,simulation]"]
-test-vision = ["ManipulaPy[dev,vision]"]
+test-vision = ["ManipulaPy[dev,vision-headless]"]
 test-gpu = ["ManipulaPy[dev,gpu-cuda11]"]
-test-full = ["ManipulaPy[dev,simulation,vision,ml]"]
+test-full = ["ManipulaPy[dev,simulation,vision-headless,ml]"]
+test-ci = ["ManipulaPy[ci]"]
 
 # All optional dependencies
 all = ["ManipulaPy[core-optional,simulation,vision,ml,docs]"]
@@ -119,11 +150,18 @@ Homepage = "https://github.com/boelnasr/ManipulaPy"
 Documentation = "https://manipulapy.readthedocs.io/"
 Repository = "https://github.com/boelnasr/ManipulaPy.git"
 "Bug Tracker" = "https://github.com/boelnasr/ManipulaPy/issues"
+"Changelog" = "https://github.com/boelnasr/ManipulaPy/blob/main/CHANGELOG.md"
+"Funding" = "https://github.com/sponsors/boelnasr"
 
 # ------------------------------------------------------------------
 [tool.setuptools]
 packages = ["ManipulaPy"]
[…]
