### 2025-07-23 b7db636
CI test discovery
