+++ .github/workflows/test.yml
@@ -35,7 +35,12 @@ jobs:
           libxcursor1 \
           libxi6 \
           libxtst6 \
-          xvfb
+          xvfb \
+          libglib2.0-0 \
+          libsm6 \
+          libxext6 \
+          libxrender-dev \
+          libfontconfig1
 
     - name: Cache pip dependencies
       uses: actions/cache@v4
@@ -51,6 +56,35 @@ jobs:
         python -m pip install --upgrade pip setuptools wheel
         pip install pytest pytest-cov pytest-mock coverage[toml]
 
+    - name: Debug TOML file
+      run: |
+        echo "🔍 Checking pyproject.toml syntax..."
+        python -c "
+        import sys
+        try:
+            import tomllib
+        except ImportError:
+            try:
+                import tomli as tomllib
+            except ImportError:
+                print('Installing tomli...')
+                import subprocess
+                subprocess.check_call([sys.executable, '-m', 'pip', 'install', 'tomli'])
+                import tomli as tomllib
+
+        with open('pyproject.toml', 'rb') as f:
+            try:
+                data = tomllib.load(f)
+                print('✅ TOML file is valid')
+                if 'project' in data and 'dependencies' in data['project']:
+                    print(f'✅ Found dependencies: {data[\"project\"][\"dependencies\"]}')
+                else:
+                    print('⚠️ No dependencies found in TOML')
+            except Exception as e:
+                print(f'❌ TOML error: {e}')
+                sys.exit(1)
+        "
+
     - name: Install PyTorch (CPU version for CI)
       run: |
         pip install torch --index-url https://download.pytorch.org/whl/cpu
@@ -58,7 +92,7 @@ jobs:
 
     - name: Install ManipulaPy in development mode
       run: |
-        pip install -e .
+        pip install -e .[dev]
 
     - name: Install test dependencies
       run: |
@@ -67,14 +101,16 @@ jobs:
           scipy>=1.6.0 \
           matplotlib>=3.3.0 \
           scikit-learn>=1.0.0 \
-          pillow>=8.0.0
+          pillow>=8.0.0 \
+          pytest-xvfb>=2.0.0 \
+          pytest-timeout>=2.1.0
 
     - name: Install optional dependencies (continue on error)
       continue-on-error: true
       run: |
         # These may not be available in CI, but install if possible
         pip install pybullet || echo "PyBullet installation failed (expected in CI)"
-        pip install opencv-python || echo "OpenCV installation failed"
+        pip install opencv-python-headless || pip install opencv-python || echo "OpenCV installation failed"
         pip install urchin || echo "URCHIN installation failed"
         pip install ultralytics || echo "Ultralytics installation failed (expected in CI)"
 
@@ -90,6 +126,54 @@ jobs:
         Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
         sleep 3
 
+    - name: Check existing test suite
+      run: |
+        echo "📁 Checking existing test suite..."
+        ls -la tests/
+        echo "Total test files: $(find tests/ -name 'test_*.py' | wc -l)"
+        
+        # Verify conftest.py exists
+        if [ -f tests/conftest.py ]; then
+            echo "✅ conftest.py found"
+        else
+            echo "⚠️ No conftest.py found"
+        fi
+        
+        # Check pytest configuration
+        if [ -f pyproject.toml ]; then
+            echo "✅ pyproject.toml found"
+        fi
+        
+        # Test pytest discovery with detailed output
+        echo "🔍 Testing pytest discovery..."
+        python -m pytest --collect-only tests/ -q || echo "⚠️ Test collection has issues"
+
[…]
