### 2025-07-23 925759c
Vision test correction
