+++ ManipulaPy/vision.py
@@ -181,12 +181,15 @@ class Vision:
         self.show_plot = show_plot
 
         # Load YOLO Model for Object Detection
-        try:
-            self.yolo_model = YOLO("yolov8m.pt")  # Ensure this model file is available
-            self.logger.info("✅ YOLO model loaded successfully.")
-        except Exception as e:
-            self.logger.error(f"❌ Failed to load YOLO model: {e}")
-            self.yolo_model = None  # Prevent crashes
+# Load YOLO (optional)
+        self.yolo_model = None
+        if ULTRALYTICS_AVAILABLE:
+            try:
+                self.yolo_model = YOLO("yolov8m.pt")
+                self.logger.info("✅ YOLO model loaded successfully.")
+            except Exception as e:
+                self.logger.warning(f"YOLO load failed ({e}); continuing without it.")
+
 
         # Camera configuration (Monocular or Stereo)
         self.cameras = {}
@@ -498,72 +501,93 @@ class Vision:
     # --------------------------------------------------------------------------
     # Basic Obstacle Detection
     # --------------------------------------------------------------------------
-
     def detect_obstacles(
-        self, depth_image, rgb_image, depth_threshold=0.0, camera_index=0, step=5
+        self,
+        depth_image: np.ndarray,
+        rgb_image: np.ndarray,
+        depth_threshold: float = 0.0,
+        camera_index: int = 0,
+        step: int = 5,
     ):
         """
-        Detects obstacles using YOLO for object detection + median depth for 3D positioning.
-        Returns positions and orientations (in the XY plane).
-        """
-        if self.yolo_model is None:
-            self.logger.warning("⚠️ YOLO model is missing! Skipping object detection.")
-            return np.empty((0, 3)), np.array([])
+        Detect obstacles with YOLO (if available) and estimate 3D positions via depth.
 
+        Returns
+        -------
+        positions : (N, 3) float32 array
+        orientations : (N,) float32 array  (yaw angles in degrees in XY plane)
+        """
+        # --------- Basic sanity checks ----------
         if depth_image is None or rgb_image is None:
             self.logger.error("❌ Invalid depth or RGB input")
-            return np.empty((0, 3)), np.array([])
+            return np.empty((0, 3), dtype=np.float32), np.empty((0,), dtype=np.float32)
 
-        # 1. YOLO inference
-        results = self.yolo_model(rgb_image, conf=0.3)
-        if not results or results[0].boxes is None or len(results[0].boxes) == 0:
-            self.logger.warning("❌ No objects detected by YOLO")
-            return np.empty((0, 3)), np.array([])
+        if camera_index not in self.cameras:
+            self.logger.error(f"Camera index {camera_index} not found.")
+            return np.empty((0, 3), dtype=np.float32), np.empty((0,), dtype=np.float32)
 
-        boxes = results[0].boxes  # No filtering by class!
+        if self.yolo_model is None:
+            self.logger.warning("⚠️ YOLO model is missing! Skipping object detection.")
+            return np.empty((0, 3), dtype=np.float32), np.empty((0,), dtype=np.float32)
+
+        # --------- YOLO inference (guarded) ----------
+        try:
+            results = self.yolo_model(rgb_image, conf=0.3)
+        except Exception as e:
+            self.logger.warning(f"YOLO inference failed ({e}); returning empty.")
+            return np.empty((0, 3), dtype=np.float32), np.empty((0,), dtype=np.float32)
+
+        if not results or getattr(results[0], "boxes", None) is None or len(results[0].boxes) == 0:
+            self.logger.info("No objects detected by YOLO.")
[…]
