+++ .github/workflows/test.yml
@@ -242,163 +242,271 @@ jobs:
           
           # Create the enhanced analysis script
           cat > analyze_coverage.py << 'EOF'
-        #!/usr/bin/env python3
-        """Enhanced Coverage Analysis for ManipulaPy"""
-        import xml.etree.ElementTree as ET
-        import sys
-        from pathlib import Path
-        
-        def analyze_coverage(xml_file="coverage.xml"):
-            """Extract detailed coverage information per module."""
-            
-            if not Path(xml_file).exists():
-                print(f"❌ Coverage file {xml_file} not found")
-                return None
-                
-            try:
-                tree = ET.parse(xml_file)
-                root = tree.getroot()
-                
-                # Overall statistics
-                overall_coverage = float(root.attrib['line-rate']) * 100
-                overall_branch = float(root.attrib.get('branch-rate', 0)) * 100
-                
-                print(f"📊 Overall Coverage: {overall_coverage:.1f}% (lines)")
-                print(f"🌿 Overall Branch Coverage: {overall_branch:.1f}%")
-                print("=" * 60)
-                
-                # Per-module detailed analysis
-                modules = []
-                for package in root.findall('.//package'):
-                    name = package.attrib['name'].replace('ManipulaPy.', '').replace('ManipulaPy', 'root')
-                    if not name or name == 'root':  # Skip empty names or root
-                        continue
-                        
-                    line_rate = float(package.attrib['line-rate']) * 100
-                    branch_rate = float(package.attrib.get('branch-rate', 0)) * 100
-                    
-                    # Count total lines and covered lines
-                    total_lines = 0
-                    covered_lines = 0
-                    
-                    for line in package.findall('.//line'):
-                        total_lines += 1
-                        if int(line.attrib['hits']) > 0:
-                            covered_lines += 1
-                    
-                    modules.append({
-                        'name': name,
-                        'line_rate': line_rate,
-                        'branch_rate': branch_rate,
-                        'total_lines': total_lines,
-                        'covered_lines': covered_lines,
-                        'uncovered_lines': total_lines - covered_lines
-                    })
-                
-                # Sort by coverage rate (descending)
-                modules.sort(key=lambda x: x['line_rate'], reverse=True)
-                
-                # Display results
-                print(f"{'Module':<20} {'Lines':<8} {'Branches':<10} {'Total':<8} {'Covered':<8} {'Missing':<8}")
-                print("-" * 70)
-                
-                for module in modules:
-                    # Status emoji
-                    if module['line_rate'] >= 90:
-                        emoji = "🟢"
-                    elif module['line_rate'] >= 80:
-                        emoji = "✅"
-                    elif module['line_rate'] >= 60:
-                        emoji = "⚠️"
-                    else:
-                        emoji = "❌"
-                        
-                    print(f"{emoji} {module['name']:<18} "
-                          f"{module['line_rate']:>6.1f}% "
-                          f"{module['branch_rate']:>8.1f}% "
-                          f"{module['total_lines']:>6} "
-                          f"{module['covered_lines']:>6} "
-                          f"{module['uncovered_lines']:>6}")
-                
-                print("=" * 60)
-                
-                # Summary statistics
-                high_coverage = sum(1 for m in modules if m['line_rate'] >= 80)
[…]
