+++ .github/workflows/test.yml
@@ -242,271 +242,163 @@ jobs:
           
           # Create the enhanced analysis script
           cat > analyze_coverage.py << 'EOF'
-#!/usr/bin/env python3
-"""Enhanced Coverage Analysis for ManipulaPy"""
-import xml.etree.ElementTree as ET
-import sys
-from pathlib import Path
-
-def format_missing_lines(lines):
-    """Format missing line numbers into ranges like coverage.py"""
-    if not lines:
-        return ""
-    
-    lines = sorted(lines)
-    ranges = []
-    start = lines[0]
-    end = lines[0]
-    
-    for line in lines[1:]:
-        if line == end + 1:
-            end = line
-        else:
-            if start == end:
-                ranges.append(str(start))
-            else:
-                ranges.append(f"{start}-{end}")
-            start = end = line
-    
-    # Add the last range
-    if start == end:
-        ranges.append(str(start))
-    else:
-        ranges.append(f"{start}-{end}")
-    
-    # Limit output length to avoid too long lines
-    result = ", ".join(ranges)
-    if len(result) > 80:
-        result = result[:77] + "..."
-    return result
-
-def analyze_coverage(xml_file="coverage.xml"):
-    """Extract detailed coverage information per module."""
-    
-    if not Path(xml_file).exists():
-        print(f"❌ Coverage file {xml_file} not found")
-        return None
+        #!/usr/bin/env python3
+        """Enhanced Coverage Analysis for ManipulaPy"""
+        import xml.etree.ElementTree as ET
+        import sys
+        from pathlib import Path
         
-    try:
-        tree = ET.parse(xml_file)
-        root = tree.getroot()
-        
-        # Overall statistics
-        overall_coverage = float(root.attrib['line-rate']) * 100
-        overall_branch = float(root.attrib.get('branch-rate', 0)) * 100
-        
-        print(f"📊 Overall Coverage: {overall_coverage:.1f}% (lines)")
-        print(f"🌿 Overall Branch Coverage: {overall_branch:.1f}%")
-        print("=" * 60)
-        
-        # Per-module detailed analysis
-        modules = []
-        for package in root.findall('.//package'):
-            name = package.attrib['name'].replace('ManipulaPy.', '').replace('ManipulaPy', 'root')
-            if not name or name == 'root':  # Skip empty names or root
-                continue
+        def analyze_coverage(xml_file="coverage.xml"):
+            """Extract detailed coverage information per module."""
+            
+            if not Path(xml_file).exists():
+                print(f"❌ Coverage file {xml_file} not found")
+                return None
                 
-            line_rate = float(package.attrib['line-rate']) * 100
-            branch_rate = float(package.attrib.get('branch-rate', 0)) * 100
-            
-            # Detailed line analysis from classes
-            total_stmts = 0
-            missed_stmts = 0
-            total_branches = 0
-            partial_branches = 0
-            missing_lines = []
-            
-            for cls in package.findall('.//class'):
-                for line in cls.findall('.//line'):
-                    line_num = int(line.attrib['number'])
-                    hits = int(line.attrib['hits'])
-                    branch = line.attrib.get('branch', 'false') == 'true'
-                    
-                    total_stmts += 1
-                    if hits == 0:
-                        missed_stmts += 1
-                        missing_lines.append(line_num)
-                    
-                    if branch:
-                        total_branches += 1
-                        # Check for partial branch coverage
-                        condition_coverage = line.attrib.get('condition-coverage', '')
-                        if condition_coverage and '(' in condition_coverage:
-                            # Parse condition coverage like "50% (1/2)"
-                            try:
[…]
