+++ ManipulaPy/ManipulaPy_data/xarm/__init__.py
@@ -1,281 +1,71 @@
 #!/usr/bin/env python3
 """
-xArm Robot Data Module
+XARM Robot Model Data
 
-This module provides URDF files and mesh data for the xArm robotic manipulator.
+This module provides access to XARM robot URDF files and related assets.
 """
 
 import os
-from pathlib import Path
+import pkg_resources
 
-# Get the directory containing this module
-_MODULE_DIR = Path(__file__).parent
+# Get the path to this module
+_module_path = os.path.dirname(__file__)
 
-# Available URDF files in the xarm directory
-AVAILABLE_URDFS = {
-    'base': _MODULE_DIR / "base.urdf",
-    'base_com': _MODULE_DIR / "base_com.urdf", 
-    'xarm6_robot': _MODULE_DIR / "xarm6_robot.urdf",
-    'xarm6_robot_white': _MODULE_DIR / "xarm6_robot_white.urdf",
-    'xarm6_with_gripper': _MODULE_DIR / "xarm6_with_gripper.urdf",
-    'link1': _MODULE_DIR / "link1.urdf",
-    'link1_com': _MODULE_DIR / "link1_com.urdf",
-    'link2': _MODULE_DIR / "link2.urdf",
-    'link2_com': _MODULE_DIR / "link2_com.urdf",
-    'link3': _MODULE_DIR / "link3.urdf",
-    'link3_com': _MODULE_DIR / "link3_com.urdf",
-    'link4': _MODULE_DIR / "link4.urdf",
-    'link4_com': _MODULE_DIR / "link4_com.urdf",
-    'link5': _MODULE_DIR / "link5.urdf",
-    'link5_com': _MODULE_DIR / "link5_com.urdf",
-    'link6': _MODULE_DIR / "link6.urdf",
-    'link6_com': _MODULE_DIR / "link6_com.urdf",
-}
-
-# Default URDF file (the main complete robot)
-urdf_file = str(AVAILABLE_URDFS['xarm6_robot'])
-
-# Fallback if the main URDF doesn't exist
-if not AVAILABLE_URDFS['xarm6_robot'].exists():
-    # Try alternative URDFs
-    for name, path in AVAILABLE_URDFS.items():
-        if path.exists():
-            urdf_file = str(path)
-            print(f"Warning: Using fallback URDF: {name}")
-            break
-    else:
-        # If no URDF files exist, create a basic one
-        urdf_file = str(_MODULE_DIR / "basic_xarm.urdf")
-        _create_basic_xarm_urdf()
-
-def _create_basic_xarm_urdf():
-    """Create a basic xArm URDF file for testing purposes."""
-    basic_urdf_content = '''<?xml version="1.0"?>
-<robot name="xarm6">
-  <!-- Base Link -->
-  <link name="base_link">
-    <visual>
-      <geometry>
-        <box size="0.2 0.2 0.1"/>
-      </geometry>
-      <material name="gray">
-        <color rgba="0.5 0.5 0.5 1"/>
-      </material>
-    </visual>
-    <inertial>
-      <mass value="2.0"/>
-      <inertia ixx="0.01" ixy="0" ixz="0" iyy="0.01" iyz="0" izz="0.01"/>
-    </inertial>
-  </link>
-
-  <!-- Link 1 -->
-  <link name="link1">
-    <visual>
-      <geometry>
-        <cylinder radius="0.06" length="0.267"/>
-      </geometry>
-      <material name="blue">
-        <color rgba="0.2 0.2 0.8 1"/>
-      </material>
-    </visual>
-    <inertial>
-      <mass value="1.5"/>
[…]
