+++ Examples/advanced_examples/batch_processing_advanced_demo.py
@@ -18,7 +18,7 @@ Features demonstrated:
 - Performance profiling and optimization
 - Real-time batch processing capabilities
 
-Copyright (c) 2025 Mohamed Aboelnar
+Copyright (c) 2025 Mohamed Aboelnasr
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 """
 
