+++ Examples/basic_examples/urdf_processing_basic_demo.py
@@ -10,7 +10,8 @@ URDFToSerialManipulator class.
 Usage:
     python urdf_processing_basic_demo.py
 
-Author: ManipulaPy Development Team
+Copyright (c) 2025 Mohamed Aboelnasr
+Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 """
 
 import numpy as np
