+++ Examples/intermediate_examples/simulation_intermediate_demo.py
@@ -14,7 +14,7 @@ This script demonstrates the intermediate capabilities of the ManipulaPy library
 
 All generated plots are saved in the same folder as this script.
 
-Copyright (c) 2025 Mohamed Aboelnar
+Copyright (c) 2025 Mohamed Aboelnasr
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 """
 
