+++ Examples/intermediate_examples/trajectory_planning_intermediate_demo.py
@@ -13,7 +13,7 @@ This demo showcases advanced trajectory planning capabilities including:
 - Trajectory smoothing and filtering
 - Performance benchmarking and optimization
 
-Copyright (c) 2025 Mohamed Aboelnar
+Copyright (c) 2025 Mohamed Aboelnasr
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 """
 
