+++ ManipulaPy/vision.py
@@ -6,7 +6,7 @@ Vision Module - ManipulaPy
 This module provides computer vision capabilities for robotic perception including
 stereo vision, object detection, and camera calibration with optional YOLO integration.
 
-Copyright (c) 2025 Mohamed Aboelnar
+Copyright (c) 2025 Mohamed Aboelnasr
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 
 This file is part of ManipulaPy.
