+++ ManipulaPy/dynamics.py
@@ -6,7 +6,7 @@ Dynamics Module - ManipulaPy
 This module provides classes and functions for manipulator dynamics analysis including
 mass matrix computation, Coriolis forces, gravity compensation, and inverse/forward dynamics.
 
-Copyright (c) 2025 Mohamed Aboelnar
+Copyright (c) 2025 Mohamed Aboelnasr
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 
 This file is part of ManipulaPy.
