+++ Examples/basic_examples/kinematics_basic_demo.py
@@ -12,7 +12,7 @@ This demo showcases the fundamental control capabilities of ManipulaPy including
 - Controller tuning with Ziegler-Nichols method
 - Performance analysis and visualization
 
-Copyright (c) 2025 Mohamed Aboelnar
+Copyright (c) 2025 Mohamed Aboelnasr
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 """
 
