### 2025-07-23 boelnasr/ManipulaPy#18 @boelnasr
Thanks for the remark, i have corrected the name in all the files it's **Aboelnasr**
