### 2025-07-23 boelnasr/ManipulaPy#17 @boelnasr
This point is already discussed extensively in the last paper draft, you can check it now

let me know if this is sufficient
