+++ Benchmark/performance_benchmark.py
@@ -15,7 +15,7 @@ Features:
 - Scalability analysis
 - Hardware capability assessment
 
-Copyright (c) 2025 Mohamed Aboelnar
+Copyright (c) 2025 Mohamed Aboelnasr
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 """
 
