### 2025-07-23 boelnasr/ManipulaPy#4 @boelnasr
i have added checks for the benchmark scripts to check for pandas and seaborn if not installed to be installed automatically
