### 2025-07-23 boelnasr/ManipulaPy#11 @boelnasr
The "No module named 'ManipulaPy.ManipulaPy_data'" error that occurred with PyPI installations (but not editable installs) has been completely fixed through packaging improvements.
* MANIFEST.in now uses graft ManipulaPy/ManipulaPy_data to comprehensively include all robot data files
* setup.py simplified to use include_package_data=True as the single source of truth
Eliminates the editable vs. non-editable install discrepancy
