+++ pyproject.toml
@@ -9,9 +9,9 @@ build-backend = "setuptools.build_meta"
 # ------------------------------------------------------------------
 [project]
 name = "ManipulaPy"
-version = "1.1.1"                      
+version = "1.1.2"                      
 authors = [
-    { name = "Mohamed Aboelnar", email = "aboelnasr1997@gmail.com" },
+    { name = "Mohamed Aboelnasr", email = "aboelnasr1997@gmail.com" },
 ]
 description = "A comprehensive Python package for robotic manipulator analysis and control"
 readme = "README.md"
