+++ Examples/basic_examples/kinematics_basic_demo.py
@@ -1,1222 +1,777 @@
 #!/usr/bin/env python3
 # SPDX-License-Identifier: AGPL-3.0-or-later
 """
-Basic Control Demo - ManipulaPy
+Basic Kinematics Demo: Forward and Inverse Kinematics Fundamentals
 
-This demo showcases the fundamental control capabilities of ManipulaPy including:
-- PID Control
-- Computed Torque Control  
-- PD with Feedforward Control
-- Joint Space Control
-- Cartesian Space Control
-- Controller tuning with Ziegler-Nichols method
-- Performance analysis and visualization
+This example demonstrates fundamental kinematic operations for robotic manipulators including
+forward kinematics computation, inverse kinematics solving, Jacobian analysis, and basic
+visualization of robot configurations and workspace.
 
-Copyright (c) 2025 Mohamed Aboelnasr
-Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
+Usage:
+    python kinematics_basic_demo.py
+
+Expected Output:
+    - Console output showing joint angles and end-effector poses
+    - Matplotlib plots of robot configurations and Jacobian analysis
+    - Convergence information for inverse kinematics
+    - Workspace visualization and joint limit analysis
+
+Author: ManipulaPy Development Team
 """
 
 import numpy as np
 import matplotlib.pyplot as plt
-import time
-import logging
 from pathlib import Path
-import matplotlib
-matplotlib.use("Agg")  # prevents any GUI backend from being invoked
+import sys
+import time
 
-# ManipulaPy imports
+# Add ManipulaPy to path if needed
 try:
-    from ManipulaPy.kinematics import SerialManipulator
-    from ManipulaPy.dynamics import ManipulatorDynamics
-    from ManipulaPy.control import ManipulatorController
-    from ManipulaPy.urdf_processor import URDFToSerialManipulator
-    from ManipulaPy import utils
-    from ManipulaPy.ManipulaPy_data.xarm import urdf_file
-except ImportError as e:
-    print(f"Error importing ManipulaPy modules: {e}")
-    print("Please ensure ManipulaPy is properly installed.")
-    exit(1)
-
-# Optional GPU acceleration
-try:
-    import cupy as cp
-    GPU_AVAILABLE = True
-    print("✅ GPU acceleration available")
+    import ManipulaPy
 except ImportError:
-    GPU_AVAILABLE = False
-    print("⚠️ GPU acceleration not available, using CPU only")
+    sys.path.append(str(Path(__file__).parent.parent))
+    import ManipulaPy
 
-# Set up logging
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
-logger = logging.getLogger(__name__)
+from ManipulaPy.urdf_processor import URDFToSerialManipulator
+from ManipulaPy.kinematics import SerialManipulator
+import matplotlib
+matplotlib.use('TkAgg')
 
-class BasicControlDemo:
+class KinematicsBasicDemo:
     """
-    Demonstrates basic control algorithms for robotic manipulators using the built-in XArm robot.
+    Comprehensive demonstration of basic kinematics operations.
     """
     
-    def __init__(self, use_simple_robot=False):
-        """
-        Initialize the control demo.
+    def __init__(self):
+        """Initialize the demo with robot model loading."""
+        self.robot = None
+        self.joint_limits = None
+        self.n_joints = 0
         
-        Args:
-            use_simple_robot: If True, creates a simple 3-DOF robot. 
-                             If False (default), uses the built-in XArm robot.
-        """
-        self.use_simple_robot = use_simple_robot
-        self.setup_robot()
-        self.setup_controller()
[…]
