+++ ManipulaPy/__init__.py
@@ -8,7 +8,7 @@ This package provides tools for the analysis and manipulation of robotic systems
 including kinematics, dynamics, singularity analysis, path planning, and URDF processing utilities.
 
 License: GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
-Copyright (c) 2025 Mohamed Aboelnasr
+Copyright (c) 2025 Mohamed Aboelnar
 
 This program is free software: you can redistribute it and/or modify
 it under the terms of the GNU Affero General Public License as published by
@@ -26,15 +26,52 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 import warnings
 import sys
+import os
+import pkg_resources
 
 # Package metadata
-__version__ = "1.1.2"
-__author__ = "Mohamed Aboelnasr"
+__version__ = "1.1.2"  # Updated version
+__author__ = "Mohamed Aboelnar"
 __license__ = "AGPL-3.0-or-later"
 
 # Core modules that should always be available
 _CORE_MODULES = []
 _OPTIONAL_MODULES = []
+
+# Check if ManipulaPy_data is available and accessible
+def _check_data_availability():
+    """Check if ManipulaPy_data is available and accessible."""
+    try:
+        # Method 1: Try to access as package resource
+        try:
+            pkg_resources.resource_exists(__name__, 'ManipulaPy_data')
+            return True
+        except:
+            pass
+        
+        # Method 2: Try direct path access
+        try:
+            package_dir = os.path.dirname(__file__)
+            data_dir = os.path.join(package_dir, 'ManipulaPy_data')
+            return os.path.exists(data_dir) and os.path.isdir(data_dir)
+        except:
+            pass
+        
+        return False
+    except:
+        return False
+
+# Set global flag for data availability
+_DATA_AVAILABLE = _check_data_availability()
+
+if not _DATA_AVAILABLE:
+    warnings.warn(
+        "ManipulaPy_data directory not found. Some examples may not work. "
+        "This can happen with certain pip installations. "
+        "Try installing from source: pip install git+https://github.com/boelnasr/ManipulaPy.git",
+        ImportWarning,
+        stacklevel=2
+    )
 
 # Import core utilities (should always work)
 try:
@@ -143,6 +180,84 @@ except ImportError as e:
 # Define what gets imported with "from ManipulaPy import *"
 __all__ = _CORE_MODULES + _OPTIONAL_MODULES
 
+# Add data availability info to __all__
+__all__ += ['_DATA_AVAILABLE']
+
+def get_data_path(relative_path=''):
+    """
+    Get the path to ManipulaPy_data files.
+    
+    Args:
+        relative_path (str): Relative path within ManipulaPy_data directory
+        
+    Returns:
+        str: Full path to the requested file/directory
+        
+    Raises:
+        FileNotFoundError: If ManipulaPy_data is not available
+    """
+    if not _DATA_AVAILABLE:
+        raise FileNotFoundError(
+            "ManipulaPy_data directory not found. "
+            "Try installing from source: pip install git+https://github.com/boelnasr/ManipulaPy.git"
+        )
+    
+    try:
+        # Method 1: Use pkg_resources (works with pip installs)
+        try:
+            if relative_path:
+                return pkg_resources.resource_filename(__name__, f'ManipulaPy_data/{relative_path}')
+            else:
+                return pkg_resources.resource_filename(__name__, 'ManipulaPy_data')
+        except:
+            pass
+        
+        # Method 2: Direct path (works with development installs)
+        package_dir = os.path.dirname(__file__)
+        data_dir = os.path.join(package_dir, 'ManipulaPy_data')
+        if relative_path:
+            return os.path.join(data_dir, relative_path)
+        else:
+            return data_dir
[…]
