+++ tests/test_cuda_kernels.py
@@ -6,7 +6,7 @@ This test suite provides thorough testing of all CUDA-accelerated functionality
 including trajectory generation, dynamics computation, potential fields, and
 performance benchmarking with proper error handling and fallback mechanisms.
 
-Copyright (c) 2025 Mohamed Aboelnar
+Copyright (c) 2025 Mohamed Aboelnasr
 Licensed under the GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 """
 
