### 2025-07-24 boelnasr/ManipulaPy#11 @boelnasr
Right now the issue is resolved, the current pypi package version is v1.1.3
