### 2025-07-25 #8490 @sea-bass
@boelnasr has satisfied all of my comments, and in doing so has (in my opinion) also met the quality bar for publication in JOSS. Thank you for addressing these!

I do have one moderate concern about the author's liberal use of generative AI tools to respond to issues and assist them in generating fixes (evidenced by the extensive use of things like emojis in doc/examples and "helpful assistant" tone in responses). I had to test the outcomes and iterate a few times with the author until my issues were solved, and new issues popped up during this process, which we also identified and fixes. As such, I worry that in maintaining this package, users will also receive similar treatment that uses LLMs to hastily address issues in a way that compromises quality.

I only see a brief [policy on AI usage](https://joss.readthedocs.io/en/latest/review_criteria.html#using-ai-to-generate-code), so I will led the editor (@lockwo) and the second reviewer (@cadojo) decide how to proceed. For me, the code is currently in good shape but the use of generative AI during this review experience puts me at a "borderline accept".
