### 2025-07-30 #8490 @sea-bass
@lockwo thanks for your response. I think this assessment is fair, so pending the second review I think we should press on!

@boelnasr thanks for clarifying your use of LLMs and amending your repo to address the concerns.
