### 2025-08-16 boelnasr/ManipulaPy#17 @cadojo
Thanks @boelnasr. The latest paper looks good, though I don't see any mention of limitations. That is okay --- there is good discussion about alternative software tools --- but I recommend you add clarity about what features require a GPU, and what features can be used in CPU-only environments. I have been able to replicate some kinematics examples without a GPU, but no control or visualization features.
