### 2025-08-16 boelnasr/ManipulaPy#19 @cadojo
[JOSS Review] Incomplete Example in Dynamics.rst
There appears to be some incompleteness in the Dynamics user guide. Does this trailing content imply there's more you meant to add?

https://github.com/boelnasr/ManipulaPy/blob/f76866a7aae08fa9e275ff6ac9559c1fd38554ee/docs/source/user_guide/Dynamics.rst?plain=1#L354

---

This issue was opened as part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8490).
