### 2025-08-16 boelnasr/ManipulaPy#20 @jdc-pub
[JOSS Review] Clarify Meaning of Real Time Applications
There is a section of documentation referring to real-time applications. There are many different uses of the phrase _real-time_. Could you clarify what real-time means in this case? Hard real-time? Soft real-time? There's no true real-time behavior in Python, so I was confused reading through this section.

https://manipulapy.readthedocs.io/en/latest/user_guide/Trajectory_Planning.html#real-time-applications

---

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8490).
