### 2025-08-16 boelnasr/ManipulaPy#21 @cadojo
[JOSS Review] Missing Newline for List Syntax
In various places in documentation, and in the paper, there is no newline between the list's header (bold text) and the list items. As a result, the list does not render correctly. These should be fixed by adding a newline before the first list item.

Example: https://manipulapy.readthedocs.io/en/latest/user_guide/Trajectory_Planning.html#real-time-applications

---

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8490).
