### 2025-08-16 boelnasr/ManipulaPy#22 @cadojo
[JOSS Review] Clarify Requirements for Controls and Trajectory Planning
I was unable to execute the controls and trajectory planning examples shown in documentation. There is a helpful note clarifying CUDA requirements in the [Simulation](https://manipulapy.readthedocs.io/en/latest/user_guide/Simulation.html) section of the documentation. This should also be shown for all sections which require CUDA or GPU support.

As described in #17, it would be helpful to summarize what features are CPU-only compatible in documentation, and also in the paper.

I also see failures executing vision code examples. On importing `cv2` in `ManipulaPy.vision`, the following `ImportError` is raised. This does not need to be fixed, but requirements should be clarified.

```
ImportError: libGL.so.1: cannot open shared object file: No such file or directory
```

---

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8490).
