### 2025-08-16 boelnasr/ManipulaPy#23 @cadojo
[JOSS Review] Clarify Performance Benefits for e.g. 7-DOF Manipulators
The paper discusses custom kernels for 6-DOF manipulators. When I read this I wondered: are there performance benefits for higher-DOF robots? I expect there are, and I see 7-DOF examples, but I'm unclear on whether 7-DOF (and higher) manipulators would benefit from any custom kernels.

https://github.com/boelnasr/ManipulaPy/blob/f76866a7aae08fa9e275ff6ac9559c1fd38554ee/paper/paper.md?plain=1#L147

---

This issue is part of a [JOSS Review](https://github.com/openjournals/joss-reviews/issues/8490).
