### 2025-08-16 #8490 @cadojo
I have completed my first review pass. I have opened several new issues, mostly requesting clarification in documentation and in the paper. I have also opened one pull request, which is not required for acceptance, but includes some (hopefully helpful) syntax fixes to documentation and the paper. I am ready to recommend acceptance once these issues are addressed, pending my questions below.

---

I have two questions for you, @lockwo. Thanks for your patience and help with this process.

1. The paper liberally uses bullet lists to communicate. The information is presented clearly, but this format is a bit unusual for journal papers. I am okay with this format personally. **Is this bullet-list format acceptable?**

2. I have been unable to replicate performance claims due to my not having a GPU accessible, and I have not been able to execute all usage examples (see [#22](https://github.com/boelnasr/ManipulaPy/issues/22)) for the same reason. I tried to find cloud GPU options for my evaluation of this project, but that led to more delays, so I simply verified what I could with my CPU-only computer. There is ample documentation and benchmark scripts, so performance claims _can_ be verified. I am okay with accepting the submission without these examples verified. **Is this partial verification acceptable?**
