### 2025-08-16 #8490 @sea-bass
@cadojo if it helps, I have a GPU and was able to run the benchmark code that was added.
