### 2025-08-17 38c2bd1
Issue #22 and Issue #17
addressing the issues #22 and #17
