+++ README.md
@@ -59,12 +59,65 @@ ManipulaPy is a modern, comprehensive framework that bridges the gap between bas
 
 ---
 
+## 📋 Feature Availability Matrix
 
+ManipulaPy automatically enables features based on available dependencies. Here's what you can expect:
+
+### Core Features (Always Available)
+
+| Feature | CPU Performance | Dependencies | Notes |
+|---------|----------------|--------------|-------|
+| **Kinematics** | Excellent | numpy, scipy | Forward/inverse kinematics, Jacobians |
+| **Basic Dynamics** | Good | numpy, scipy | Mass matrix, Coriolis, gravity |
+| **Control Systems** | Excellent | numpy, scipy | PID, computed torque, adaptive |
+| **URDF Processing** | Fast | pybullet, urchin | Robot model conversion |
+| **Small Trajectories** | Good | numba | N < 1000 points, auto-optimized |
+
+### GPU-Accelerated Features (Optional)
+
+| Feature | CPU vs GPU | Requirements | Speedup |
+|---------|------------|--------------|---------|
+| **Large Trajectories** | 40x+ faster | CUDA, cupy | N > 1000 points |
+| **Batch Processing** | 20x+ faster | CUDA, cupy | Multiple trajectories |
+| **Inverse Dynamics** | 100x+ faster | CUDA, cupy | Large datasets |
+| **Workspace Analysis** | 10x+ faster | CUDA, cupy | Monte Carlo sampling |
+
+### Vision Features (Requires System Dependencies)
+
+| Feature | Requirements | Common Issues | Solutions |
+|---------|-------------|---------------|-----------|
+| **Camera Capture** | OpenCV, libGL.so.1 | ImportError: libGL.so.1 | `apt install libgl1-mesa-glx` |
+| **Object Detection** | ultralytics, internet | YOLO download fails | Check network, manual download |
+| **Stereo Vision** | OpenCV, calibration | Poor depth quality | Camera calibration required |
+| **3D Point Clouds** | OpenCV, numpy | Memory issues | Reduce point cloud density |
+
+### Installation Check Commands
+
+```python
+import ManipulaPy
+
+# Quick check - shows ✅/❌ for each feature
+ManipulaPy.check_dependencies()
+
+# Detailed system information
+ManipulaPy.print_system_info()
+
+# Get missing dependencies install commands
+print(ManipulaPy.get_installation_command())
+
+# Check specific feature
+try:
+    ManipulaPy.require_feature('cuda')
+    print("GPU acceleration ready!")
+except ImportError as e:
+    print(f"GPU not available: {e}")
+```
+
+---
 
 ## <a id="quick-start"></a>🚀 Quick Start
 
-### Prerequisites
-
+### Installation
 Before installing ManipulaPy, make sure your system has:
 
 1. **NVIDIA Drivers & CUDA Toolkit**  
@@ -80,49 +133,121 @@ Before installing ManipulaPy, make sure your system has:
    - Verify headers/libs under `/usr/include` and `/usr/lib/x86_64-linux-gnu` (or your distro’s equivalent).
 
 ---
-
-<h2 id="installation">Installation</h2>
+ManipulaPy attempts to install all dependencies by default for the best user experience. Missing dependencies are handled gracefully - the package will still work with available features.
 
 ```bash
-# Basic installation (CPU-only)
+# One command installs everything (recommended)
 pip install ManipulaPy
-
-# With GPU support (CUDA 11.x)
-pip install ManipulaPy[gpu-cuda11]
-
-# With GPU support (CUDA 12.x)
-pip install ManipulaPy[gpu-cuda12]
-
-# Development installation (with dev extras)
[…]
