+++ pyproject.toml
@@ -47,56 +47,55 @@ keywords = [
 ]
 
 # ------------------------------------------------------------------
-# Core runtime dependencies
+# Install all dependencies by default - let individual imports handle missing ones gracefully
 dependencies = [
+    # Core numerical computing (required)
+    "numpy>=1.19.2",
+    "scipy>=1.5.2", 
+    "matplotlib>=3.3",
+    
+    # Core robotics (required)
+    "numba>=0.55",
+    "pybullet>=3.0.6",
+    "urchin>=0.0.27",
+    "trimesh>=3.9.14",
+    
+    # Machine learning (install but handle gracefully if fails)
+    "torch>=1.8.0",
+    "scikit-learn>=1.0",
+    
+    # Vision and perception (install but handle gracefully if fails)
+    "opencv-python>=4.5",
+    "ultralytics>=8.0",
+    "pillow>=8.0.0",
+    
+    # GPU acceleration (install but handle gracefully if fails)
+    # Note: This will attempt CUDA 11.x first, fallback handled in code
+    "cupy-cuda11x>=10.0.0; sys_platform != 'darwin'",  # No CUDA on macOS typically
+]
+
+[project.optional-dependencies]
+# Minimal installation for constrained environments
+minimal = [
     "numpy>=1.19.2",
     "scipy>=1.5.2", 
     "matplotlib>=3.3",
     "numba>=0.55",
     "pybullet>=3.0.6",
     "urchin>=0.0.27",
-    "trimesh>=3.9.14",
-    "torch>=1.8.0",
-    "scikit-learn>=1.0",
-    "opencv-python>=4.5",
-    "ultralytics>=8.0",
-    "pillow>=8.0.0",
 ]
 
-[project.optional-dependencies]
-# Core optional dependencies
-core-optional = [
-    "scikit-learn>=1.0",
-]
+# Alternative GPU options for different CUDA versions
+gpu-cuda12 = ["cupy-cuda12x>=12.0.0"]
+gpu-rocm = ["cupy-rocm-4-3>=10.0.0"]
+gpu-pycuda = ["pycuda>=2021.1"]
 
-
-
-# Vision dependencies  
-vision = [
-    "opencv-python>=4.5",
-    "ultralytics>=8.0",
-    "pillow>=8.0.0",
-]
-
-# Vision dependencies for headless environments (CI-friendly)
+# Vision for headless environments (CI-friendly)
 vision-headless = [
     "opencv-python-headless>=4.5",
     "ultralytics>=8.0",
     "pillow>=8.0.0",
 ]
-
-# AI/ML dependencies
-ml = [
-    "torch>=1.8.0",
-    "scikit-learn>=1.0",
-]
-
-# GPU support
-gpu-cuda11 = ["cupy-cuda11x>=10.0.0"]
-gpu-cuda12 = ["cupy-cuda12x>=12.0.0"]
-gpu-rocm = ["cupy-rocm-4-3>=10.0.0"]
-gpu-pycuda = ["pycuda>=2021.1"]
-gpu = ["cupy-cuda11x>=10.0.0"]
 
 # Development and testing dependencies
 dev = [
@@ -137,16 +136,10 @@ ci = [
     "tomli>=1.2.0;python_version<'3.11'",
 ]
 
-# Testing combinations
-test-core = ["ManipulaPy[dev]"]
-test-simulation = ["ManipulaPy[dev,simulation]"]
-test-vision = ["ManipulaPy[dev,vision-headless]"]
-test-gpu = ["ManipulaPy[dev,gpu-cuda11]"]
-test-full = ["ManipulaPy[dev,simulation,vision-headless,ml]"]
-test-ci = ["ManipulaPy[ci]"]
-
-# All optional dependencies
[…]
