### 2025-08-17 1e8b9ed
Issue #22 and Issue #17
addressing the issues #22 and #17
