+++ ManipulaPy/__init__.py
@@ -4,282 +4,512 @@
 """
 ManipulaPy Package
 
-This package provides tools for the analysis and manipulation of robotic systems,
-including kinematics, dynamics, singularity analysis, path planning, and URDF processing utilities.
+This package provides tools for the analysis and manipulation of robotic systems, including kinematics,
+dynamics, singularity analysis, path planning, and URDF processing utilities.
 
 License: GNU Affero General Public License v3.0 or later (AGPL-3.0-or-later)
 Copyright (c) 2025 Mohamed Aboelnar
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU Affero General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-GNU Affero General Public License for more details.
-
-You should have received a copy of the GNU Affero General Public License
-along with this program. If not, see <https://www.gnu.org/licenses/>.
 """
 
 import warnings
 import sys
 import os
-import pkg_resources
+import platform
+from typing import Dict, List, Optional
 
 # Package metadata
-__version__ = "1.1.2"  # Updated version
-__author__ = "Mohamed Aboelnar"
+__version__ = "1.1.3"            
+__author__  = "Mohamed Aboelnar"
 __license__ = "AGPL-3.0-or-later"
 
-# Core modules that should always be available
-_CORE_MODULES = []
-_OPTIONAL_MODULES = []
+# Dependency availability tracking
+_available_features = {
+    'core': True,          # Always available (numpy, scipy, etc.)
+    'cuda': False,         # GPU acceleration
+    'vision': False,       # Computer vision features
+    'simulation': False,   # PyBullet simulation
+    'ml': False,          # Machine learning features
+}
 
-# Check if ManipulaPy_data is available and accessible
-def _check_data_availability():
-    """Check if ManipulaPy_data is available and accessible."""
-    try:
-        # Method 1: Try to access as package resource
-        try:
-            pkg_resources.resource_exists(__name__, 'ManipulaPy_data')
-            return True
-        except:
-            pass
-        
-        # Method 2: Try direct path access
-        try:
-            package_dir = os.path.dirname(__file__)
-            data_dir = os.path.join(package_dir, 'ManipulaPy_data')
-            return os.path.exists(data_dir) and os.path.isdir(data_dir)
-        except:
-            pass
-        
-        return False
-    except:
-        return False
+# Track missing dependencies for helpful error messages
+_missing_dependencies = {}
 
-# Set global flag for data availability
-_DATA_AVAILABLE = _check_data_availability()
-
-if not _DATA_AVAILABLE:
-    warnings.warn(
-        "ManipulaPy_data directory not found. Some examples may not work. "
-        "This can happen with certain pip installations. "
-        "Try installing from source: pip install git+https://github.com/boelnasr/ManipulaPy.git",
-        ImportWarning,
-        stacklevel=2
-    )
-
-# Import core utilities (should always work)
-try:
-    from ManipulaPy.utils import *
-    from ManipulaPy import utils
-    _CORE_MODULES.append("utils")
-except ImportError as e:
-    warnings.warn(f"Failed to import utils module: {e}", ImportWarning)
-
-# Import core functionality that depends on torch
-try:
-    from ManipulaPy.kinematics import *
-    from ManipulaPy import kinematics
-    _CORE_MODULES.append("kinematics")
-except ImportError as e:
-    warnings.warn(f"Failed to import kinematics module: {e}. Ensure PyTorch is installed.", ImportWarning)
-
-try:
[…]
