### 2025-08-17 boelnasr/ManipulaPy#22 @boelnasr
Thank you @cadojo for the detailed feedback! I've implemented comprehensive changes to resolve all these concerns. Fixed: Simplified Installation & Graceful Dependency Handling
Manipulate now uses a single installation command that attempts to install all dependencies while gracefully handling missing ones. now can check which modules can be used after the [installation](https://github.com/boelnasr/ManipulaPy?tab=readme-ov-file#check-your-installation)
