### 2025-08-17 fe60a57
Paper Updates
addressing issue #17
