+++ paper/paper.md
@@ -146,8 +146,100 @@ $$\tau = M(\theta)\ddot{\theta} + C(\theta,\dot{\theta}) + G(\theta)$$
 
 The mass matrix $M(\theta) = \sum_{i=1}^{n}\operatorname{Ad}_{T_i}^T G_i \operatorname{Ad}_{T_i}$ computation is optimized for 256-thread blocks, achieving up to **3600× speedup for inverse dynamics** and **8× speedup for trajectory generation** on 6-DOF manipulators compared to NumPy implementations.
 
+# CPU vs GPU Module Requirements
+
+ManipulaPy provides tiered functionality that gracefully scales from CPU-only to GPU-accelerated operation:
+
+## CPU-Only Features (Always Available)
+
+**Core Robotics Modules:**
+- URDF Processing: Model loading and screw axis extraction [@lynch2017modern]
+- Forward/Inverse Kinematics: Single-point pose calculations using Product-of-Exponentials formulation [@lynch2017modern]  
+- Jacobian Analysis: Velocity relationships and singularity detection
+- Small Trajectory Planning: N < 1000 points using CPU fallback with Numba optimization
+- Basic Control: PID, computed torque without real-time constraints
+- Simulation Setup: PyBullet loading and basic joint control [@coumans2019]
+
+**Performance Characteristics:**
+- Single trajectory generation: ~10-50ms for 6-DOF robots
+- Real-time control limited to ~100 Hz due to Python Global Interpreter Lock (GIL)
+
+## GPU-Required Features (CUDA Hardware)
+
+**High-Performance Modules:**
+- Large Trajectory Planning: N > 1000 points with 40× speedup over CPU implementations [@liang2018gpu; @sundaralingam2023curobo]
+- Batch Processing: Multiple trajectories simultaneously using CuPy acceleration [@cupy2021]
+- Real-time Inverse Dynamics: >1 kHz computation rates enabled by custom CUDA kernels
+- Workspace Analysis: Monte Carlo sampling with GPU parallelization [@liang2018gpu]
+- GPU-Accelerated Potential Fields: Real-time collision avoidance using parallel gradient computation
+
+**Performance Characteristics:**
+- Large trajectory generation: ~1-5ms for 6-DOF robots with optimized CUDA kernels
+- Batch inverse dynamics: 3600× speedup for multiple robot configurations
+- Real-time control: 1 kHz rates achievable with GPU acceleration
+
+## Vision Features (Additional Dependencies)
+
+**System Requirements:**
+- OpenCV: Camera calibration, stereo rectification, image processing
+- Graphics Libraries: libGL.so.1 for visualization
+- YOLO Models: Ultralytics for object detection [@Jocher_Ultralytics_YOLO_2023]
+
+**Modules:**
+- Camera Operations: Image capture, stereo processing, calibration
+- Object Detection: YOLO integration, 3D point generation [@Jocher_Ultralytics_YOLO_2023]
+- Clustering: DBSCAN spatial grouping for obstacle representation [@chu2021boundary]
+
+# Limitations and Design Trade-offs
+
+## Performance and Scalability Limitations
+
+**GPU Memory Constraints**: Consumer GPUs (8GB) limit trajectory planning to ~50,000 points; professional GPUs (24GB+) required for larger problems. Memory usage scales as O(N × joints × 4 bytes) as demonstrated in CuRobo's parallel collision-free motion generation [@sundaralingam2023curobo].
+
+**Small Problem Overhead**: GPU acceleration only beneficial for N > 1000 trajectory points due to kernel launch overhead [@liang2018gpu]. CPU implementation remains faster for small-scale problems, consistent with findings in GPU-based robot dynamics computation.
+
[…]
