### 2025-08-17 70145ab
dynamics documentation update
