+++ docs/source/user_guide/Dynamics.rst
@@ -351,4 +351,2025 @@ Separate centrifugal (velocity²) and Coriolis (cross-coupling) effects:
        centrifugal = np.zeros(n)
        coriolis = np.zeros(n)
        
-       # Centrifugal forces (diagonal
+       # Centrifugal forces (diagonal terms: i=j=k)
+       for i in range(n):
+           for j in range(n):
+               if i == j:
+                   christoffel = dynamics.partial_derivative(i, j, j, theta)
+                   centrifugal[i] += christoffel * theta_dot[j] * theta_dot[j]
+       
+       # Coriolis forces (off-diagonal coupling: i≠j or j≠k)
+       for i in range(n):
+           for j in range(n):
+               for k in range(n):
+                   if not (i == j == k):
+                       christoffel = dynamics.partial_derivative(i, j, k, theta)
+                       coriolis[i] += christoffel * theta_dot[j] * theta_dot[k]
+       
+       return centrifugal, coriolis
+
+   # Example usage
+   theta = np.array([0.1, 0.3, -0.2])
+   theta_dot = np.array([1.0, -0.5, 0.8])
+
+   centrifugal, coriolis = decompose_velocity_forces(dynamics, theta, theta_dot)
+   total_c = dynamics.velocity_quadratic_forces(theta, theta_dot)
+
+   print(f"Centrifugal forces: {centrifugal}")
+   print(f"Coriolis forces: {coriolis}")
+   print(f"Total forces: {total_c}")
+   print(f"Sum check: {np.allclose(centrifugal + coriolis, total_c)}")
+
+Gravity Compensation
+----------------------
+
+Gravity forces act continuously on robot links and must be compensated for precise control.
+
+Computing Gravity Forces
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. code-block:: python
+
+   # Standard Earth gravity
+   g_earth = [0, 0, -9.81]  # m/s²
+   
+   # Different orientations
+   g_upright = [0, 0, -9.81]      # Robot upright
+   g_inverted = [0, 0, 9.81]      # Robot inverted
+   g_sideways = [9.81, 0, 0]      # Robot on its side
+
+   # Compute gravity forces for different configurations
+   configurations = [
+       np.array([0, 0, 0]),           # Home position
+       np.array([np.pi/2, 0, 0]),     # Joint 1 at 90°
+       np.array([0, np.pi/2, 0]),     # Joint 2 at 90°
+       np.array([0, 0, np.pi/2]),     # Joint 3 at 90°
+   ]
+
+   for i, theta in enumerate(configurations):
+       g_torques = dynamics.gravity_forces(theta, g_earth)
+       print(f"Config {i+1}: θ={theta}, G(θ)={g_torques}")
+
+Gravity in Different Orientations
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. code-block:: python
+
+   def analyze_gravity_effects(dynamics, gravity_vectors):
+       """Analyze gravity effects for different robot orientations."""
+       
+       theta = np.array([np.pi/4, np.pi/6, -np.pi/3])  # Test configuration
+       
+       fig, axes = plt.subplots(1, len(gravity_vectors), figsize=(15, 4))
+       if len(gravity_vectors) == 1:
+           axes = [axes]
+       
+       for idx, (g_vec, label) in enumerate(gravity_vectors):
+           g_forces = dynamics.gravity_forces(theta, g_vec)
+           
+           axes[idx].bar(range(len(g_forces)), g_forces)
+           axes[idx].set_title(f'Gravity: {label}')
+           axes[idx].set_xlabel('Joint')
+           axes[idx].set_ylabel('Torque (N⋅m)')
+           axes[idx].grid(True, alpha=0.3)
+       
[…]
