### 2025-08-17 1d63eff
dynamics documentation update
