### 2025-08-17 a99065f
docs typo fixes
