### 2025-08-24 2c60b67
Issue #21
adding newline between the list's header (bold text) and the list items
