+++ paper/paper.md
@@ -1,12 +1,12 @@
 ---
-title: "ManipulaPy: A GPU‑Accelerated Python Framework for Robotic Manipulation, Perception, and Control"
+title: "ManipulaPy: A GPU-Accelerated Python Framework for Robotic Manipulation, Perception, and Control"
 tags: [robotics, manipulator, simulation, kinematics, dynamics, perception, cuda, trajectory-planning, computer-vision]
 authors:
   - name: M.I.M. AboElNasr
     orcid: 0000-0002-1768-2031
     affiliation: 1
 affiliations:
-  - name: Universität Duisburg‑Essen
+  - name: Universität Duisburg-Essen
     index: 1
 date: "2025-05-03"
 bibliography: paper.bib
@@ -14,11 +14,11 @@ bibliography: paper.bib
 
 # Summary
 
-**ManipulaPy** is an open‑source Python toolbox that unifies the entire manipulation pipeline—from URDF parsing to GPU‑accelerated dynamics, vision‑based perception, planning and control—within a single API. Built on the Product‑of‑Exponentials model [@lynch2017modern] (similar to Pinocchio [@Pinocchio2025] but with GPU acceleration), PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library enables researchers to move from robot description to real‑time control with up to **13× overall performance improvement** and **3600× faster inverse dynamics** on a 6‑DOF UR5 compared to NumPy baseline. Performance claims are reproducible via benchmarks described in `Benchmark/README.md`.
+**ManipulaPy** is an open-source Python toolbox that unifies the entire manipulation pipeline—from URDF parsing to GPU-accelerated dynamics, vision-based perception, planning and control—within a single API. Built on the Product-of-Exponentials model [@lynch2017modern] (similar to Pinocchio [@Pinocchio2025] but with GPU acceleration), PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library enables researchers to move from robot description to real-time control with up to **13× overall performance improvement** and **3600× faster inverse dynamics** on a 6-DOF UR5 compared to NumPy baseline. Performance claims are reproducible via benchmarks in `benchmarks/README.md`.
 
 # Statement of Need
 
-Modern manipulation research requires tight integration of geometry, dynamics, perception, planning, and control—ideally within a single, real-time computational loop with GPU acceleration. However, existing open-source tools address only portions of this pipeline, forcing researchers to write substantial integration code:
+Modern manipulation research requires tight integration of geometry, dynamics, perception, planning, and control—ideally within a single, real-time computational loop on GPU hardware. However, existing open-source tools address only portions of this pipeline, forcing researchers to write substantial integration code:
 
 | Library | Core Strengths | Integration Challenges |
 |---------|---------------|------------------------|
@@ -31,16 +31,16 @@ Modern manipulation research requires tight integration of geometry, dynamics, p
 
 These integration challenges manifest as:
 
-- **Sensor-planner gaps**: Converting camera data to collision geometries requires custom OpenCV → ROS → MoveIt pipelines.
-- **Dynamics-control mismatches**: Real-time controllers need consistent mass matrices, but most libraries compute dynamics separately from control loops.
-- **GPU memory fragmentation**: Transferring data between CPU planners and GPU dynamics creates performance bottlenecks.
-- **Synchronization complexity**: Keeping sensors, planners, and controllers temporally aligned requires careful threading and message passing.
[…]
