+++ docs/source/api/potential_field.rst
@@ -226,9 +226,13 @@ Collision Detection
    **Detection Algorithm:**
    
    1. **Forward Kinematics**: Compute link transforms using robot.link_fk(cfg=thetalist)
+
    2. **Hull Transformation**: Transform each link's convex hull to world coordinates
+
    3. **Pairwise Intersection**: Check all link pairs for hull intersections
+
    4. **Self-Exclusion**: Skip collision checks between identical links
+
    5. **Boolean Result**: Return True on first intersection found
 
 ---
@@ -255,8 +259,11 @@ ConvexHull Properties
 Each ConvexHull object contains:
 
 - **points**: Original vertex coordinates
+
 - **vertices**: Indices of vertices forming the hull
+
 - **simplices**: Triangular faces of the hull
+
 - **equations**: Hyperplane equations for faces
 
 Forward Kinematics Integration
@@ -265,7 +272,9 @@ Forward Kinematics Integration
 Integration with urchin URDF processing:
 
 - **Input**: Joint configuration (thetalist)
+
 - **Method**: robot.link_fk(cfg=thetalist)
+
 - **Output**: Dictionary {link_name: 4×4_transform_matrix}
 
 ---
@@ -331,7 +340,9 @@ Potential Field Scaling
 Applied scaling factors in implementation:
 
 - **Repulsive potential**: 10× multiplier on final result
+
 - **Repulsive gradient**: 5× multiplier on gradient magnitude
+
 - **Distance threshold**: influence_distance parameter cutoff
 
 Distance Computations
@@ -340,7 +351,9 @@ Distance Computations
 All distance calculations use numpy.linalg.norm():
 
 - **Method**: Euclidean L2 norm
+
 - **Input**: Configuration space vectors
+
 - **Precision**: Double precision floating point
 
 Matrix Operations
@@ -349,7 +362,9 @@ Matrix Operations
 Convex hull transformations use optimized NumPy operations:
 
 - **Rotation**: 3×3 matrix multiplication with broadcasted points
+
 - **Translation**: Vector addition with reshaped translation vector
+
 - **Memory layout**: Contiguous arrays for efficient computation
 
 Intersection Detection
@@ -358,7 +373,9 @@ Intersection Detection
 Hull intersection testing relies on scipy.spatial.ConvexHull:
 
 - **Algorithm**: Built-in intersection testing methods
+
 - **Precision**: Computational geometry tolerance handling
+
 - **Robustness**: Handles degenerate cases automatically
 
 ---
@@ -370,21 +387,27 @@ URDF Processing Errors
 ----------------------
 
 - **Missing files**: URDF.load() exception handling
+
 - **Invalid geometry**: hasattr() validation for mesh.vertices
+
 - **Empty meshes**: Automatic skipping of links without valid geometry
 
 Convex Hull Failures
 --------------------
 
 - **Coplanar points**: scipy.spatial.ConvexHull internal handling
+
 - **Insufficient vertices**: Minimum 4 points required for 3D hulls
+
 - **Numerical precision**: Tolerance-based geometric computations
 
 Potential Field Singularities
 -----------------------------
 
 - **Zero distance**: Protected by influence_distance > 0 requirement
+
 - **Infinite gradients**: Prevented by distance threshold checking
+
 - **Numerical overflow**: Managed through finite gain parameters
 
 ---
@@ -398,7 +421,9 @@ Convex Hull Storage
 Memory allocation for collision checking:
 
 - **Hull objects**: Persistent storage throughout object lifetime
+
 - **Transformed hulls**: Temporary allocation during collision checks
+
 - **Point arrays**: Copy operations for transformation safety
 
 Potential Field Computations
[…]
