### 2025-08-24 9350257
Issue #21
adding newline between the list's header (bold text) and the list items
