+++ docs/source/user_guide/Trajectory_Planning.rst
@@ -946,203 +946,277 @@ Processing multiple trajectories efficiently:
 Real-Time Applications
 -------------------------
 
-Trajectory Execution
-~~~~~~~~~~~~~~~~~~~~~~
+High-Frequency Control Simulation
+---------------------------------
 
-Real-time trajectory following for robot control:
+.. important:: Real-time semantics
+
+   In ManipulaPy, *real-time* refers to **soft real-time (best-effort)** loops:
+   a fixed-rate cycle (e.g., 100–1000 Hz) driven by a monotonic clock where the
+   *average* period is met and small jitter/missed deadlines are acceptable.
+   Python does **not** provide hard real-time guarantees. For **hard real-time**
+   (deterministic deadlines), implement the inner loop on RT hardware/OS and
+   use ManipulaPy for trajectory generation, validation, and supervision.
+
+Control Loop Simulation at Industrial Rates
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+This section demonstrates trajectory-following **simulation** at typical industrial control frequencies (100–1000 Hz). While Python cannot guarantee hard real-time, these examples show how to design trajectories suitable for deployment on dedicated RT controllers.
+
+Control Frequency Considerations
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+**Soft Real-Time Simulation (Python):**
+
+- Simulated control loops at 100–1000 Hz.
+- Suitable for algorithm validation and controller tuning.
+- No timing guarantees — for development only.
+
+**Hard Real-Time Implementation (Target Hardware):**
+
+- Requires a real-time OS (e.g., PREEMPT_RT Linux, Xenomai, VxWorks) or MCU/PLC.
+- Deterministic timing guarantees.
+- Use ManipulaPy for offline trajectory design and export to the RT stack.
+
+Trajectory Execution Simulation (100 Hz)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The example below simulates a **100 Hz** control loop for trajectory tracking. It also includes optional soft RT scheduling to align each iteration with the 10 ms period and report deadline misses.
 
 .. code-block:: python
 
-   def real_time_trajectory_execution():
-       """Simulate real-time trajectory execution."""
-       
-       # Generate reference trajectory
+   import time
+   import numpy as np
+   import matplotlib.pyplot as plt
+
+   def simulate_high_frequency_control(enforce_timing: bool = True):
+       """
+       Simulate high-frequency trajectory following control (100 Hz).
+
+       Notes:
+       - This is a simulation, not hard real-time.
+       - Assumes 'planner' is already constructed (robot, dynamics, limits).
+       """
+
+       # Generate reference trajectory offline
        theta_start = np.array([0.1, 0.2, -0.1, 0.0, 0.3, 0.0])
-       theta_end = np.array([0.8, -0.3, 0.5, -0.2, 0.6, -0.4])
-       
+       theta_end   = np.array([0.8, -0.3,  0.5, -0.2, 0.6, -0.4])
+
        ref_trajectory = planner.joint_trajectory(
-           theta_start, theta_end, Tf=4.0, N=400, method=5  # 100 Hz
+           theta_start, theta_end, Tf=4.0, N=400, method=5  # 100 Hz (dt = 0.01 s)
        )
-       
-       # Simulation parameters
-       dt = 0.01  # 100 Hz control rate
-       n_steps = ref_trajectory['positions'].shape[0]
-       
-       # Control parameters
+
+       # Control loop target
+       dt = 0.01  # 100 Hz
[…]
