### 2025-08-24 2d0d717
Docs format update
updating syntax errors in the docs
