+++ docs/source/user_guide/Collision_Checker.rst
@@ -1,61 +1,67 @@
+.. _user_guide_collision_checker:
+
 Collision Checker Module User Guide
 ===================================
 
-This guide covers the `CollisionChecker` class in ManipulaPy, which performs
-geometry-based self-collision detection for serial manipulators using URDF meshes.
+This guide covers the :class:`ManipulaPy.potential_field.CollisionChecker` class,
+which performs geometry-based **self-collision detection** for serial manipulators
+using URDF meshes.
+
+.. contents:: **Quick navigation**
+   :local:
+   :depth: 2
 
 .. note::
-   This guide uses Python 3.10.12, SciPy 1.10+, and the `urchin.urdf` parser
-   to load URDF meshes and compute convex hulls for collision tests.
+   Examples target **Python 3.10+**, **SciPy 1.10+**, and the ``urchin.urdf`` parser
+   for loading URDF meshes and building convex hulls for collision tests.
 
 Introduction
 ------------
 
-Collision checking is essential for safe motion planning.  The
-`CollisionChecker` builds convex‐hull approximations of each link’s visual mesh
+Collision checking is essential for safe motion planning. The
+``CollisionChecker`` builds **convex-hull** approximations of each link’s visual mesh
 and tests for pairwise intersection at a given robot configuration.
 
-Key features:
+Key features
+~~~~~~~~~~~~
 
-- **Convex hull construction** from URDF mesh vertices  
-
-- **Fast pairwise collision tests** using SciPy’s `ConvexHull.intersects`  
-
-- **Transforms hulls** by forward‐kinematics link poses  
-
-- **Self-collision detection** between any two non-adjacent links  
+- **Convex hull construction** from URDF mesh vertices
+- **Fast pairwise collision tests** using convex polytope intersection checks
+- **Pose transforms** of hulls via forward kinematics
+- **Self-collision detection** between any two **non-adjacent** links
 
 Mathematical Background
 -----------------------
 
-Given a robot configuration :math:`\mathbf q` (joint angles), the forward‐kinematics
+Given a robot configuration :math:`\mathbf q` (joint angles), the forward-kinematics
 function :math:`T_i(\mathbf q)\in SE(3)` returns the homogeneous transform of link *i*.
-Each link’s mesh is approximated by its convex hull :math:`H_i`.  Under transform
-:math:`T_i`, the hull vertices move to
+Each link’s mesh is approximated by its convex hull :math:`H_i`. Under transform
+:math:`T_i`, hull vertices map to
 
 .. math::
-   \{\,T_i(\mathbf q)\,p \mid p\in H_i.\}
+   \{\,T_i(\mathbf q)\,p \mid p\in H_i\,\}.
 
-Two convex polyhedra intersect iff their convex hulls intersect.  SciPy’s
-`ConvexHull` can test intersection via oriented‐bounding‐boxes or
-half‐space tests under the hood.
+Two convex polyhedra intersect iff their convex hulls intersect. In practice, the
+check can be implemented by half-space evaluations or separating-axis tests on the
+transformed hulls.
 
-Workflow:
+Workflow
+~~~~~~~~
 
 1. **Load URDF** → extract mesh vertices.  
-
 2. **Build hulls** :math:`\{H_i\}` offline.  
+3. For each configuration :math:`\mathbf q`:
+   - Compute :math:`T_i(\mathbf q)` for each link.
+   - Transform :math:`H_i \rightarrow T_i(H_i)`.
[…]
