+++ docs/source/index.rst
@@ -24,14 +24,14 @@ A modern, GPU-accelerated Python toolbox for **robot kinematics, dynamics, traje
             <a href="https://joss.theoj.org/papers/e0e68c2dcd8ac9dfc1354c7ee37eb7aa">
                <img src="https://joss.theoj.org/papers/e0e68c2dcd8ac9dfc1354c7ee37eb7aa/status.svg?style=for-the-badge" alt="JOSS Paper">
             </a>
-            <a href="https://github.com/boelnasr/ManipulaPy/blob/main/LICENSE">
+            <a href="https://github.com/boelnar/ManipulaPy/blob/main/LICENSE">
                <img src="https://img.shields.io/badge/License-AGPL%20v3-blue.svg?style=for-the-badge" alt="License">
             </a>
-            <a href="https://github.com/boelnasr/ManipulaPy/actions">
+            <a href="https://github.com/boelnar/ManipulaPy/actions">
                <img src="https://img.shields.io/github/actions/workflow/status/boelnasr/ManipulaPy/test.yml?branch=main&style=for-the-badge&logo=github&label=CI" alt="CI Status">
             </a>
             <a href="https://pypi.org/project/manipulapy/">
-               <img src="https://static.pepy.tech/badge/manipulapy" alt="Downloads">
+               <img src="https://static.pepy.tech/badge/manipulapy">
             </a>
             <a href="https://github.com/boelnasr/ManipulaPy">
                <img src="https://img.shields.io/github/stars/boelnasr/ManipulaPy?style=for-the-badge&logo=github&logoColor=white&label=Stars&color=yellow" alt="GitHub Stars">
@@ -54,6 +54,7 @@ A modern, GPU-accelerated Python toolbox for **robot kinematics, dynamics, traje
                <strong>Real-time Control</strong><br>
                PyBullet simulation and advanced controllers
             </div>
+
          </div>
       </div>
    </div>
@@ -61,7 +62,6 @@ A modern, GPU-accelerated Python toolbox for **robot kinematics, dynamics, traje
 .. contents:: **Quick links**
    :local:
    :depth: 1
-
 
 Getting Started
 ---------------
@@ -90,9 +90,8 @@ If you're in a hurry, install the package into a fresh virtual-env and try the e
    python -m pip install manipulapy[cuda]  # or just manipulapy
    python -c "import ManipulaPy; print('🎉 Installation successful!')"
 
-.. note::
-   The docs you're reading are generated from the source in ``docs/``; feel free to improve them and send a `pull request <https://github.com/boelnasr/ManipulaPy/pull/new>`_. 🚀
-
+.. note:: 
+   The docs you're reading are generated from the source in `docs/`; feel free to improve them and send a [pull request](https://github.com/boelnasr/ManipulaPy/pull/new). 🚀
 
 🚀 Quick Start Examples
 ~~~~~~~~~~~~~~~~~~~~~~~
@@ -297,7 +296,6 @@ If you're in a hurry, install the package into a fresh virtual-env and try the e
       </div>
    </div>
 
-
 Key Features at a Glance
 ~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -333,7 +331,6 @@ Key Features at a Glance
       </div>
    </div>
 
-
 Documentation map
 -----------------
 
@@ -347,12 +344,13 @@ Documentation map
    :maxdepth: 2
    :caption: 🛠️ API Reference
 
-   api/index
-
+   api/index   
 .. toctree::
    :maxdepth: 2
    :caption: 📚 User Guides
 
+   
+   
    user_guide/index
[…]

[…]
