### 2025-08-24 6d924c9
Docs format update
updating syntax errors in the docs
