+++ docs/makefile
@@ -7,7 +7,6 @@
 # =============================================================================
 
 # Sphinx configuration
-SPHINXOPTS    ?= -W --keep-going
 SPHINXBUILD   ?= sphinx-build
 SPHINXAPIDOC  ?= sphinx-apidoc
 SOURCEDIR     = source
@@ -17,6 +16,27 @@ PACKAGE_DIR   = ../ManipulaPy
 # Python configuration
 PYTHON        ?= python3
 PIP           ?= pip3
+
+# Headless / CI-friendly env for all sphinx calls
+SPHINXENV     = MPLBACKEND=Agg SPHINX_BUILD=1
+
+# Build strictness
+#   0 = default (keep going, do not fail on warnings)
+#   1 = treat warnings as errors (-W)
+SPHINXSTRICT  ?= 0
+
+# Baseline Sphinx options: parallel (-j auto), show full tracebacks (-T)
+SPHINXOPTS_COMMON = -j auto -T
+
+# Strict toggle
+ifeq ($(SPHINXSTRICT),1)
+	SPHINXOPTS := $(SPHINXOPTS_COMMON) -W
+else
+	SPHINXOPTS := $(SPHINXOPTS_COMMON) --keep-going
+endif
+
+# Optional fixer script that inserts blank lines after headers (safe if missing)
+DOCS_FIXER ?= fix_docs.py
 
 # Colors for output
 RED    = \033[0;31m
@@ -35,38 +55,40 @@ help: ## Show this help message
 	@echo "$(BLUE)====================================$(NC)"
 	@echo ""
 	@echo "$(GREEN)Available targets:$(NC)"
-	@echo "  $(YELLOW)html$(NC)          Build HTML documentation"
-	@echo "  $(YELLOW)dirhtml$(NC)       Build HTML pages as directories"
-	@echo "  $(YELLOW)singlehtml$(NC)    Build single HTML page"
-	@echo "  $(YELLOW)pickle$(NC)        Build pickle documentation"
-	@echo "  $(YELLOW)json$(NC)          Build JSON documentation"
-	@echo "  $(YELLOW)htmlhelp$(NC)      Build HTML help documentation"
-	@echo "  $(YELLOW)qthelp$(NC)        Build Qt help documentation"
-	@echo "  $(YELLOW)devhelp$(NC)       Build DevHelp documentation"
-	@echo "  $(YELLOW)epub$(NC)          Build EPUB documentation"
-	@echo "  $(YELLOW)latex$(NC)         Build LaTeX documentation"
-	@echo "  $(YELLOW)latexpdf$(NC)      Build PDF documentation via LaTeX"
-	@echo "  $(YELLOW)latexpdfja$(NC)    Build PDF documentation via LaTeX (Japanese)"
-	@echo "  $(YELLOW)text$(NC)          Build plain text documentation"
-	@echo "  $(YELLOW)man$(NC)           Build manual pages"
-	@echo "  $(YELLOW)texinfo$(NC)       Build Texinfo documentation"
-	@echo "  $(YELLOW)info$(NC)          Build info documentation"
-	@echo "  $(YELLOW)gettext$(NC)       Build gettext documentation"
-	@echo "  $(YELLOW)changes$(NC)       Build changes documentation"
-	@echo "  $(YELLOW)xml$(NC)           Build XML documentation"
-	@echo "  $(YELLOW)pseudoxml$(NC)     Build pseudo-XML documentation"
[…]
