### 2025-08-24 dd4d5ed
dynamics guide syntax
fixing title underline
