+++ docs/source/user_guide/Dynamics.rst
@@ -354,7 +354,6 @@ Separate centrifugal (velocity²) and Coriolis (cross-coupling) effects:
        
        n = len(theta)
        centrifugal = np.zeros(n)
-<<<<<<< HEAD
        coriolis = np.zeros(n)
        
        # Centrifugal forces (diagonal terms: i=j=k)
@@ -1033,813 +1032,6 @@ Mass Matrix Caching
    print(f"Speedup: {uncached_time/cached_time:.2f}x")
    print(f"Cache stats: {optimized_dynamics.cache_stats()}")
 
-Parallel Computation
-~~~~~~~~~~~~~~~~~~~~~~~
-
-.. code-block:: python
-
-   def parallel_trajectory_dynamics(dynamics, trajectory_data, num_processes=4):
-       """Compute dynamics for trajectory using parallel processing."""
-       
-       import multiprocessing as mp
-       from functools import partial
-       
-       def compute_point_dynamics(point_data):
-           """Compute dynamics for a single trajectory point."""
-           i, theta, theta_dot, theta_ddot = point_data
-           
-           try:
-               # Compute all dynamic quantities
-               M = dynamics.mass_matrix(theta)
-               c = dynamics.velocity_quadratic_forces(theta, theta_dot)
-               g_forces = dynamics.gravity_forces(theta, [0, 0, -9.81])
-               
-               # Inverse dynamics
-               Ftip = [0, 0, 0, 0, 0, 0]
-               tau = dynamics.inverse_dynamics(theta, theta_dot, theta_ddot, [0, 0, -9.81], Ftip)
-               
-               return i, {
-                   'mass_matrix': M,
-                   'coriolis_forces': c,
-                   'gravity_forces': g_forces,
-                   'required_torques': tau
-               }
-           except Exception as e:
-               return i, {'error': str(e)}
-       
-       # Prepare data for parallel processing
-       positions = trajectory_data['positions']
-       velocities = trajectory_data['velocities']
-       accelerations = trajectory_data['accelerations']
-       
-       point_data = [
-           (i, positions[i], velocities[i], accelerations[i])
-           for i in range(len(positions))
-       ]
-       
-       # Process in parallel
-       with mp.Pool(processes=num_processes) as pool:
-           results = pool.map(compute_point_dynamics, point_data)
-       
-       # Reconstruct ordered results
-       dynamics_data = {}
-       for i, result in sorted(results):
-           dynamics_data[i] = result
-       
-       return dynamics_data
-
-   # Example usage (when trajectory is available)
-   if 'trajectory' in locals():
-       parallel_results = parallel_trajectory_dynamics(dynamics, trajectory)
-       print(f"Computed dynamics for {len(parallel_results)} trajectory points")
-
-GPU Acceleration Integration
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. code-block:: python
-
-   def integrate_gpu_dynamics(dynamics, use_cuda=True):
-       """Integrate with GPU-accelerated trajectory planning."""
-       
-       from ManipulaPy.path_planning import OptimizedTrajectoryPlanning
-       
-       # Create GPU-optimized planner
-       joint_limits = [(-np.pi, np.pi)] * len(dynamics.joint_limits)
-       
-       planner = OptimizedTrajectoryPlanning(
-           robot, "robot.urdf", dynamics, joint_limits,
-           use_cuda=use_cuda,
-           cuda_threshold=100,  # Use GPU for trajectories > 100 points
-           enable_profiling=True,
-           auto_optimize=True,
-           target_speedup=40.0
-       )
-       
-       # Generate high-resolution trajectory
[…]
