### 2025-08-24 9b3eae5
dynamics guide syntax
fixing title underline
