### 2025-08-24 boelnasr/ManipulaPy#20 @boelnasr
I add a clarification for it: in the docs **real-time** means soft real-time (best-effort)—fixed-rate loops with possible jitter/missed deadlines. Python can’t do hard RT; for that, run the inner loop on RT hardware/OS and use ManipulaPy for offline planning. A definition callout is now added at the start of that section.
