### 2025-08-24 boelnasr/ManipulaPy#21 @boelnasr
Fixed in the docs: I added the required blank line after bold list headers before the items, Lists now render correctly (verified with make html), let me know if i missed anything.
