### 2025-08-24 boelnasr/ManipulaPy#23 @boelnasr
Yes—the trajectory kernels are DOF-agnostic, so 7-DOF (and higher) manipulators already get GPU speedups (joint & Cartesian time-scaling, plus batched trajectories). The 3600× figure is from a specialized 6-DOF inverse-dynamics prototype; the released GPU dynamics kernels are shape-agnostic but simplified (per-joint/diagonalized), while the fully coupled n-DOF dynamics remain on CPU for exactness.
