+++ paper/paper.bib
@@ -1,126 +1,126 @@
 @book{lynch2017modern,
-  title={Modern Robotics: Mechanics, Planning, and Control},
-  author={Lynch, Kevin M. and Park, Frank C.},
-  year={2017},
-  publisher={Cambridge University Press},
-  url={http://modernrobotics.org}
+  title     = {Modern Robotics: Mechanics, Planning, and Control},
+  author    = {Lynch, Kevin M. and Park, Frank C.},
+  year      = {2017},
+  publisher = {Cambridge University Press},
+  url       = {http://modernrobotics.org}
 }
 
 @article{corke2021,
-  title={Robotics, Vision and Control: Fundamental Algorithms in {MATLAB}, {Python} and {Julia}},
-  author={Corke, Peter and Haviland, Jesse},
-  journal={Springer Tracts in Advanced Robotics},
-  volume={1},
-  year={2021},
-  publisher={Springer},
-  url={https://petercorke.com/books/robotics-vision-control-python-the-practice-of-robotics-vision/}
+  title     = {Robotics, Vision and Control: Fundamental Algorithms in {MATLAB}, {Python} and {Julia}},
+  author    = {Corke, Peter and Haviland, Jesse},
+  journal   = {Springer Tracts in Advanced Robotics},
+  volume    = {1},
+  year      = {2021},
+  publisher = {Springer},
+  url       = {https://petercorke.com/books/robotics-vision-control-python-the-practice-of-robotics-vision/}
 }
 
 @article{chitta2012moveit,
-  title={{MoveIt!}: An Introduction},
-  author={Chitta, Sachin and Sucan, Ioan A. and Cousins, Steve},
-  journal={AI Matters},
-  volume={3},
-  number={4},
-  pages={4--9},
-  year={2012},
-  publisher={ACM},
-  url={https://moveit.ros.org/}
+  title     = {{MoveIt!}: An Introduction},
+  author    = {Chitta, Sachin and Sucan, Ioan A. and Cousins, Steve},
+  journal   = {AI Matters},
+  volume    = {3},
+  number    = {4},
+  pages     = {4--9},
+  year      = {2012},
+  publisher = {ACM},
+  url       = {https://moveit.ros.org/}
 }
 
 @misc{coumans2019,
-  title={{PyBullet}: Physics simulation for games, robotics, and machine learning},
-  author={Coumans, Erwin and Bai, Yunfei},
-  year={2019},
-  url={http://pybullet.org},
-  note={Open-source physics engine}
+  title  = {{PyBullet}: Physics simulation for games, robotics, and machine learning},
+  author = {Coumans, Erwin and Bai, Yunfei},
+  year   = {2019},
+  url    = {http://pybullet.org},
+  note   = {Open-source physics engine}
 }
 
 @misc{smits2009kdl,
-  title={The Kinematics and Dynamics Library ({KDL})},
-  author={Smits, Ruben and Soetens, Peter and Bruyninckx, Herman},
-  year={2009},
-  organization={OROCOS Project},
-  url={https://www.orocos.org/kdl.html}
+  title        = {The Kinematics and Dynamics Library ({KDL})},
+  author       = {Smits, Ruben and Soetens, Peter and Bruyninckx, Herman},
+  year         = {2009},
+  organization = {OROCOS Project},
+  url          = {https://www.orocos.org/kdl.html}
 }
 
-@software{Jocher_Ultralytics_YOLO_2023,   
-  author   = {Glenn Jocher and Jing Qiu and Ayush Chaurasia},
-  title    = {{Ultralytics YOLO}},
-  version  = {8.0.0},
-  year     = {2023},
-  month    = {January},
[…]
