+++ paper/paper.md
@@ -2,7 +2,7 @@
 title: "ManipulaPy: A GPU-Accelerated Python Framework for Robotic Manipulation, Perception, and Control"
 tags: [robotics, manipulator, simulation, kinematics, dynamics, perception, cuda, trajectory-planning, computer-vision]
 authors:
-  - name: M.I.M. AboElNasr
+  - name: M.I.M.AboElNasr
     orcid: 0000-0002-1768-2031
     affiliation: 1
 affiliations:
@@ -14,7 +14,7 @@ bibliography: paper.bib
 
 # Summary
 
-**ManipulaPy** is an open-source Python toolbox that unifies the entire manipulation pipeline—from URDF parsing to GPU-accelerated dynamics, vision-based perception, planning and control—within a single API. Built on the Product-of-Exponentials model [@lynch2017modern] (similar to Pinocchio [@Pinocchio2025] but with GPU acceleration), PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library enables researchers to move from robot description to real-time control with up to **13× overall performance improvement** and **3600× faster inverse dynamics** on a 6-DOF UR5 compared to NumPy baseline. Performance claims are reproducible via benchmarks in `benchmarks/README.md`.
+**ManipulaPy** is an open-source Python toolbox that unifies the entire manipulation pipeline—from URDF parsing to GPU-accelerated dynamics, vision-based perception, planning and control—within a single API. Built on the Product-of-Exponentials model [@lynch2017modern] (similar to Pinocchio [@Pinocchio2025] but with GPU acceleration), PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library enables researchers to move from robot description to real-time control with up to **13× overall performance improvement**. On 6-DOF arms, a specialized inverse-dynamics prototype attains up to **3600×** over NumPy, while our **DOF-agnostic GPU trajectory kernels** (joint & Cartesian) accelerate **7-DOF and higher** manipulators in practice. Performance claims are reproducible via benchmarks in `benchmarks/README.md`.
 
 # Statement of Need
 
@@ -67,7 +67,7 @@ ManipulaPy's architecture centers on a **unified manipulation pipeline** that ma
 
 **Robot Model Processing** converts URDF descriptions into Product-of-Exponentials representations, extracting screw axes, mass properties, and joint constraints through PyBullet integration. This creates the fundamental `SerialManipulator` and `ManipulatorDynamics` objects used throughout the system.
 
-**Kinematics and Dynamics** implement GPU-accelerated forward/inverse kinematics, Jacobian computation, and Newton-Euler dynamics. Custom CUDA kernels optimize critical operations for 6-DOF manipulators, enabling real-time performance at 1 kHz control rates.
+**Kinematics and Dynamics** provide vectorized FK/IK and Jacobians and **GPU-accelerated trajectory time-scaling** (joint & Cartesian) that is **DOF-agnostic** (applies to 6- and 7-DOF+). A **specialized 6-DOF inverse-dynamics prototype** demonstrates up to 3600× speedup; the **released GPU dynamics kernels** are shape-agnostic but **simplified** (per-joint/diagonalized). The **fully coupled n-DOF spatial dynamics** remain on the CPU path for exactness.
 
[…]
