### 2025-09-01 #8490 @cadojo
Thanks everyone for your patience. I have completed my review. The only remaining open GitHub Issue I have (concerning this review) is [one additional paper comment](https://github.com/boelnasr/ManipulaPy/issues/17#issuecomment-3243165670) on a statement of Pinocchio's capabilities that I feel may not be fully accurate.

In my opinion, the software project and paper meet JOSS standards for publishing.
