### 2025-09-07 f408ba7
paper updates
