+++ paper/paper.md
@@ -2,7 +2,7 @@
 title: "ManipulaPy: A GPU-Accelerated Python Framework for Robotic Manipulation, Perception, and Control"
 tags: [robotics, manipulator, simulation, kinematics, dynamics, perception, cuda, trajectory-planning, computer-vision]
 authors:
-  - name: M.I.M.AboElNasr
+  - name: M.I.M. AboElNasr
     orcid: 0000-0002-1768-2031
     affiliation: 1
 affiliations:
@@ -14,256 +14,106 @@ bibliography: paper.bib
 
 # Summary
 
-**ManipulaPy** is an open-source Python toolbox that unifies the entire manipulation pipeline—from URDF parsing to GPU-accelerated dynamics, vision-based perception, planning and control—within a single API. Built on the Product-of-Exponentials model [@lynch2017modern] (similar to Pinocchio [@Pinocchio2025] but with GPU acceleration), PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library enables researchers to move from robot description to real-time control with up to **13× overall performance improvement**. On 6-DOF arms, a specialized inverse-dynamics prototype attains up to **3600×** over NumPy, while our **DOF-agnostic GPU trajectory kernels** (joint & Cartesian) accelerate **7-DOF and higher** manipulators in practice. Performance claims are reproducible via benchmarks in `benchmarks/README.md`.
+**ManipulaPy** is an open-source Python toolbox that unifies the entire manipulation pipeline—from URDF parsing to GPU-accelerated dynamics, vision-based perception, planning and control—within a single API. Built on the Product-of-Exponentials (PoE) model [@lynch2017modern], PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library enables researchers to move from robot description to real-time control with up to 40× speedup over CPU implementations. DOF-agnostic GPU trajectory kernels accelerate 6-DOF and higher manipulators, while specialized inverse-dynamics prototypes achieve up to 3600× speedup for batch processing. Performance claims are reproducible via benchmarks in the repository.
 
 # Statement of Need
 
-Modern manipulation research requires tight integration of geometry, dynamics, perception, planning, and control—ideally within a single, real-time computational loop on GPU hardware. However, existing open-source tools address only portions of this pipeline, forcing researchers to write substantial integration code:
+Modern manipulation research requires tight integration of geometry, dynamics, perception, planning, and control within a unified computational framework. However, existing open-source tools address only portions of this pipeline, forcing researchers to write substantial integration code:
 
-| Library | Core Strengths | Integration Challenges |
-|---------|---------------|------------------------|
-| **MoveIt** [@chitta2012moveit] | Mature sampling-based planners | Requires custom ROS nodes to bridge sensor data with planning; external plugins needed for real-time dynamics; no native GPU acceleration |
-| **Pinocchio** [@Pinocchio2025] | High-performance PoE dynamics (C++) | CPU-only; separate perception and planning libraries must be manually synchronized; requires Python bindings for integration |
-| **CuRobo** [@sundaralingam2023curobo] | GPU-accelerated collision checking and trajectory optimization | Planning-focused; lacks perception pipeline and closed-loop control; requires external sensor processing |
-| **Python Robotics Toolbox** [@corke2021] | Educational algorithms with clear APIs | CPU-only implementation; users must implement their own simulators, controllers, and sensor processing |
-| **PyRoKi** [@pyroki2025] | JAX-accelerated kinematics | Early development stage; limited dynamics and no perception support |
[…]
