### 2025-09-09 #8490 @lockwo
Ok, that's some good changes. Thanks @boelnasr . A few more notes.

On the paper:
- The figures have duplicate figure numbers (e.g. "Figure 1: Figure 1:").
- Figure 1, the green box overlaps with the label text "CUDA Kernels\n cupy/custom"
- We're still a little over 1,000 words, which can be ok, but I think the Theory and Implementation section can be largely cut (it could be valuable to keep somewhere in the documentation, but it represents technical details that aren't given enough depth and thus presumably are understandable by people who don't need it said).

On the repo:
- The documentation link at the top of the readme doesn't go anywhere when clicked

Next steps outside the paper iteration (once repo points are resolved):
