### 2025-09-27 b0598aa
Manuscript update
