### 2025-09-27 b9d46b0
paper triming
