### 2025-09-27 #8490 @boelnasr
@lockwo sorry for the late reply, I was busy in the last 2 weeks i applied the edits can you confirm if the manuscript passes before making the tag and archiving the repo on Zenodo.
