### 2025-10-01 #8490 @lockwo
Ah sorry, I missed this notification @boelnasr. I think we are very close now, I think this is my last round of iterations.

- "up to 40× speedup" -> "up to a 40x speedup"
- “inverse-dynamics prototypes achieve up to 3600× speedup” -> “speedups”
- The Limitations and Design Trade-offs section has some of repeated points, would it be possible to merge it with the CPU/GPU section (maybe just call it "design considerations" or something to that effect)?
