### 2025-10-05 705b517
script updates
