+++ paper/paper.md
@@ -14,7 +14,7 @@ bibliography: paper.bib
 
 # Summary
 
-**ManipulaPy** is an open-source Python toolbox that unifies the entire manipulation pipeline—from URDF parsing to GPU-accelerated dynamics, vision-based perception, planning and control—within a single API. Built on the Product-of-Exponentials (PoE) model [@lynch2017modern], PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library enables researchers to move from robot description to real-time control with up to 40× speedup over CPU implementations. DOF-agnostic GPU trajectory kernels accelerate 6-DOF and higher manipulators, while specialized inverse-dynamics prototypes achieve up to 3600× speedup for batch processing. Performance claims are reproducible via benchmarks in the repository.
+**ManipulaPy** is an open-source Python toolbox that unifies the entire manipulation pipeline—from URDF parsing to GPU-accelerated dynamics, vision-based perception, planning and control—within a single API. Built on the Product-of-Exponentials (PoE) model [@lynch2017modern], PyBullet [@coumans2019], CuPy [@cupy2021] and custom CUDA kernels [@liang2018gpu], the library enables researchers to move from robot description to real-time control with up to a 40× speedup over CPU implementations. DOF-agnostic GPU trajectory kernels accelerate 6-DOF and higher manipulators, while specialized inverse-dynamics prototypes achieve up to 3600× speedups for batch processing. Performance claims are reproducible via benchmarks in the repository.
 
 # Statement of Need
 
@@ -67,28 +67,29 @@ ManipulaPy's perception system converts raw sensor data into actionable robot kn
 
 ![GPU-accelerated trajectory execution demonstration in PyBullet simulation. A 6-DOF robotic manipulator executes a complex trajectory while avoiding dynamic obstacles in real-time. The trajectory planning utilizes GPU acceleration for 40× speedup over CPU implementation, enabling 1 kHz control rates with real-time collision avoidance through potential field methods integrated with CUDA kernels.](manipulapy_trajectory.png)
 
-# CPU vs GPU Module Requirements
+# Design Considerations
 
-ManipulaPy provides tiered functionality that gracefully scales from CPU-only to GPU-accelerated operation:
+ManipulaPy provides tiered functionality that scales gracefully from CPU-only to GPU-accelerated operation while maintaining a balance between performance, flexibility, and precision.  
+This section outlines the computational tiers, vision dependencies, and design trade-offs that guided the framework’s architecture.
 
 ## CPU-Only Features
-Core robotics modules include URDF processing, forward/inverse kinematics, Jacobian analysis, small trajectory planning (N < 1000 points), basic control, and simulation setup. Performance characteristics include single trajectory generation (~10–50 ms for 6-DOF robots) and real-time control limited to ~100 Hz due to Python's Global Interpreter Lock.
+Core robotics modules include URDF processing, forward and inverse kinematics, Jacobian analysis, small-scale trajectory planning (N < 1000 points), basic control, and simulation setup.  
+Performance characteristics include single trajectory generation in **~10–50 ms** for 6-DOF robots and real-time control frequencies up to **~100 Hz**, primarily limited by Python’s Global Interpreter Lock (GIL).
 
-## GPU-Required Features
-High-performance modules include large trajectory planning (N > 1000 points) with 40× speedup, batch processing, real-time inverse dynamics >1 kHz, workspace analysis with Monte Carlo sampling, and GPU-accelerated potential fields. Performance characteristics include large trajectory generation (~1–5 ms for 6-DOF robots) and real-time control at 1 kHz rates.
+## GPU-Accelerated Features
[…]
